### 2026-01-13 #9769 @editorialbot
[PRE REVIEW]: pedon: A Python package for analyzing unsaturated soil hydraulic properties
**Submitting author:** @martinvonk (<a href="http://orcid.org/0009-0007-3528-2991">M.A. Vonk</a>)
**Repository:** https://github.com/martinvonk/pedon
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.11
