### 2026-01-14 da0507d
add urls for some refs
