+++ paper/paper.bib
@@ -14,7 +14,8 @@
   author       = {{van Genuchten}, M. Th. and Leij, F. J. and Yates, S. R.},
   organization = {U.S. Salinity Laboratory, USDA, ARS},
   address      = {Riverside, CA},
-  year         = {1991}
+  year         = {1991},
+  url          = {https://www.pc-progress.com/documents/programs/retc.pdf}
 }
 
 @article{carsel_dataset_1988,
@@ -75,7 +76,8 @@
   title        = {The {HYDRUS-1D} software package for simulating the one-dimensional movement of water, heat, and multiple solutes in variably-saturated media. {Version 4.08}},
   organization = {University of California Riverside},
   address      = {Riverside, CA},
-  year         = {2009}
+  year         = {2009},
+  url          = {https://www.pc-progress.com/downloads/pgm_hydrus1d/hydrus1d-4.08.pdf}
 }
 
 @article{fuentes_burdine_1992,
@@ -206,7 +208,8 @@
   institution = {Alterra},
   number      = {153},
   year        = {2001},
-  language    = {Dutch}
+  language    = {Dutch},
+  url         = {https://edepot.wur.nl/43272}
 }
 
 @article{cooper_pedotransfer_2021,
