### 2026-01-14 0ceef43
add research impact and AI statement
