+++ paper/paper.md
@@ -124,5 +124,16 @@ bc = pe.SoilSample(h=h, theta=theta, k=k).fit(pe.Brooks)
 
 ![Resulting fit of the Brooks-Corey soil model on the Mualem-van Genuchten soil model \label{fig:swrc_fit}](figures/swrc_fit.png)
 
+# Research impact statement
+Soil hydraulic functions and their parameters are essential for simulating variably saturated groundwater flow. Determining these parameters experimentally is difficult, time-consuming, expensive, and uncertain [@genuchten_retc_1991; @brandhorst_uncertain_2017]. As a result, parameters are often approximated or taken from reference databases. `pedon` bundles soil hydraulic models and parameter sources in a single framework, enabling efficient parameter derivation without extensive literature searches or ad hoc reimplementation.
+
+As a result, `pedon` is already used in scientific workflows for variably saturated groundwater flow modeling, including in published research workflows from @vonk_nonlinear_2024 and @collenteur_signatures_2025. Additionally, `pedon` is a dependency of the Python package [`dutchsoils`](https://github.com/markvdbrink/dutchsoils), which is used in a scientific context to access and process Dutch soil datasets.
+
+# AI usage disclosure
+Generative AI tools (GitHub Copilot) were used during the development of this project for reviewing pull requests, writing unit tests, providing code completion suggestions, and sanity-checking proposed bug solutions.
+
+For the preparation of this manuscript, generative AI (OpenAI ChatGPT) was used to review the reference list, identify linguistic, grammatical, and textual errors, and verify that the manuscript complies with the requirements of the Journal of Open Source Software.
+
+All outputs generated by AI tools were thoroughly reviewed and validated by the authors. Human authors made all primary architectural, design, and scientific decisions. The authors take full responsibility for the accuracy, originality, licensing compliance, and ethical and legal standards of all submitted materials.
 
 # References
