+++ paper/paper.bib
@@ -333,3 +333,35 @@
   doi     = {10.2136/sssaj1977.03615995004100020024x},
   year    = {1977}
 }
+
+@article{collenteur_signatures_2025,
+  author  = {Collenteur, R. A. and Vonk, M. A. and Haaf, E.},
+  title   = {Quantification and Analysis of Hydrograph Behavior Using Groundwater Signatures},
+  journal = {Groundwater},
+  volume  = {63},
+  number  = {5},
+  pages   = {779--789},
+  doi     = {10.1111/gwat.13486},
+  year    = {2025}
+}
+
+@article{vonk_nonlinear_2024,
+  author  = {Vonk, M. A. and Collenteur, R. A. and Panday, S. and Schaars, F. and Bakker, M.},
+  title   = {Time Series Analysis of Nonlinear Head Dynamics Using Synthetic Data Generated with a Variably Saturated Model},
+  journal = {Groundwater},
+  volume  = {62},
+  number  = {5},
+  pages   = {748--760},
+  doi     = {10.1111/gwat.13403},
+  year    = {2024}
+}
+
+@article{brandhorst_uncertainty_2017,
+  author  = {Brandhorst, N. and Erdal, D. and Neuweiler, I.},
+  title   = {Soil moisture prediction with the ensemble Kalman filter: Handling uncertainty of soil hydraulic parameters},
+  journal = {Advances in Water Resources},
+  volume  = {110},
+  pages   = {360--370},
+  doi     = {10.1016/j.advwatres.2017.10.022},
+  year    = {2017}
+}
