### 2026-01-14 819e01e
add part on software design
