### 2026-01-14 bcb03f8
trim the manuscript
