+++ paper/paper.md
@@ -32,22 +32,22 @@ bibliography: paper.bib
 `pedon` is a Python package for describing and analyzing unsaturated soil hydraulic properties. It provides a framework for soil hydraulic models, along with tools for retrieving parameters from soil databases, applying pedotransfer functions, and fitting model parameters to measurements.
 
 # Statement of need
-Researchers and engineers working with unsaturated soils often need estimates of soil parameters for their variably saturated groundwater flow models. `pedon` addresses this need by providing a modern Python toolkit that brings together commonly used soil hydraulic models, parameter databases, pedotransfer functions, and fitting routines. This makes soil analysis faster, more reproducible, and easier to integrate into existing groundwater modeling workflows.
+Researchers and engineers working with unsaturated soils need estimates of soil parameters for variably saturated groundwater flow models. `pedon` provides a Python toolkit that brings together soil hydraulic models, parameter databases, pedotransfer functions, and fitting routines, making soil analysis faster, more reproducible, and easier to integrate into existing groundwater modeling workflows.
 
 # Soil hydraulic models
-Different soil hydraulic models (hereafter referred to as soil models) are available in `pedon`. A soil model is a parametric description of soil hydraulic functions, namely the soil water retention curve (SWRC) and the (unsaturated) hydraulic conductivity function (HCF). These link the soil water content and flow to pressure head or saturation for use in variably saturated groundwater flow models. At this time, the following soil models are available:
+`pedon` provides several soil hydraulic models. A soil model is a parametric description of soil hydraulic functions: the soil water retention curve (SWRC) and the unsaturated hydraulic conductivity function (HCF). These relate soil water content and flow to pressure head for use in variably saturated groundwater flow models. At this time, the following soil models are available:
 
 - Mualem-van Genuchten [@genuchten_mualem_1980]: `pedon.Genuchten`
 - Brooks-Corey [@brooks_corey_1964]: `pedon.Brooks`
-- Combination of the van Genuchten SWRC and Brooks-Corey HCF [@fuentes_burdine_1992; @panday_mfusgt_2025]: `pedon.Panday`
+- Mualem-van Genuchten SWRC and Brooks-Corey HCF [@fuentes_burdine_1992; @panday_mfusgt_2025]: `pedon.Panday`
 - Fredlund-Xing [@fredlund_xing_1994]: `pedon.Fredlund`
 - Haverkamp [@haverkamp_model_1977]: `pedon.Haverkamp`
 - Gardner(-Kozeny) [@gardner_model_1958; @brutsaert_kozeny_1967; @bakker_gardner_2009; @mathias_gardner_2006]: `pedon.Gardner`
 - Gardner-Rucker [@rucker_gardner_2005]: `pedon.Rucker`
-- Combination of the van Genuchten SWRC and Gardner HCF [@genuchten_mualem_1980; @gardner_model_1958]: `pedon.GenuchtenGardner`
+- Mualem-van Genuchten SWRC and Gardner HCF [@genuchten_mualem_1980; @gardner_model_1958]: `pedon.GenuchtenGardner`
 
 # Software design
-The soil models are implemented as Python classes, providing a clear and consistent structure in which model-specific methods, such as those for evaluating the SWRC and HCF, are defined. For example, the Mualem–van Genuchten model can be instantiated and used as follows:
+The soil models are implemented as Python classes with model-specific methods for evaluating the SWRC and HCF. For example, the Mualem–van Genuchten model can be used as follows:
 
 ```python
 import numpy as np
@@ -66,17 +66,17 @@ theta = mg.theta(h)  # water content (-) at pressure head values
 k = mg.k(h)  # hydraulic conductivity (cm/d) at pressure head values
 ```
 
[…]
