### 2026-01-14 96a2336
last textual changes
