+++ paper/paper.md
@@ -29,24 +29,24 @@ bibliography: paper.bib
 ---
 
 # Summary
-`pedon` is a Python package for describing and analyzing unsaturated soil hydraulic properties. It provides a framework for soil hydraulic models, along with tools for retrieving parameters from soil databases, applying pedotransfer functions, and fitting model parameters to measurements.
+`pedon` is a Python package for describing and analyzing unsaturated soil hydraulic properties. It provides a framework for soil hydraulic models, along with tools for retrieving parameters from soil databases, applying pedotransfer functions, and fitting soil hydraulic parameters to measurements.
 
 # Statement of need
 Researchers and engineers working with unsaturated soils need estimates of soil parameters for variably saturated groundwater flow models. `pedon` provides a Python toolkit that brings together soil hydraulic models, parameter databases, pedotransfer functions, and fitting routines, making soil analysis faster, more reproducible, and easier to integrate into existing groundwater modeling workflows.
 
 # Soil hydraulic models
-`pedon` provides several soil hydraulic models. A soil model is a parametric description of soil hydraulic functions: the soil water retention curve (SWRC) and the unsaturated hydraulic conductivity function (HCF). These relate soil water content and flow to pressure head for use in variably saturated groundwater flow models. At this time, the following soil models are available:
+A soil hydraulic model is a parametric description of soil hydraulic functions: the soil water retention curve (SWRC) and the unsaturated hydraulic conductivity function (HCF). These relate soil water content and flow to pressure head and vice versa for use in variably saturated groundwater flow models. At this time, `pedon` provides the following soil models:
 
-- Mualem-van Genuchten [@genuchten_mualem_1980]: `pedon.Genuchten`
-- Brooks-Corey [@brooks_corey_1964]: `pedon.Brooks`
-- Mualem-van Genuchten SWRC and Brooks-Corey HCF [@fuentes_burdine_1992; @panday_mfusgt_2025]: `pedon.Panday`
-- Fredlund-Xing [@fredlund_xing_1994]: `pedon.Fredlund`
-- Haverkamp [@haverkamp_model_1977]: `pedon.Haverkamp`
-- Gardner(-Kozeny) [@gardner_model_1958; @brutsaert_kozeny_1967; @bakker_gardner_2009; @mathias_gardner_2006]: `pedon.Gardner`
-- Gardner-Rucker [@rucker_gardner_2005]: `pedon.Rucker`
-- Mualem-van Genuchten SWRC and Gardner HCF [@genuchten_mualem_1980; @gardner_model_1958]: `pedon.GenuchtenGardner`
+- `pedon.Genuchten`: Mualem-van Genuchten [@genuchten_mualem_1980]
+- `pedon.Brooks`: Brooks-Corey [@brooks_corey_1964]
+- `pedon.Panday`: Mualem-van Genuchten SWRC and Brooks-Corey HCF [@fuentes_burdine_1992; @panday_mfusgt_2025]
+- `pedon.Fredlund`: Fredlund-Xing [@fredlund_xing_1994]
+- `pedon.Haverkamp`: Haverkamp [@haverkamp_model_1977]
+- `pedon.Gardner`: Gardner(-Kozeny) [@gardner_model_1958; @brutsaert_kozeny_1967; @bakker_gardner_2009; @mathias_gardner_2006]
+- `pedon.Rucker`: Gardner-Rucker [@rucker_gardner_2005]
+- `pedon.GenuchtenGardner`: Mualem-van Genuchten SWRC and Gardner HCF [@genuchten_mualem_1980; @gardner_model_1958]
 
-# Software design
+## Software design
[…]
