### 2026-01-27 7421811
feat: add two methods for Gardner parameter approximation based on Peche et al. (in prep) and based on Ghezzehei et al. (2007)
