### 2026-01-30 b1340d6
updated our model for the approximation of Gardner parameter c
