+++ paper/paper.bib
@@ -365,3 +365,24 @@
   doi     = {10.1016/j.advwatres.2017.10.022},
   year    = {2017}
 }
+
+@article{seki_unsatfit_2023,
+  author  = {Seki, K. and Toride, N. and van Genuchten, M. Th.},
+  doi     = {10.2478/johh-2022-0039},
+  title   = {Evaluation of a general model for multimodal unsaturated soil hydraulic properties},
+  journal = {Journal of Hydrology and Hydromechanics},
+  number  = {1},
+  volume  = {71},
+  year    = {2023},
+  pages   = {22--34}
+}
+
+@article{memari_pyswr_2021,
+  author  = {S. S. Memari and T. P. Clement},
+  title   = {PySWR - A Python code for fitting soil water retention functions},
+  journal = {Computers & Geosciences},
+  volume  = {156},
+  pages   = {104897},
+  year    = {2021},
+  doi     = {10.1016/j.cageo.2021.104897}
+}
