+++ paper/paper.md
@@ -34,6 +34,9 @@ bibliography: paper.bib
 # Statement of need
 Researchers and engineers working with unsaturated soils need estimates of soil parameters for variably saturated groundwater flow models. `pedon` provides a Python toolkit that brings together soil hydraulic models, parameter databases, pedotransfer functions, and fitting routines, making soil analysis faster, more reproducible, and easier to integrate into existing groundwater modeling workflows.
 
+## State of the field
+pedon is a contribution to the groundwater modeling field that integrates, combining and extending on  existing tools, for example by incorporating parameter estimation algorithms from HYPAGS [@peche_hypags_2024]. Its primary novelty lies in the object-oriented representation of soil hydraulic models, which enables the definition of custom soil models and facilitates efficient comparison across formulations using the integrated parameter datasets and pedotransfer functions. The main functional overlap with existing software lies in the fitting of soil hydraulic functions using least-squares optimization, as implemented in tools such as `unsatfit` [@seki_unsatfit_2023], `PySWR` [@memari_pyswr_2021], and `RETC` [@genuchten_retc_1991]. However, these tools are primarily standalone fitting utilities and do not provide an integrated framework that combines (custom) soil models, parameter datasets, and estimation methods within reproducible workflows.
+
 # Soil hydraulic models
 A soil hydraulic model is a parametric description of soil hydraulic functions: the soil water retention curve (SWRC) and the unsaturated hydraulic conductivity function (HCF). These relate soil water content and flow to pressure head and vice versa for use in variably saturated groundwater flow models. At this time, `pedon` provides the following soil models:
 
@@ -103,7 +106,7 @@ hypags: pe.Genuchten = pe.SoilSample(k=ks).hypags()
 ```
 
 ### Soil hydraulic measurements
-`pedon` can estimate parameters directly when measurements of soil water retention and/or unsaturated hydraulic conductivity are available. A soil model, together with its SWRC and HCF, is fitted to the data by minimizing the difference between measured and simulated values. This uses nonlinear least-squares algorithm from SciPy [@scipy_paper_2020] and follows the well-established methodology of the RETC software [@genuchten_retc_1991].
+`pedon` can estimate parameters directly when measurements of soil water retention and/or unsaturated hydraulic conductivity are available. A soil model, together with its SWRC and HCF, is fitted to the data by minimizing the difference between measured and simulated values. This uses nonlinear least-squares algorithm from SciPy [@scipy_paper_2020] and follows the well-established methodology of the `RETC` software [@genuchten_retc_1991].
 
 ### Soil model conversion
 The same fitting procedure can translate between soil models. The SWRC and HCF generated by one model are sampled over a range of pressure heads and refitted using another formulation. This enables direct model comparison (Figure \ref{fig:swrc_fit}) and facilitates integration with external tools when a different model is required.
