### 2026-01-30 aefb90d
add docstring to protocol
