### 2026-02-02 361c776
add type checking with ty
