### 2026-02-02 df3d5ae
proper ty command
