+++ paper/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 Researchers and engineers working with unsaturated soils need estimates of soil parameters for variably saturated groundwater flow models. `pedon` provides a Python toolkit that brings together soil hydraulic models, parameter databases, pedotransfer functions, and fitting routines, making soil analysis faster, more reproducible, and easier to integrate into existing groundwater modeling workflows.
 
 ## State of the field
-pedon is a contribution to the groundwater modeling field that integrates, combining and extending on  existing tools, for example by incorporating parameter estimation algorithms from HYPAGS [@peche_hypags_2024]. Its primary novelty lies in the object-oriented representation of soil hydraulic models, which enables the definition of custom soil models and facilitates efficient comparison across formulations using the integrated parameter datasets and pedotransfer functions. The main functional overlap with existing software lies in the fitting of soil hydraulic functions using least-squares optimization, as implemented in tools such as `unsatfit` [@seki_unsatfit_2023], `PySWR` [@memari_pyswr_2021], and `RETC` [@genuchten_retc_1991]. However, these tools are primarily standalone fitting utilities and do not provide an integrated framework that combines (custom) soil models, parameter datasets, and estimation methods within reproducible workflows.
+`pedon` contributes to the field of groundwater modeling by providing a modular, object-oriented framework that integrates soil hydraulic models, established parameter databases, and pedotransfer functions into a single, reproducible workflow. Existing tools such as unsatfit [@seki_unsatfit_2023], PySWR [@memari_pyswr_2021], and RETC [@genuchten_retc_1991] are utilities primarily focused on least-squares fitting of predefined soil hydraulic models. In contrast, `pedon` enables researchers to define custom soil hydraulic model formulations and systematically evaluate them against different datasets and pedotransfer functions. This level of interoperability is unique and provides a broader framework for the analysis of unsaturated soil hydraulic properties. At the same time, pedon is developed with a collaborative mindset, exemplified by the integration and extension of the soil hydraulic model parameter estimation algorithms from HYPAGS [@peche_hypags_2024].
 
 # Soil hydraulic models
 A soil hydraulic model is a parametric description of soil hydraulic functions: the soil water retention curve (SWRC) and the unsaturated hydraulic conductivity function (HCF). These relate soil water content and flow to pressure head and vice versa for use in variably saturated groundwater flow models. At this time, `pedon` provides the following soil models:
