+++ paper/paper.md
@@ -29,16 +29,13 @@ bibliography: paper.bib
 ---
 
 # Summary
-`pedon` is a Python package for describing and analyzing unsaturated soil hydraulic properties. It provides a framework for soil hydraulic models, along with tools for retrieving parameters from soil databases, applying pedotransfer functions, and fitting soil hydraulic parameters to measurements.
+`pedon` is a Python package for describing and analyzing unsaturated soil hydraulic properties. It provides a framework for soil hydraulic models, along with tools for retrieving parameters from soil databases, applying pedotransfer functions, and fitting soil hydraulic model parameters to measurements.
 
 # Statement of need
-Researchers and engineers working with unsaturated soils need estimates of soil hydraulic parameters for variably saturated groundwater flow models. `pedon` provides a Python toolkit that brings together soil hydraulic models, parameter databases, pedotransfer functions, and fitting routines, making soil analysis faster, more reproducible, and easier to integrate into existing groundwater modeling workflows.
-
-## State of the field
-`pedon` contributes to the field of groundwater modeling by providing a modular, object-oriented framework that integrates soil hydraulic models, established parameter databases, and pedotransfer functions into a single, reproducible workflow. Existing tools such as unsatfit [@seki_unsatfit_2023], PySWR [@memari_pyswr_2021], and RETC [@genuchten_retc_1991] are utilities primarily focused on least-squares fitting of predefined soil hydraulic models. In contrast, `pedon` enables researchers to define custom soil hydraulic model formulations and systematically evaluate them against different datasets and pedotransfer functions. This level of interoperability is unique and provides a broader framework for the analysis of unsaturated soil hydraulic properties. At the same time, pedon is developed with a collaborative mindset, exemplified by the integration and extension of the soil hydraulic model parameter estimation algorithms from HYPAGS [@peche_hypags_2024].
+Researchers and engineers working with unsaturated soils need estimates of soil hydraulic model parameters for variably saturated groundwater flow models. `pedon` provides a Python toolkit that brings together soil hydraulic models, parameter databases, pedotransfer functions, and fitting routines, making soil analysis faster, more reproducible, and easier to integrate into existing groundwater modeling workflows.
 
 # Soil hydraulic models
-A soil hydraulic model is a parametric description of soil hydraulic functions: the soil water retention curve (SWRC) and the unsaturated hydraulic conductivity function (HCF). These relate soil water content and flow to pressure head and vice versa for use in variably saturated groundwater flow models. At this time, `pedon` provides the following soil models:
+A soil hydraulic model (soil model in short) is a parametric description of soil hydraulic functions: the soil water retention curve (SWRC) and the unsaturated hydraulic conductivity function (HCF). These relate soil water content and flow to pressure head and vice versa for use in variably saturated groundwater flow models. At this time, `pedon` provides the following soil models:
 
 - `pedon.Genuchten`: Mualem-van Genuchten [@genuchten_mualem_1980]
 - `pedon.Brooks`: Brooks-Corey [@brooks_corey_1964]
@@ -89,10 +86,10 @@ staring = pe.Soil("B01").from_name(pe.Genuchten, source="Staring_2018")
 ```
 
 ## Parameter estimation
-`pedon` provides two approaches for obtaining soil hydraulic model parameters from soil data. The first uses pedotransfer functions based on easily measured soil properties. The second relies on direct measurements of soil water content and hydraulic conductivity.
+`pedon` provides two approaches for obtaining soil model parameters from soil data. The first uses pedotransfer functions based on easily measured soil properties. The second relies on direct measurements of soil water content and hydraulic conductivity.
 
[…]
