+++ docs/index.md
@@ -24,7 +24,7 @@ In addition, Pedon implements multiple pedotransfer functions that estimate soil
 
 By combining soil hydraulic models, reference datasets, pedotransfer functions, and fitting routines in a single, consistent framework, Pedon makes it straightforward to move from soil information—whether coarse texture data or detailed laboratory measurements—to parameterized soil models ready for use in variably saturated flow simulations.
 
-The pedon package is open-source and hosted on [GitHub]((https://github.com/martinvonk/pedon)), where you can find more information about the ongoing developments. The package is published on [PyPi](https://pypi.org/project/pedon/) from which it can be installed using `pip install pedon`. If you use this software in your research or analyses, please cite the package via its Zenodo archive:
+The pedon package is open-source and hosted on [GitHub](https://github.com/martinvonk/pedon), where you can find more information about the ongoing developments. The package is published on [PyPi](https://pypi.org/project/pedon/) from which it can be installed using `pip install pedon`. If you use this software in your research or analyses, please cite the package via its Zenodo archive:
 > Vonk, M. A. & Peche, A. (XXXX). Pedon: A Python package for analyzing unsaturated soil hydraulic properties (vX.X.X). Zenodo. [doi.org/10.5281/zenodo.18222514](https://doi.org/10.5281/zenodo.18222514).
 
 
