+++ _sources/examples/02_datasets.ipynb.txt
@@ -9,12 +9,21 @@
     "\n",
     "*Martin Vonk (2025)*\n",
     "\n",
-    "Generally either the Brooks-Corey or Mualem-van Genuchten soil models are used. Pedon has some built-in datasets with parameter sets that can be used for both soil models."
+    "## Overview\n",
+    "\n",
+    "When measured data is unavailable, soil hydraulic parameters can be obtained from established parameter databases. These databases contain pre-determined parameter sets for common soils based on hundreds of laboratory measurements and professional judgment.\n",
+    "\n",
+    "`pedon` integrates soil parameter data from three major sources:\n",
+    "1. **HYDRUS database** - Parameter sets from Carsel & Parrish (1988), widely used in HYDRUS and other simulators\n",
+    "2. **VS2D database** - Additional parameter sets in both Brooks-Corey and van Genuchten formats\n",
+    "3. **Staring series** - Dutch soil parameters based on BOFEK classification with many measured samples\n",
+    "\n",
+    "This notebook shows how to access these databases and understand their characteristics."
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 6,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -23,19 +32,108 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 7,
    "metadata": {},
-   "outputs": [],
+   "outputs": [
+    {
+     "data": {
+      "text/plain": [
+       "['Sand',\n",
+       " 'Loamy Sand',\n",
+       " 'Sandy Loam',\n",
+       " 'Loam',\n",
+       " 'Silt',\n",
+       " 'Silt Loam',\n",
+       " 'Sandy Clay Loam',\n",
+       " 'Clay Loam',\n",
+       " 'Silty Clay Loam',\n",
+       " 'Sandy Clay',\n",
+       " 'Silty Clay',\n",
+       " 'Clay',\n",
+       " 'B01',\n",
+       " 'B02',\n",
+       " 'B03',\n",
+       " 'B04',\n",
+       " 'B05',\n",
+       " 'B06',\n",
+       " 'B07',\n",
[…]
