### 2026-02-02 #9769 @martinvonk
Hi @arfon and @kthyng,

No worries at all for missing the State of the Field section. I missed it on the example page as well.

@apeche and I have revised the manuscript so it now includes the state of the field section as well. The updated draft is available in the repository.

Looking forward to the rest of your evaluation, and please let us know if any further adjustments are needed.

Best regards,
Martin
