### 2026-02-02 6644a62
make sure figure fits on page
