### 2026-03-13 5ded3df
Add converter class to soil.py
