+++ src/pedon/__init__.py
@@ -4,6 +4,7 @@ from ._params import get_params as get_params
 from ._version import __version__ as __version__
 from ._version import show_versions as show_versions
 from .soil import Soil as Soil
+from .soil import SoilModelConverter as SoilModelConverter
 from .soil import SoilSample as SoilSample
 from .soilmodel import Brooks as Brooks
 from .soilmodel import Fredlund as Fredlund
