+++ pyproject.toml
@@ -94,7 +94,6 @@ commands = [
                 "check",
                 "--extend-select",
                 "I",
-                "--preview",
         ],
         [
                 "ruff",
