+++ pyproject.toml
@@ -41,7 +41,7 @@ documentation = [
         "ipykernel",
         "pandoc",
 ]
-dev = ["pedon[ruffing,typing,pytesting]", "tox"]
+dev = ["pedon[ruffing,typing,pytesting]", "tox-uv"]
 
 
 [tool.setuptools]
@@ -77,7 +77,7 @@ show-fixes = true
 fix = true
 
 [tool.tox]
-requires = ["tox>=4"]
+requires = ["tox-uv>=1.0"]
 env_list = ["py311", "py312", "py313", "py314", "type", "ruff", "typety"]
 
 [tool.tox.env_run_base]
@@ -104,7 +104,7 @@ commands = [
 
 
 [tool.tox.env.type]
-description = "run type checks"
+description = "run type checks with mypy"
 extras = ["typing"]
 commands = [["mypy", "src"]]
 
