+++ .github/workflows/documentation.yml
@@ -15,23 +15,26 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v6
 
-      - name: Set up Python
-        uses: actions/setup-python@v6
+      - name: Set up Python via uv
+        uses: astral-sh/setup-uv@v7
         with:
           python-version: '3.11'
           check-latest: true
 
       - name: Set up tox environment
         run: |
-          python -m pip install --upgrade pip
-          pip install tox
-          tox -e docu --notest
+          uv venv
+          uv pip install tox-uv
+          uv run tox -e docu --notest
 
-      - name: Install Pandoc and build documentation
+      - name: Install Pandoc
         run: |
           sudo apt-get update
           sudo apt-get install -y pandoc
-          tox -e docu --skip-pkg-install
+
+      - name: Build documentation
+        run: |
+          uv run tox -e docu --skip-pkg-install
 
       - name: Publish HTML output on gh-pages branch
         if: github.event_name == 'push'
