+++ .github/workflows/tests.yml
@@ -18,7 +18,7 @@ jobs:
             toxenv: py311
             experimental: false
           - name: Test suite with py312-ubuntu
-            python: "3.12-dev"
+            python: "3.12"
             os: ubuntu-latest
             toxenv: py312
             experimental: false
@@ -59,7 +59,7 @@ jobs:
     steps:
       - uses: actions/checkout@v6
 
-      - name: Set up Python via tox ${{ matrix.python }}
+      - name: Set up Python ${{ matrix.python }} via uv
         uses: astral-sh/setup-uv@v7
         with:
           python-version: ${{ matrix.python }}
