### 2026-03-26 #9769 @elbeejay
@davidlitwin, @tmweigand, @mgcooper, @frank1010111 - would any of you be willing and able to review the submission *pedon: A Python package for analyzing unsaturated soil hydraulic properties* for JOSS?

You've all published in JOSS, so I assume you are familiar with the general process but please reach out to me if you have any questions. At this time we're asking reviewers to complete their initial reviews within 6 weeks of the review formally beginning.
