+++ paper/paper.bib
@@ -14,8 +14,9 @@
   author       = {{van Genuchten}, M. Th. and Leij, F. J. and Yates, S. R.},
   organization = {U.S. Salinity Laboratory, USDA, ARS},
   address      = {Riverside, CA},
+  note         = {EPA Report 600/2-91/065},
   year         = {1991},
-  url          = {https://www.pc-progress.com/documents/programs/retc.pdf}
+  url          = {https://cfpub.epa.gov/si/si_public_record_report.cfm?dirEntryId=130162}
 }
 
 @article{carsel_dataset_1988,
@@ -57,6 +58,7 @@
   booktitle = {Hydrology Papers 3},
   publisher = {Colorado State University},
   address   = {Fort Collins, CO},
+  url       = {https://mountainscholar.org/items/3c7b98df-13e3-486c-9d1e-949a7a869f76},
   year      = {1964}
 }
 
@@ -130,6 +132,7 @@
   volume  = {85},
   number  = {4},
   pages   = {228--232},
+  doi     = {10.1097/00010694-195804000-00006},
   year    = {1958}
 }
 
@@ -293,12 +296,12 @@
   year    = {2020}
 }
 
-@manual{panday_mfusgt_2025,
+@manual{panday_mfusgt_2026,
   author       = {Panday, S.},
-  title        = {{USG-Transport}: Transport and other Enhancements to {MODFLOW-USG}. Version: 2.6.2},
+  title        = {{USG-Transport}: Transport and other Enhancements to {MODFLOW-USG}. Version: 2.7.0},
   organization = {GSI Environmental},
-  year         = {2025},
-  note         = {\url{https://www.gsienv.com/product/modflow-usg/}}
+  year         = {2026},
+  url          = {https://www.gsienv.com/software/modflow-usg/modflow-usg/}
 }
 
 @article{heinen_bofek_2022,
@@ -368,12 +371,12 @@
 
 @article{seki_unsatfit_2023,
   author  = {Seki, K. and Toride, N. and van Genuchten, M. Th.},
-  doi     = {10.2478/johh-2022-0039},
   title   = {Evaluation of a general model for multimodal unsaturated soil hydraulic properties},
   journal = {Journal of Hydrology and Hydromechanics},
   number  = {1},
   volume  = {71},
   year    = {2023},
+  doi     = {10.2478/johh-2022-0039},
   pages   = {22--34}
 }
 
