### 2026-03-28 #9769 @martinvonk
> Hi [@martinvonk](https://github.com/martinvonk), things generally look fine to me, please check that one DOI the bot flagged above. In the meantime I will be contacting potential reviewers. Thanks!

Hi @elbeejay, thank you for stepping in as the editor, and a thanks to @tmweigand for agreeing to review! :)

I've just updated the missing DOI that the bot flagged and took the opportunity to double-check and improve the URLs for a few of the older publications that don't have official DOIs.

Looking forward to the review process!
