### 2026-03-29 #9769 @DavidLitwin
> [@DavidLitwin](https://github.com/DavidLitwin), [@tmweigand](https://github.com/tmweigand), [@mgcooper](https://github.com/mgcooper), [@frank1010111](https://github.com/frank1010111) - would any of you be willing and able to review the submission _pedon: A Python package for analyzing unsaturated soil hydraulic properties_ for JOSS?
>
> You've all published in JOSS, so I assume you are familiar with the general process but please reach out to me if you have any questions. At this time we're asking reviewers to complete their initial reviews within 6 weeks of the review formally beginning.

I'm sorry, I'm not available at the moment -- too busy with teaching and other reviews.
