### 2026-03-31 29b8e8a
Revert "better type errors and list_methods update"
This reverts commit 54a74cd7849e51bc5aaa99f673a0131968c4a20b.
