### 2026-03-31 3f4cbb2
Reapply "better type errors and list_methods update"
This reverts commit 29b8e8a03e3b09804db1bddaf79a2d27f90d0223.
