### 2026-03-31 7426b74
Compute h_b in Ghezzehei van Genuchten->Gardner
