### 2026-04-01 0fa695d
move SoilModelConverter to soilmodel.py
