+++ src/pedon/soil.py
@@ -24,7 +24,7 @@ from scipy.optimize import fixed_point, least_squares
 
 from ._params import get_params
 from ._typing import FloatArray
-from .soilmodel import Brooks, Gardner, Genuchten, SoilModel, get_soilmodel
+from .soilmodel import Brooks, Genuchten, SoilModel, get_soilmodel
 
 
 @dataclass
@@ -906,293 +906,3 @@ class Soil:
         ss = SoilSample().from_staring(name=self.name, year=year)
         self.__setattr__("sample", ss)
         return self
-
-
-@dataclass
-class SoilModelConverter:
-    """
-    Container class for soil hydraulic model conversion methods.
-
-    Converts between different soil hydraulic models (e.g., van Genuchten to Gardner).
-
-    Attributes
-    ----------
-    sm : SoilModel
-        The soil model instance to convert from.
-
-    Example
-    --------
-    >>> vg = Genuchten(k_s=1e-5, alpha=2.0, n=2.5, theta_s=0.45, theta_r=0.05)
-    >>> converter = SoilModelConverter(sm=vg)
-    >>> gardner = converter.peche()
-    >>> print(gardner.c)  # Output: 4.4
-
-    """
-
-    sm: SoilModel
-
-    def list_methods(self) -> list[str]:
-        """
-        Get all available conversion methods for the current SoilModel type.
-
-        Returns
-        -------
-        list[str]
-            List of method names that are applicable for the current SoilModel.
-
-        Example
-        --------
-        >>> vg = Genuchten(k_s=1e-5, alpha=2.0, n=2.5, theta_s=0.45, theta_r=0.05)
-        >>> converter = SoilModelConverter(sm=vg)
-        >>> converter.list_methods()
-        ['ghezzehei', 'peche']
-        """
-        all_methods = []
-
-        for attr_name in dir(self):
-            if attr_name.startswith("_") or attr_name == "list_methods":
-                continue
-            attr = getattr(self, attr_name)
-            if not callable(attr):
-                continue
-            try:
-                attr()
-                all_methods.append(attr_name)
-            except TypeError:
-                pass  # method does not support the current model type
-
-        return all_methods
-
-    def morel(self) -> Brooks | Genuchten:
-        """
-        Convert between van Genuchten and Brooks-Corey models using the Morel-Seytoux
-        equivalence method, which preserves the effective capillary drive.
-
-        This method implements the parameter equivalence from Morel-Seytoux et al. (1996),
-        providing conversion between Brooks-Corey (BC) and van Genuchten (vG) parameters
-        while preserving the maximum value of the effective capillary drive Hc.M,
-        defined as the integral of relative permeability over capillary pressure.
-
-        The conversion is based on two criteria:
-        1. Preserve the effective capillary drive (primary criterion)
-        2. Preserve asymptotic behavior at low water contents (secondary criterion)
-
-        Returns
-        -------
-        Brooks or Genuchten
-            Converted soil model instance:
-            - If input is Genuchten, returns Brooks model
-            - If input is Brooks, returns Genuchten model
-
-        References
-        ----------
-        Morel-Seytoux, H. J., Meyer, P. D., Nachabe, M., Touma, J., van Genuchten, M. T.,
-        & Lenhard, R. J. (1996). Parameter equivalence for the Brooks-Corey and van
-        Genuchten soil characteristics: Preserving the effective capillary drive.
-        Water Resources Research, 32(5), 1251-1258.
[…]
