+++ src/pedon/soilmodel.py
@@ -2,6 +2,7 @@ from dataclasses import dataclass, field
 from typing import Protocol, Type, runtime_checkable
 
 import matplotlib.pyplot as plt
+from matplotlib.axes import Axes
 from numpy import abs as npabs
 from numpy import exp, full, linspace, log, log10, logspace, maximum
 
@@ -43,9 +44,9 @@ class SoilModel(Protocol):
         ...         # Inverse of theta method
         ...         return ...
         ...
-        ...     def plot(self, ax: plt.Axes | None = None) -> plt.Axes:
+        ...     def plot(self, ax: Axes | None = None) -> Axes:
         ...         # Plot the soil water retention curve by calling `plot_swrc`
-        ...         return plot_swrc(self, ax=ax)
+        ...         return plot_swrc(sm=self, ax=ax)
     """
 
     def theta(self, h: FloatArray) -> FloatArray:
@@ -72,8 +73,22 @@ class SoilModel(Protocol):
         """Calculate pressure head h from soil moisture content (inverse of theta)."""
         ...
 
-    def plot(self, ax: plt.Axes | None = None) -> plt.Axes:
+    def plot(self, ax: Axes | None = None) -> Axes:
         """Plot the soil water retention curve by calling `plot_swrc`."""
+        ...
+
+    def convert(self, method: str | None = None) -> "list[str] | SoilModel":
+        """Convert this soil model to another model type.
+
+        If `method` is None, returns a list of available conversion methods for this
+        soil model. Otherwise, performs the specified conversion and returns the
+        converted soil model instance.
+        """
+        ...
+
+    def fit(self, sm: "Type[SoilModel]", h: FloatArray, **kwargs) -> "SoilModel":
+        """Fit this soil model to another soil model type using the provided pressure
+        head data `h` and optional keyword arguments for fitting."""
         ...
 
 
@@ -120,8 +135,20 @@ class Genuchten:
         h = 1 / self.alpha * ((1 / se) ** (1 / self.m) - 1) ** (1 / self.n)
         return h
 
-    def plot(self, ax: plt.Axes | None = None) -> plt.Axes:
-        return plot_swrc(self, ax=ax)
+    def plot(self, ax: Axes | None = None) -> Axes:
+        return plot_swrc(sm=self, ax=ax)
+
+    def convert(self, method: str | None = None) -> list[str] | SoilModel:
+
+        if method is None:
+            return SoilModelConverter.list_methods(sm=self)
+        return getattr(SoilModelConverter, method)(self)
+
+    def fit(self, sm: Type[SoilModel], h: FloatArray, **kwargs):
+        from .soil import SoilSample
+
+        ss = SoilSample(h=h, theta=self.theta(h), k=self.k(h))
+        return ss.fit(sm=sm, **kwargs)
 
 
 @dataclass
@@ -189,8 +216,20 @@ class Brooks:
             ) ** (-1 / self.l)
             return h
 
-    def plot(self, ax: plt.Axes | None = None) -> plt.Axes:
-        return plot_swrc(self, ax=ax)
+    def plot(self, ax: Axes | None = None) -> Axes:
+        return plot_swrc(sm=self, ax=ax)
+
+    def convert(self, method: str | None = None) -> list[str] | SoilModel:
+
+        if method is None:
+            return SoilModelConverter.list_methods(sm=self)
+        return getattr(SoilModelConverter, method)(self)
+
+    def fit(self, sm: Type[SoilModel], h: FloatArray, **kwargs):
+        from .soil import SoilSample
+
+        ss = SoilSample(h=h, theta=self.theta(h), k=self.k(h))
[…]
