+++ src/pedon/__init__.py
@@ -4,7 +4,6 @@ from ._params import get_params as get_params
 from ._version import __version__ as __version__
 from ._version import show_versions as show_versions
 from .soil import Soil as Soil
-from .soil import SoilModelConverter as SoilModelConverter
 from .soil import SoilSample as SoilSample
 from .soilmodel import Brooks as Brooks
 from .soilmodel import Fredlund as Fredlund
@@ -15,5 +14,6 @@ from .soilmodel import Haverkamp as Haverkamp
 from .soilmodel import Panday as Panday
 from .soilmodel import Rucker as Rucker
 from .soilmodel import SoilModel as SoilModel
+from .soilmodel import SoilModelConverter as SoilModelConverter
 from .soilmodel import plot_hcf as plot_hcf
 from .soilmodel import plot_swrc as plot_swrc
