### 2026-04-01 667e044
update type hints
