+++ src/pedon/soilmodel.py
@@ -2,11 +2,10 @@ from dataclasses import dataclass, field
 from typing import Protocol, Type, runtime_checkable
 
 import matplotlib.pyplot as plt
-from matplotlib.axes import Axes
 from numpy import abs as npabs
 from numpy import exp, full, linspace, log, log10, logspace, maximum
 
-from ._typing import FloatArray, SoilModelNames
+from ._typing import FloatArray, MatplotlibAxes, SoilModelNames
 
 
 @runtime_checkable
@@ -44,7 +43,7 @@ class SoilModel(Protocol):
         ...         # Inverse of theta method
         ...         return ...
         ...
-        ...     def plot(self, ax: Axes | None = None) -> Axes:
+        ...     def plot(self, ax: MatplotlibAxes | None = None) -> MatplotlibAxes:
         ...         # Plot the soil water retention curve by calling `plot_swrc`
         ...         return plot_swrc(sm=self, ax=ax)
     """
@@ -73,7 +72,7 @@ class SoilModel(Protocol):
         """Calculate pressure head h from soil moisture content (inverse of theta)."""
         ...
 
-    def plot(self, ax: Axes | None = None) -> Axes:
+    def plot(self, ax: MatplotlibAxes | None = None) -> MatplotlibAxes:
         """Plot the soil water retention curve by calling `plot_swrc`."""
         ...
 
@@ -135,7 +134,7 @@ class Genuchten:
         h = 1 / self.alpha * ((1 / se) ** (1 / self.m) - 1) ** (1 / self.n)
         return h
 
-    def plot(self, ax: Axes | None = None) -> Axes:
+    def plot(self, ax: MatplotlibAxes | None = None) -> MatplotlibAxes:
         return plot_swrc(sm=self, ax=ax)
 
     def convert(self, method: str | None = None) -> list[str] | SoilModel:
@@ -216,7 +215,7 @@ class Brooks:
             ) ** (-1 / self.l)
             return h
 
-    def plot(self, ax: Axes | None = None) -> Axes:
+    def plot(self, ax: MatplotlibAxes | None = None) -> MatplotlibAxes:
         return plot_swrc(sm=self, ax=ax)
 
     def convert(self, method: str | None = None) -> list[str] | SoilModel:
@@ -270,7 +269,7 @@ class Haverkamp:
         s = (theta - self.theta_r) / (self.theta_s - self.theta_r)
         return (self.alpha * ((1.0 / s) - 1.0)) ** (1.0 / self.beta)
 
-    def plot(self, ax: Axes | None = None) -> Axes:
+    def plot(self, ax: MatplotlibAxes | None = None) -> MatplotlibAxes:
         return plot_swrc(sm=self, ax=ax)
 
     def convert(self, method: str | None = None) -> list[str] | SoilModel:
@@ -322,7 +321,7 @@ class Gardner:
     def h(self, theta: FloatArray) -> FloatArray:
         return self.h_b - (1.0 / self.m) * log(theta / self.theta_s)
 
-    def plot(self, ax: Axes | None = None) -> Axes:
+    def plot(self, ax: MatplotlibAxes | None = None) -> MatplotlibAxes:
         return plot_swrc(sm=self, ax=ax)
 
     def convert(self, method: str | None = None) -> list[str] | SoilModel:
@@ -378,7 +377,7 @@ class Rucker:
             (theta - self.theta_r) / (self.theta_s - self.theta_r)
         )
 
-    def plot(self, ax: Axes | None = None) -> Axes:
+    def plot(self, ax: MatplotlibAxes | None = None) -> MatplotlibAxes:
[…]
