+++ src/pedon/_typing.py
@@ -1,8 +1,10 @@
 from typing import Literal
 
+from matplotlib.axes import Axes
 from numpy import float64
 from numpy.typing import NDArray
 
+MatplotlibAxes = Axes
 FloatArray = float | NDArray[float64]
 SoilModelNames = Literal[
     "Genuchten",
