### 2026-04-01 f6aa28d
Reapply "create tests"
This reverts commit c9deecd1401b248204c9eaa25c47ae498086074f.
