### 2026-04-01 7c9e501
update protocol docstring
