### 2026-04-05 #9769 @keurfonluu
> @WilkAndy, @keurfonluu - would either of you be willing and able to review the submission *pedon: A Python package for analyzing unsaturated soil hydraulic properties* for JOSS?
>
> You've both published in JOSS, so I assume you are familiar with the general process but please reach out to me if you have any questions. At this time we're asking reviewers to complete their initial reviews within 6 weeks of the review formally beginning.

I would be glad to review this package as it perfectly fits my field of work.
