### 2026-04-13 #10409 @editorialbot
[REVIEW]: pedon: A Python package for analyzing unsaturated soil hydraulic properties
**Submitting author:** @martinvonk (<a href="http://orcid.org/0009-0007-3528-2991">M.A. Vonk</a>)
**Repository:** https://github.com/martinvonk/pedon
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.11

@tmweigand, @keurfonluu, @gmarupilla, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
