### 2026-04-14 #10409 @gmarupilla
Hi @elbeejay, I noticed this review and wanted to check if additional reviewer support would be helpful. I have some familiarity with soil datasets/workflows (including SSURGO) and soil hydraulic modeling, and would be happy to contribute if needed. Totally fine if the current reviewer coverage is sufficient.
