### 2026-04-15 #10409 @elbeejay
@openjournals/dev is there some other way we're supposed to be adding reviewers to a review issue? It doesn't look like @gmarupilla is able to create their reviewer checklist.
