### 2026-04-18 #10409 @gmarupilla
Hi @martinvonk, thanks for the submission. I have completed my review of `pedon` v0.0.11. Overall this is a well-scoped, well-tested package that clearly meets the JOSS criteria — I am recommending **acceptance after minor revisions**. A small set of non-blocking issues follow.

1. **Broken PyPI badge link** — `README.md` line 3 has `https://pypi.org/projectpedonspei/` (should be `https://pypi.org/project/pedon/`). The image renders but the hyperlink 404s.
2. **Typo in paper** — `paper/paper.md` line 122: "soil hydralulic model parameters" → "soil hydraulic". Also same line: "in a academic context" → "in an academic context".
3. **Version drift between submission and release** — the review tag is `v0.0.11` but PyPI is on `0.0.12`. Consider either tagging a JOSS-specific release at the reviewed commit or updating the version metadata in the review issue so the archived artifact matches what reviewers install.
4. **HYPAGS `rosetta` optional extra** — `SoilSample.hypags()` is advertised in the paper alongside `.cosby()` but Rosetta v1/2/3 (mentioned in the README) requires the `pedon[rosetta]` extra (`httpx`). A short note in the README / paper about which optional extras unlock which PTFs would help users avoid import surprises.

None of these block acceptance in my view. Happy to re-verify once the small revisions above are in. Thanks again for a clean and well-engineered submission.
