### 2026-04-28 #10409 @tmweigand
Greetings @martinvonk. I have started to look through the code and wanted to share my initial thoughts. While `pedon` provides a solid foundation, the package is currently fairly limited in scope and the range of use cases it can support is narrow. Expanding the functionality and validation would go a long way toward making it a go-to tool for the community.

- I think a modern and open-source implementation of `RETC` is needed as it is quite dated. However, this is still being highly cited (a few hundred times over the past few years!). For `pedon` to be a useful replacement, I think it should include tests to demonstrate similar results are given as compared to `RETC`.
- I have noticed there are a significant number of hard coded parameters throughout the code. While there are many call-level tests, I am not seeing validation tests to ensure these parameters (and functions) are correct and the results are expected. For example, the tests for the pedotransfer functions (`tests/test_ptf.py`) are only asserting the model type is correct and not validating the results. Improved testing that includes boundary checks, model formula checks, and complete invertibility checks (such as `tests/test_soilmodel.py::test_h_haverkamp_inverse`) should be included.
- Solutions to variably saturated groundwater models can be quite difficult and require refined grids and efficient methods. Have you checked performance to ensure that `pedon` is a viable option for these cases as mentioned in the `Statement of Need?` I suspect the Python implementation may result in long simulation times for realistic systems.
- Extending methods to include scanning curves, trapping models, and hysteresis models would be a valuable addition.
