### 2026-04-29 martinvonk/pedon#32 @martinvonk
Compare curve fitting routine outcomes to RETC
### Describe the enhancement

Originally posted by @tmweigand in https://github.com/openjournals/joss-reviews/issues/10409#issuecomment-4331811927
> I think a modern and open-source implementation of RETC is needed as it is quite dated. However, this is still being highly cited (a few hundred times over the past few years!). For pedon to be a useful replacement, I think it should include tests to demonstrate similar results are given as compared to RETC.

I got RETC working locally. I'll create a test to see if you the fitting routine estimates the same parameters (more or less).
