### 2026-04-29 martinvonk/pedon#33 @martinvonk
Extend soil models for hysteresis
### Describe the enhancement

Originally posted by @tmweigand in https://github.com/openjournals/joss-reviews/issues/10409#issuecomment-4331811927

> Extending methods to include scanning curves, trapping models, and hysteresis models would be a valuable addition.

Some good references:
Rudiyanto, Minasny, B., Shah, R. M., Setiawan, B. I., & van Genuchten, M. T. (2020). Simple functions for describing soil water retention and the unsaturated hydraulic conductivity from saturation to complete dryness. Journal of Hydrology, 588, Article 125041. https://doi.org/10.1016/j.jhydrol.2020.125041
