### 2026-04-29 martinvonk/pedon#34 @martinvonk
Add regression tests
### Describe the enhancement

Originally brought up by @tweigand in https://github.com/openjournals/joss-reviews/issues/10409#issuecomment-4331811927

The current test suite focuses primarily on execution-level coverage (verifying that the code runs and returns the correct types) rather than mathematical validation (verifying the stability and correctness of the outcomes). Because the models and Pedotransfer Functions (PTFs) rely on specific formulas and numerous hard-coded parameters, it is critical to ensure that refactoring or updates do not silently alter the numerical results.

Implementing Regression Testing should lock down the output of the mathematical models.
