### 2026-04-29 #10409 @martinvonk
Hi @tmweigand,

Thank you for taking the time to look into this package and for sharing your detailed thoughts.

I completely agree with your suggestions regarding expanded validation and broader scope. I have created GitHub issues to address the RETC validation tests (https://github.com/martinvonk/pedon/issues/32), the improved validation checks (https://github.com/martinvonk/pedon/issues/34), and the addition of hysteresis models (https://github.com/martinvonk/pedon/issues/33). These additions would substantially improve the Python package.

Regarding your comment on performance:
> * Solutions to variably saturated groundwater models can be quite difficult and require refined grids and efficient methods. Have you checked performance to ensure that `pedon` is a viable option for these cases as mentioned in the `Statement of Need?` I suspect the Python implementation may result in long simulation times for realistic systems.

I want to clarify the intended scope of the package. `pedon` is not designed to be a numerical solver for Richards' equation or to run the groundwater simulations itself. Instead, its primary purpose is to generate the soil hydraulic properties required as input for those computationally heavy numerical solvers.

For example, I frequently use `pedon` alongside [FloPy](https://github.com/modflowpy/flopy) to create inputs for [MODFLOW-USG Transport](https://www.gsienv.com/software/modflow-usg/modflow-usg/), which models 3D variably saturated flow. In workflows like this, including post-processing the results, `pedon` performs great (for academic research) in my experience.

Does that help clarify how pedon fits into the modeling pipeline, or were you referring something else with regards to the performance of a specific function within the package? Let me know!
