### 2026-04-29 0602fdb
create test to compare retc results
Co-authored-by: Copilot <copilot@github.com>
