+++ tests/test_retc_compare.py
@@ -1,0 +1,330 @@
+# %%
+import logging
+
+import numpy as np
+
+import pedon as pe
+
+logging.basicConfig(level="DEBUG")
+# %%
+# Dummy model for Silt in order to create SWRC and HCF
+p_true = {
+    "k_s": 24.96,  # cm/d
+    "theta_r": 0.078,  # cm3/cm3
+    "theta_s": 0.43,  # cm3/cm3
+    "alpha": 0.036,  # cm-1
+    "n": 1.56,  # -
+}
+gen = pe.Genuchten(**p_true)
+
+# %%
+# Create pressure head, water content and hydraulic conductivity arrays
+print("Pressure head (cm):")
+h = np.logspace(-3, 7, 11)
+_ = [print(f"{x:0.3f}".replace(".", ",")) for x in h]
+print("Water content:")
+theta = gen.theta(h)
+_ = [print(f"{x:0.4f}".replace(".", ",")) for x in theta]
+print("Conductivity (cm/d):")
+k = gen.k(h)
+_ = [print(f"{x:0.3e}".replace(".", ",")) for x in k]
+
+
+# %%
+#  RETC Output
+
+#  RETC, version 6.02, Released 13.02.2009
+#  M. Th. van Genuchten, J. Simunek, F. J. Leij, and M. Sejna
+#  Code for Quantifying the Hydraulic Functions of Unsaturated Soils
+#  Copyright © 2005-2009 Authors, All Rights Reserved
+
+#  RETC - Output ASCI File:
+#  *******************************************************************
+#  *                                                                 *
+#  *     Analysis of soil hydraulic properties                       *
+#  *                                                                 *
+#  *     Welcome to RETC                                             *
+#  *                                                                 *
+#  *     Mualem-based restriction, M=1-1/N                           *
+#  *     Simultaneous fit of retention and Conductivity data         *
+#  *     MType= 3     Method= 4                                      *
+#  *                                                                 *
+#  *******************************************************************
+
+
+#  INITial values of the coefficients
+#  ==================================
+#  No      Name        INITial value   Index
+#   1     ThetaR           .0670         1
+#   2     ThetaS           .4500         1
+#   3     Alpha            .0200         1
+#   4        n            1.4100         1
+#   5        m             .2908         0
+#   6        l             .5000         0
+#   7       Ks           10.8000         1
+
+
+#  Observed data
+#  =============
+#  Obs. No.    Pressure head     Water content     Weighting coefficient
+#      1            .001              .4300             1.0000
+#      2            .010              .4300             1.0000
+#      3            .100              .4300             1.0000
+#      4           1.000              .4293             1.0000
+#      5          10.000              .4074             1.0000
+#      6         100.000              .2421             1.0000
+#      7        1000.000              .1253             1.0000
+#      8       10000.000              .0910             1.0000
[…]
