### 2026-04-29 111ad65
update routine to allow kwargs for least_squares
Co-authored-by: Copilot <copilot@github.com>
