### 2026-04-29 f881ca8
implement some good ideas from copilot
