### 2026-04-29 8673141
fix broken url https://github.com/openjournals/joss-reviews/issues/10409#issuecomment-4272332656
