### 2026-04-29 0e19404
fix typo https://github.com/openjournals/joss-reviews/issues/10409#issuecomment-4272332656
