### 2026-04-29 ac9615a
mention rosetta is an optional dependency
https://github.com/openjournals/joss-reviews/issues/10409#issuecomment-4272332656
