### 2026-04-29 #10409 @martinvonk
> Hi [@martinvonk](https://github.com/martinvonk), thanks for the submission. I have completed my review of `pedon` v0.0.11. Overall this is a well-scoped, well-tested package that clearly meets the JOSS criteria — I am recommending **acceptance after minor revisions**. A small set of non-blocking issues follow.
>
> 1. **Broken PyPI badge link** — `README.md` line 3 has `https://pypi.org/projectpedonspei/` (should be `https://pypi.org/project/pedon/`). The image renders but the hyperlink 404s.
> 2. **Typo in paper** — `paper/paper.md` line 122: "soil hydralulic model parameters" → "soil hydraulic". Also same line: "in a academic context" → "in an academic context".
> 3. **Version drift between submission and release** — the review tag is `v0.0.11` but PyPI is on `0.0.12`. Consider either tagging a JOSS-specific release at the reviewed commit or updating the version metadata in the review issue so the archived artifact matches what reviewers install.
> 4. **HYPAGS `rosetta` optional extra** — `SoilSample.hypags()` is advertised in the paper alongside `.cosby()` but Rosetta v1/2/3 (mentioned in the README) requires the `pedon[rosetta]` extra (`httpx`). A short note in the README / paper about which optional extras unlock which PTFs would help users avoid import surprises.
>
> None of these block acceptance in my view. Happy to re-verify once the small revisions above are in. Thanks again for a clean and well-engineered submission.

Hi @gmarupilla,

I fixed your comments 1, 2 and 4 in https://github.com/martinvonk/pedon/commit/8673141377076d15cfd935badf4deec633abf396, https://github.com/martinvonk/pedon/commit/0e19404083c81313e45c29ceb590528e8d1b9626 and https://github.com/martinvonk/pedon/commit/ac9615a0fc10a5e3abf18a2fee1a926c00cd096f.

Regarding the verison drift I want to publish a new release (v0.1.0) just before the publication of the JOSS article. That way we have a nice stable release with all the comments from the reviewers processed.

Thanks again!
