### 2026-04-29 #10409 @tmweigand
> Hi [@tmweigand](https://github.com/tmweigand),
>
> Thank you for taking the time to look into this package and for sharing your detailed thoughts.
>
> I completely agree with your suggestions regarding expanded validation and broader scope. I have created GitHub issues to address the RETC validation tests ([martinvonk/pedon#32](https://github.com/martinvonk/pedon/issues/32)), the improved validation checks ([martinvonk/pedon#34](https://github.com/martinvonk/pedon/issues/34)), and the addition of hysteresis models ([martinvonk/pedon#33](https://github.com/martinvonk/pedon/issues/33)). These additions would substantially improve the Python package.
>
> Regarding your comment on performance:
>
> > * Solutions to variably saturated groundwater models can be quite difficult and require refined grids and efficient methods. Have you checked performance to ensure that `pedon` is a viable option for these cases as mentioned in the `Statement of Need?` I suspect the Python implementation may result in long simulation times for realistic systems.
>
> I want to clarify the intended scope of the package. `pedon` is not designed to be a numerical solver for Richards' equation or to run the groundwater simulations itself. Instead, its primary purpose is to generate the soil hydraulic properties required as input for those computationally heavy numerical solvers.
>
> For example, I frequently use `pedon` alongside [FloPy](https://github.com/modflowpy/flopy) to create inputs for [MODFLOW-USG Transport](https://www.gsienv.com/software/modflow-usg/modflow-usg/), which models 3D variably saturated flow. In workflows like this, including post-processing the results, `pedon` performs great (for academic research) in my experience.
>
> Does that help clarify how pedon fits into the modeling pipeline, or were you referring something else with regards to the performance of a specific function within the package? Let me know!

Thanks for the information. Are the `pedon` functions being called within the nonlinear solver iterations of MODFLOW-USG or are you collecting the parameters from `pedon` and passing those into existing functions within MODFLOW-USG? In other words, is `pedon` used for only pre-processing or is `pedon` coupled at runtime?

Providing an interface to use `pedon` at runtime would greatly improve the utility but that is also where my concern about performance lies.
