### 2026-04-30 #10409 @martinvonk
> Thanks for the information. Are the `pedon` functions being called within the nonlinear solver iterations of MODFLOW-USG or are you collecting the parameters from `pedon` and passing those into existing functions within MODFLOW-USG? In other words, is `pedon` used for only pre-processing or is `pedon` coupled at runtime?
>
Hi @tmweigand,

At the moment, pedon is used strictly for pre-processing. There is currently no runtime coupling possible with MODFLOW-USG Transport (MF-USGT).

Pre-processing is actually where pedon provides a lot of its value. Setting up inputs for MF-USGT can be quite tedious because users must format their data to match specific internal soil models, such as:
1. van Genuchten-Mualem
2. van Genuchten (for the soil water retention curve) combined with Brooks-Corey (for the hydraulic conductivity function)
3. Tabular formats for full flexibility in the retention and conductivity curves

I completely agree with you that runtime coupling would be a fantastic feature. Looking ahead, this is something we could eventually explore with MODFLOW 6 (the sister code to MF-USGT), which supports runtime Python coupling via the [MODFLOW API](https://doi.org/10.1016/j.envsoft.2021.105257). In my experience, this API works very well without causing big performance drops.

However, MODFLOW 6 does not [yet support solving Richards' equation](https://digitalcommons.unl.edu/civilengdiss/212/), though it is currently in development. Once that capability is officially released, pedon's architecture is flexible enough that it could certainly be adapted for runtime coupling. Until then, runtime integration remains outside the current viable scope of the package.
