### 2026-05-05 458bfde
update test with expected values
Co-authored-by: Copilot <copilot@github.com>
