+++ tests/test_soilmodel.py
@@ -4,9 +4,13 @@ from numpy import allclose, array, logspace
 import pedon as pe
 from pedon._typing import FloatArray
 
-h = -logspace(-2, 6, num=50)
+h = -logspace(-2, 6, num=9, dtype=float)
 
 theta = array([0.1, 0.2, 0.3, 0.4])
+
+
+def assert_close(actual: FloatArray, expected: FloatArray) -> None:
+    assert allclose(actual, expected, rtol=1e-12, atol=1e-12)
 
 
 @pytest.fixture
@@ -55,127 +59,416 @@ def hav() -> pe.soilmodel.Haverkamp:
 
 
 def test_theta_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
-    gen.theta(h=h)
+    expected = array(
+        [
+            0.42999674186266773,
+            0.4299590045910226,
+            0.42948737007237936,
+            0.4240392694915515,
+            0.3884681132693863,
+            0.3202485060610565,
+            0.25718963002041634,
+            0.20639799991988222,
+            0.166007528803122,
+        ]
+    )
+    assert_close(gen.theta(h=h), expected)
 
 
 def test_s_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
-    gen.s(h=h)
+    expected = array(
+        [
+            0.9999922425301613,
+            0.9999023918833871,
+            0.9987794525532842,
+            0.9858077845036941,
+            0.9011145554033008,
+            0.7386869191929917,
+            0.5885467381438484,
+            0.4676142855235291,
+            0.3714464971502905,
+        ]
+    )
+    assert_close(gen.s(h=h), expected)
 
 
 def test_k_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
-    gen.k(h=h)
+    expected = array(
+        [
+            3.2869753941674986,
+            2.1425970739706632,
+            1.0528919260290555,
+            0.25654138503050566,
+            0.011109108559820792,
+            9.373293594158181e-05,
+            5.47578373681202e-07,
+            3.0887429051900053e-09,
+            1.7373526209279636e-11,
+        ]
+    )
+    assert_close(gen.k(h=h), expected)
 
 
 def test_h_genuchten(gen: pe.SoilModel, theta: FloatArray = theta) -> None:
-    gen.h(theta=theta)
+    expected = array(
+        [244927639.31052694, 139271.6694845786, 1998.5290326524432, 61.672352811394724]
+    )
+    assert_close(gen.h(theta=theta), expected)
 
 
 def test_theta_brooks(bro: pe.SoilModel, h: FloatArray = h) -> None:
-    bro.theta(h=h)
+    expected = array(
+        [
+            0.43,
+            0.43,
+            0.43,
+            0.43,
+            0.0142,
+            0.010042,
+            0.01000042,
+            0.0100000042,
+            0.010000000042000001,
+        ]
+    )
+    assert_close(bro.theta(h=h), expected)
 
 
 def test_s_brooks(bro: pe.SoilModel, h: FloatArray = h) -> None:
-    bro.s(h=h)
+    expected = array([1.0, 1.0, 1.0, 1.0, 0.01, 0.0001, 1e-06, 1e-08, 1e-10])
+    assert_close(bro.s(h=h), expected)
 
 
 def test_k_brooks(bro: pe.SoilModel, h: FloatArray = h) -> None:
-    bro.k(h=h)
+    expected = array(
+        [
+            10.0,
+            10.0,
+            10.0,
+            10.0,
+            1e-07,
+            1.0000000000000003e-15,
+            9.999999999999998e-24,
+            1e-31,
+            1.0000000000000001e-39,
+        ]
+    )
+    assert_close(bro.k(h=h), expected)
 
 
[…]
