### 2026-05-06 martinvonk/pedon#33 @APeche
Here is a first and straightforward mathematical model for primary drying and wetting curves.

**Limitations**: This model is limited to the van Genuchten soil hydraulic functions (but pedon can be used for model conversion by (1) using the fitting routine or (2) the new conversion class). Also, due to the simplicity of the model, a pumping effect may occur when drying and wetting is cyclic (Bashir et al., 2015). Last, scanning curves must be calculated manually (e.g. within the numerical simulator). For that purpose, simple methods using linear functions or, for better numerical stability, Bézier-Splines may be used (Cremer et al., 2017). Also, the model by Parker & Lenhard (1987) may be used.

**Assumptions**: The 'ink-bottle-effect' (pores empty at larger suction than filling up) leads to a shift of the primary wetting curve compared with the primary drying curve. The van Genuchten shape parameters n and m remain similar for both, drying and wetting (Bashir et al., 2015). Also, no air entrappment is assumed - values for the residual and saturated water contents remain similar (Kool & Parker, 1987; Bashir et al., 2015). Note that, due to entrapped air, the saturated water content of the primary drying curve is generally larger than the saturated water content of the primary wetting curve (Rudiyanto et al., 2020).

**Methodology**: Set van Genuchten parameters for the primary drying curve. Adjust the van Genuchten $\alpha$ for the wetting process by
$\alpha_{drying} \cdot \xi = \alpha_{wetting}$
The value for $\xi$ may best be fitted. In broad calculations, it may be assumed as $\xi=2$. Values for $\xi$ for loam, silty sand silty loam can be set to $\xi_{loam}= 1.875$, $\xi_{silty\ sand}= 0.66$, $\xi_{silty\ loam}= 0.018$ (Bashir et a., 2015) [note that the shift is reversed in the last two cases - the 'ink-bottle-effect'-assumption hast o be adjusted accordingly].

References:
Bashir, R., Sharma, J., and Stefaniak, H. (2016). Effect of hysteresis of soil-water characteristic curves on infiltration under different climatic conditions. Canadian Geotechnical Journal, 53(2), 273–284. [https://doi.org/10.1139/cgj-2015-0004](https://www.google.com/search?q=https://doi.org/10.1139/cgj-2015-0004)
Cremer, C.J.M., A. Peche, L-B. Thiele, T. Graf, and I. Neuweiler (2017): A Bézier-Spline-based Model for the Simulation of Hysteresis in Variably Saturated Soil,  GU 04/2017, Poster Programme, Vienna, Austria
Kool, J. B., and Parker, J. C. (1987). Development and evaluation of closed-form expressions for hysteretic soil hydraulic properties. Water Resources Research, 23(1), 105–114. [https://doi.org/10.1029/WR023i001p00105](https://www.google.com/search?q=https://doi.org/10.1029/WR023i001p00105)
Parker, J.C., and Lenhard, R.J. 1987. A model for hysteretic constitutive relations governing multiphase ﬂow: 1. Saturation-pressure  elations. Water Resources Research, 23(12): 2187–2196. doi:10.1029/WR023i012p02187
Rudiyanto, Minasny, B., Shah, R. M., Setiawan, B. I., & van Genuchten, M. T. (2020). Simple functions for describing soil water retention and the unsaturated hydraulic conductivity from saturation to complete dryness. Journal of Hydrology, 588, Article 125041. https://doi.org/10.1016/j.jhydrol.2020.125041
