### 2026-05-06 c419124
first attempt at hysteresis
