### 2026-05-07 84c003b
make alpha_w a property
