+++ src/pedon/soil.py
@@ -869,7 +869,7 @@ class Soil:
             smn = sm
             sm = get_soilmodel(smn)
         else:
-            raise ValueError(
+            raise TypeError(
                 f"Argument must either be Type[SoilModel] | SoilModel | str,"
                 f"not {type(sm)}"
             )
@@ -920,7 +920,7 @@ class Soil:
             sm = get_soilmodel(smn)
 
         else:
-            raise ValueError(
+            raise TypeError(
                 f"Argument must either be Type[SoilModel] | SoilModel | str,"
                 f"not {type(sm)}"
             )
