### 2026-05-09 b018c9b
add test values for pedotransfer tests
