+++ src/pedon/__init__.py
@@ -1,5 +1,5 @@
 from . import soil as soil
-from . import soilmodel as soilmodel  # noqa: F401
+from . import soilmodel as soilmodel
 from ._params import get_params as get_params
 from ._version import __version__ as __version__
 from ._version import show_versions as show_versions
