+++ src/pedon/soil.py
@@ -3,7 +3,7 @@ from bisect import bisect_right
 from dataclasses import dataclass, field
 from logging import getLogger
 from pathlib import Path
-from typing import Literal, Self, Type
+from typing import Literal, Self
 
 from numpy import abs as npabs
 from numpy import (
@@ -31,8 +31,7 @@ logger = getLogger(__name__)
 
 @dataclass
 class SoilSample:
-    """
-    A container for measured soil properties and in-situ soil hydraulic data, with
+    """A container for measured soil properties and in-situ soil hydraulic data, with
     convenience routines to predict hydraulic parameters using a range of pedo-
     transfer functions and empirical models.
 
@@ -70,6 +69,7 @@ class SoilSample:
         - Verify units before use as different methods expect specific conventions.
         - Methods return soil hydraulic model instances (e.g., Genuchten, Brooks).
         - Some methods make external API calls (rosetta) or use empirical relationships.
+
     """
 
     sand_p: float | None = None  # sand %
@@ -117,7 +117,7 @@ class SoilSample:
 
     def fit(
         self,
-        sm: Type[SoilModel],
+        sm: type[SoilModel],
         pbounds: DataFrame | None = None,
         weights: FloatArray | float = 1.0,
         W1: float = 0.1,
@@ -126,8 +126,7 @@ class SoilSample:
         silent: bool = True,
         **kwargs,
     ) -> SoilModel:
-        """
-        Fit the provided SoilModel (e.g., van Genuchten, Brooks-Corey class) to the
+        """Fit the provided SoilModel (e.g., van Genuchten, Brooks-Corey class) to the
         stored measurements (theta, k, h) using nonlinear least squares. If
         pbounds is not provided, default parameter bounds are retrieved for the
         requested model name. The objective combines water retention and log10(k)
@@ -163,6 +162,7 @@ class SoilSample:
             - Requires theta and k (and optionally h) to be set.
             - If k_s is provided it will be fixed during optimization.
             - Input/outputs and bounds are expected in the units used by the soil model.
+
         """
         theta = self.theta
         N = len(theta)
@@ -193,7 +193,8 @@ class SoilSample:
         def get_diff(p: FloatArray) -> FloatArray:
             """Objective function for least squares optimization. Computes the difference
             between measured and model-predicted theta and log10(k) values, applying
-            the specified weights and scaling factors."""
+            the specified weights and scaling factors.
+            """
             est_pars = dict(zip(pbounds.index, p))
             if k_s is not None:
                 est_pars["k_s"] = k_s
@@ -224,8 +225,8 @@ class SoilSample:
 
     def wosten(self, ts: bool = False) -> Genuchten:
         """Wosten et al (1999) - Development and use of a database of hydraulic
-        properties of European soils"""
-
+        properties of European soils
+        """
         topsoil = 1.0 * ts
 
         theta_s = (
@@ -314,8 +315,8 @@ class SoilSample:
 
     def wosten_sand(self, ts: bool = False) -> Genuchten:
         """Wosten et al. (2001) - Waterretentie- en doorlatendheidskarakteristieken
-        van boven- en ondergronden in Nederland: de Staringreeks"""
-
+        van boven- en ondergronden in Nederland: de Staringreeks
+        """
         topsoil = 1.0 * ts
         theta_s = (
             -35.7
@@ -376,7 +377,7 @@ class SoilSample:
             + 2.364 * self.silt_p**-1
             + 1.014 * log(self.silt_p)
         )
-        l = min(max(2 * (exp(l_) - 1) / (exp(l_) + 1), -2.0), 2.0)  # noqa: E741
+        l = min(max(2 * (exp(l_) - 1) / (exp(l_) + 1), -2.0), 2.0)
 
         return Genuchten(
[…]
