### 2026-05-10 0db5873
test docstrings with ruff
