+++ src/pedon/soil.py
@@ -31,9 +31,10 @@ logger = getLogger(__name__)
 
 @dataclass
 class SoilSample:
-    """A container for measured soil properties and in-situ soil hydraulic data, with
-    convenience routines to predict hydraulic parameters using a range of pedo-
-    transfer functions and empirical models.
+    """A container for measured soil properties and in-situ soil hydraulic data.
+
+    Class provides convenience routines to predict hydraulic parameters using
+    a range of pedotransfer functions and empirical models.
 
     Attributes
     ----------
@@ -95,7 +96,7 @@ class SoilSample:
     )  # moisture content measurement
 
     def from_staring(self, name: str, year: str = "2018") -> "SoilSample":
-        """Get properties and measurements from Staring series"""
+        """Get properties and measurements from Staring series."""
         if year not in ("2001", "2018"):
             raise ValueError(
                 f"No Staring series available for year '{year}'"
@@ -126,7 +127,9 @@ class SoilSample:
         silent: bool = True,
         **kwargs,
     ) -> SoilModel:
-        """Fit the provided SoilModel (e.g., van Genuchten, Brooks-Corey class) to the
+        """Fit the provided SoilModel to the measurements.
+
+        Fit the provided SoilModel (e.g., van Genuchten, Brooks-Corey class) to the
         stored measurements (theta, k, h) using nonlinear least squares. If
         pbounds is not provided, default parameter bounds are retrieved for the
         requested model name. The objective combines water retention and log10(k)
@@ -191,9 +194,11 @@ class SoilSample:
             logger.debug(f"Computed W2: {W2}")
 
         def get_diff(p: FloatArray) -> FloatArray:
-            """Objective function for least squares optimization. Computes the difference
-            between measured and model-predicted theta and log10(k) values, applying
-            the specified weights and scaling factors.
+            """Objective function for least squares optimization.
+
+            Computes the difference between measured and model-predicted
+            theta and log10(k) values, applying the specified weights and
+            scaling factors.
             """
             est_pars = dict(zip(pbounds.index, p))
             if k_s is not None:
@@ -224,9 +229,7 @@ class SoilSample:
         return sm(**opt_pars)
 
     def wosten(self, ts: bool = False) -> Genuchten:
-        """Wosten et al (1999) - Development and use of a database of hydraulic
-        properties of European soils
-        """
+        """Wosten et al (1999) - Development and use of a database of hydraulic properties of European soils."""
         topsoil = 1.0 * ts
 
         theta_s = (
@@ -314,8 +317,10 @@ class SoilSample:
         )
 
     def wosten_sand(self, ts: bool = False) -> Genuchten:
-        """Wosten et al. (2001) - Waterretentie- en doorlatendheidskarakteristieken
-        van boven- en ondergronden in Nederland: de Staringreeks
+        """Pedotransfer function for sandy soils.
+
+        Wosten et al. (2001) - Waterretentie- en doorlatendheidskarakteristieken
+        van boven- en ondergronden in Nederland: de Staringreeks.
         """
         topsoil = 1.0 * ts
         theta_s = (
@@ -377,7 +382,7 @@ class SoilSample:
             + 2.364 * self.silt_p**-1
             + 1.014 * log(self.silt_p)
         )
-        l = min(max(2 * (exp(l_) - 1) / (exp(l_) + 1), -2.0), 2.0)
+        l = min(max(2 * (exp(l_) - 1) / (exp(l_) + 1), -2.0), 2.0)  # noqa: E741
 
         return Genuchten(
             k_s=round(k_s, 4),
@@ -389,8 +394,10 @@ class SoilSample:
         )
 
[…]
