+++ src/pedon/soilmodel.py
@@ -10,7 +10,7 @@ from ._typing import FloatArray, SoilModelNames
 
 @runtime_checkable
 class SoilModel(Protocol):
-    """Protocol for soil models
+    """Protocol for soil models.
 
     This protocol defines the interface for custom soil models. To create a custom
     soil model, implement all methods defined below. Your class will automatically
@@ -50,19 +50,19 @@ class SoilModel(Protocol):
     """
 
     def theta(self, h: FloatArray) -> FloatArray:
-        """Calculate soil moisture content (water content) from
-        pressure head h.
-        """
+        """Calculate soil moisture content (water content) from pressure head h."""
         ...
 
     def s(self, h: FloatArray) -> FloatArray:
         """Calculate effective saturation from pressure head h.
+
         Effective saturation is normalized between 0 (dry) and 1 (saturated).
         """
         ...
 
     def k_r(self, h: FloatArray, s: FloatArray | None = None) -> FloatArray:
         """Calculate relative permeability (or relative hydraulic conductivity).
+
         Relative permeability is normalized between 0 (dry) and 1 (saturated).
         Can be calculated from either pressure head h or saturation s.
         """
@@ -83,10 +83,10 @@ class SoilModel(Protocol):
 
 @dataclass
 class Genuchten:
-    """Mualem-van Genuchten Soil Model
+    """Mualem-van Genuchten Soil Model.
 
     van Genuchten, M. Th. (1980) - A Closed-form Equation for Predicting the
-    Hydraulic Conductivity of Unsaturated Soil
+    Hydraulic Conductivity of Unsaturated Soil.
     """
 
     k_s: float
@@ -130,9 +130,9 @@ class Genuchten:
 
 @dataclass
 class Brooks:
-    """Brooks and Corey Soil Model
+    """Brooks and Corey Soil Model.
 
-    Brooks, R.H. and Corey, A.T. (1964) - Hydraulic Properties of Porous Media
+    Brooks, R.H. and Corey, A.T. (1964) - Hydraulic Properties of Porous Media.
     """
 
     k_s: float
@@ -199,7 +199,7 @@ class Brooks:
 
 @dataclass
 class Haverkamp:
-    """Haverkamp Soil Model
+    """Haverkamp Soil Model.
 
     Haverkamp, R., Vauclin, M., Touma, J., Wierenga, P. J., & Vachaud, G. (1977).
     A comparison of numerical simulation models for one-dimensional infiltration.
@@ -241,12 +241,12 @@ class Haverkamp:
 
 @dataclass
 class Gardner:
-    """Gardner(-Kozeny) Soil Model
+    """Gardner(-Kozeny) Soil Model.
 
     Gardner, W.H. (1958) - Some steady-state solutions of the unsaturated
-    moisture flow equation with application to evaporation from soils
+    moisture flow equation with application to evaporation from soils.
     Bakker and Nieber (2009) - Damping of Sinusoidal Surface Flux Fluctuations
-    with Soil Depth
+    with Soil Depth.
     """
 
     k_s: float
@@ -278,7 +278,7 @@ class Gardner:
 
 @dataclass
 class Rucker:
-    """Gardner(-Rucker) Soil Model
+    """Gardner(-Rucker) Soil Model.
 
     Gardner, W.H. (1958) - Some steady-state solutions of the unsaturated
     moisture flow equation with application to evaporation from soils
@@ -322,9 +322,9 @@ class Rucker:
 
 @dataclass
 class Panday:
-    """Panday Soil Model (MODFLOW-USG)
+    """Panday Soil Model as used in MODFLOW-USG Tranport and MODFLOW's UZF package.
 
-    Panday, S. - USG-Transport: Transport and other Enhancements to MODFLOW-USG
[…]
