### 2026-05-10 2a357ae
add a lot of tests
