+++ docs/conf.py
@@ -1,3 +1,4 @@
+"""Sphinx configuration for pedon documentation."""
 # Configuration file for the Sphinx documentation builder.
 #
 # For the full list of built-in configuration values, see the documentation:
