+++ tests/test_soilmodel.py
@@ -1,3 +1,5 @@
+"""Test soilmodels."""
+
 import pytest
 from numpy import allclose, array, logspace
 
@@ -10,36 +12,43 @@ theta = array([0.1, 0.2, 0.3, 0.4])
 
 
 def assert_close(actual: FloatArray, expected: FloatArray) -> None:
+    """Assert if two arrays are close within a reasonable tolerance."""
     assert allclose(actual, expected, rtol=1e-12, atol=1e-12)
 
 
 @pytest.fixture
 def gen() -> pe.SoilModel:
+    """Fixture for a van Genuchten soil model with specific parameters for testing."""
     return pe.Genuchten(k_s=10, theta_r=0.01, theta_s=0.43, alpha=0.02, n=1.1, l=0.5)
 
 
 @pytest.fixture
 def bro() -> pe.SoilModel:
+    """Fixture for a Brooks-Corey soil model with specific parameters for testing."""
     return pe.Brooks(k_s=10, theta_r=0.01, theta_s=0.43, h_b=10, l=2)
 
 
 @pytest.fixture
 def sor() -> pe.SoilModel:
+    """Fixture for a Panday soil model with specific parameters for testing."""
     return pe.Panday(k_s=10, theta_r=0.01, theta_s=0.43, alpha=0.02, beta=1.1, brook=3)
 
 
 @pytest.fixture
 def gar() -> pe.SoilModel:
+    """Fixture for a Gardner soil model with specific parameters for testing."""
     return pe.Gardner(k_s=10, theta_s=0.43, c=0.02, m=1.1)
 
 
 @pytest.fixture
 def gr() -> pe.SoilModel:
+    """Fixture for a Rucker soil model with specific parameters for testing."""
     return pe.Rucker(k_s=10, theta_r=0.01, theta_s=0.43, c=0.02, m=1.1)
 
 
 @pytest.fixture
 def gg() -> pe.SoilModel:
+    """Fixture for a Genuchten-Gardner soil model with specific parameters for testing."""
     return pe.GenuchtenGardner(
         k_s=10, theta_s=0.43, theta_r=0.01, c=0.02, alpha=0.04, n=1.4
     )
@@ -47,6 +56,7 @@ def gg() -> pe.SoilModel:
 
 @pytest.fixture
 def hav() -> pe.soilmodel.Haverkamp:
+    """Fixture for a Haverkamp soil model with specific parameters for testing."""
     # Example parameters similar in style to other model tests
     return pe.Haverkamp(
         k_s=10.0,
@@ -59,6 +69,7 @@ def hav() -> pe.soilmodel.Haverkamp:
 
 
 def test_theta_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
+    """Test water content calculation for the van Genuchten model."""
     expected = array(
         [
             0.42999674186266773,
@@ -76,6 +87,7 @@ def test_theta_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
 
 
 def test_s_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
+    """Test degree of saturation calculation for the van Genuchten model."""
     expected = array(
         [
             0.9999922425301613,
@@ -93,6 +105,7 @@ def test_s_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
 
 
 def test_k_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
+    """Test hydraulic conductivity calculation for the van Genuchten model."""
     expected = array(
         [
             3.2869753941674986,
[…]
