### 2026-05-10 3383f5b
add flopy to documentation dependencies
