### 2026-05-10 f6f7b5e
create mock test for rosetta
