+++ tests/test_ptf.py
@@ -4,12 +4,12 @@ import pytest
 
 import pedon as pe
 
-REL = 1e-5
+REL = 1e-3
 
 
 @pytest.fixture
 def ss() -> pe.soil.SoilSample:
-    """Fixutre for a soil sample with specific properties for testing pedotransfer functions."""
+    """Fixture for a soil sample with specific properties for testing pedotransfer functions."""
     return pe.soil.SoilSample(
         sand_p=50.0, silt_p=10.0, clay_p=40.0, rho=1.5, om_p=20.0, m50=150.0
     )
@@ -66,20 +66,55 @@ def test_cosby(ss: pe.soil.SoilSample) -> None:
 
 
 def test_rosetta(ss: pe.soil.SoilSample) -> None:
-    """Test ROSETTA pedotransfer function."""
-    sm = ss.rosetta()
-    assert isinstance(sm, pe.Genuchten)
-    assert sm.k_s == pytest.approx(13.7019747459841, rel=REL)
-    assert sm.theta_r == pytest.approx(0.1142278836409842, rel=REL)
-    assert sm.theta_s == pytest.approx(0.42963731868993743, rel=REL)
-    assert sm.alpha == pytest.approx(0.013518839874177326, rel=REL)
-    assert sm.n == pytest.approx(1.274931762885784, rel=REL)
-    assert sm.l == pytest.approx(0.5, rel=REL)
-    assert sm.m == pytest.approx(0.2156442963374613, rel=REL)
+    """Test ROSETTA pedotransfer function with mocked HTTP response."""
+    from unittest.mock import Mock, patch
+
+    # Mock response data matching the structure returned by the ROSETTA API
+    mock_response = Mock()
+    mock_response.is_error = False
+    # van_genuchten_params: [theta_r, theta_s, log10(alpha), log10(n), log10(k_s)]
+    mock_response.json.return_value = {
+        "van_genuchten_params": [
+            [
+                0.1142278836409842,  # theta_r
+                0.42963731868993743,  # theta_s
+                -1.8689880666646325,  # log10(alpha)
+                0.10546903898903125,  # log10(n)
+                1.1367547932846525,  # log10(k_s)
+            ]
+        ]
+    }
+
+    with patch("httpx.post", return_value=mock_response) as mock_post:
+        sm = ss.rosetta()
+
+        # Verify the API was called with correct parameters
+        assert mock_post.called
+        call_args = mock_post.call_args
+        assert "rosetta/3" in call_args[0][0]  # Default version is 3
+
+        # Verify response is parsed and mapped correctly
+        assert isinstance(sm, pe.Genuchten)
+        assert sm.k_s == pytest.approx(13.7019747459841, rel=REL)
+        assert sm.theta_r == pytest.approx(0.1142278836409842, rel=REL)
+        assert sm.theta_s == pytest.approx(0.42963731868993743, rel=REL)
+        assert sm.alpha == pytest.approx(0.013518839874177326, rel=REL)
+        assert sm.n == pytest.approx(1.274931762885784, rel=REL)
+        assert sm.l == pytest.approx(0.5, rel=REL)
+        assert sm.m == pytest.approx(0.2156442963374613, rel=REL)
 
 
 def test_rosetta_invalidpercentage(ss: pe.soil.SoilSample) -> None:
-    """Test that ROSETTA raises ValueError for invalid soil percentages."""
-    ss.sand_p = 10
-    with pytest.raises(ValueError):
-        _ = ss.rosetta()
+    """Test that ROSETTA raises ValueError when API returns None values."""
+    from unittest.mock import Mock, patch
+
+    mock_response = Mock()
+    mock_response.is_error = False
+    # Response with None values in van_genuchten_params
[…]
