### 2026-05-25 1346c47
fix a lot of type checking issues
