+++ src/pedon/soil.py
@@ -1,11 +1,10 @@
 """Soil sample class and pedotransfer functions."""
 
-# type: ignore
 from bisect import bisect_right
 from dataclasses import dataclass, field
 from logging import getLogger
 from pathlib import Path
-from typing import Literal, Self
+from typing import Literal, Self, cast
 
 from numpy import abs as npabs
 from numpy import (
@@ -20,12 +19,13 @@ from numpy import (
     log10,
     multiply,
     ndarray,
+    asarray,
 )
-from pandas import DataFrame, isna, read_csv
+from pandas import Series, DataFrame, isna, read_csv
 from scipy.optimize import fixed_point, least_squares
 
 from ._params import get_params
-from ._typing import FloatArray
+from ._typing import FloatArray, SoilModelNames
 from .soilmodel import Brooks, Genuchten, SoilModel, get_soilmodel
 
 logger = getLogger(__name__)
@@ -106,16 +106,16 @@ class SoilSample:
             )
         path = Path(__file__).parent / "datasets/soilsamples.csv"
         properties = read_csv(path, delimiter=";")
-        staring_properties = properties[
-            properties["source"] == f"Staring_{year}"
-        ].set_index("name")
+        staring_properties = properties.query(f"source == 'Staring_{year}'").set_index(
+            "name"
+        )
 
-        self.silt_p = staring_properties.loc[name, "silt_p"]
-        self.clay_p = staring_properties.loc[name, "clay_p"]
-        self.om_p = staring_properties.loc[name, "om_p"]
-        self.m50 = staring_properties.loc[name, "m50"]
+        self.silt_p = float(staring_properties.at[name, "silt_p"])
+        self.clay_p = float(staring_properties.at[name, "clay_p"])
+        self.om_p = float(staring_properties.at[name, "om_p"])
+        self.m50 = float(staring_properties.at[name, "m50"])
         if year == "2001":
-            self.rho = staring_properties.loc[name, "rho"]
+            self.rho = float(staring_properties.at[name, "rho"])
         return self
 
     def fit(
@@ -169,13 +169,13 @@ class SoilSample:
             - Input/outputs and bounds are expected in the units used by the soil model.
 
         """
-        theta = self.theta
+        theta = asarray(self.theta, dtype=float)
         N = len(theta)
-        k = self.k
+        k = asarray(self.k, dtype=float)
         M = N + len(k)
 
         if pbounds is None:
-            pbounds = get_params(sm.__name__)
+            pbounds = get_params(sm)
             if k_s is not None:
                 pbounds = pbounds.drop("k_s")
             else:
@@ -184,8 +184,11 @@ class SoilSample:
             pbounds.loc["theta_s", "p_ini"] = max(theta)
             pbounds.loc["theta_s", "p_max"] = max(theta) + 0.02
 
-        if isinstance(weights, float):
-            weights = full(M, weights)
+        weights = (
+            full(M, weights)
+            if isinstance(weights, float)
+            else asarray(weights, dtype=float)
+        )
 
         if W2 is None:
             W2 = (
@@ -202,6 +205,7 @@ class SoilSample:
             theta and log10(k) values, applying the specified weights and
             scaling factors.
             """
+            p = asarray(p, dtype=float)
             est_pars = dict(zip(pbounds.index, p))
             if k_s is not None:
                 est_pars["k_s"] = k_s
@@ -211,7 +215,7 @@ class SoilSample:
             diff = append(weights[0:N] * theta_diff, weights[N:M] * W1 * W2 * k_diff)
             return diff
 
-        kwargs = {"jac": "3-point", "x_scale": "jac"} | (kwargs or {})
+        kwargs_merged: dict = {"jac": "3-point", "x_scale": "jac"} | (kwargs or {})
[…]
