### 2026-05-25 bed2c82
final type fixes
