### 2026-05-25 0d6eaab
small bug fix in soil.py
