### 2026-05-25 c342a5d
create plot.py file for plotting
