### 2026-05-25 62f064b
update notebooks accordinglhy
