### 2026-05-25 48f4fc0
create conftest.py for fixtures
