### 2026-05-25 e2379b4
fix copilot comments
