+++ pyproject.toml
@@ -39,7 +39,6 @@ pytesting = [
         "pytest-mpl",
         "pedon[rosetta]",
 ]
-coveraging = ["coverage"]
 documentation = [
         "sphinx>=8.2",
         "myst-parser",
@@ -47,8 +46,9 @@ documentation = [
         "ipykernel",
         "pandoc",
         "flopy",
+        "pedon[rosetta]",
 ]
-dev = ["pedon[ruffing,typing,pytesting,documentation]", "tox-uv"]
+dev = ["pedon[documentation,rosetta,ruffing,typing,pytesting]", "tox-uv"]
 
 
 [tool.setuptools]
