### 2026-05-25 1e513a4
update all docstrings with parameter descriptions
