+++ src/pedon/soilmodel.py
@@ -88,8 +88,36 @@ class SoilModel(Protocol):
 class Genuchten:
     """Mualem-van Genuchten Soil Model.
 
+    Parameters
+    ----------
+    k_s: float
+        Saturated hydraulic conductivity [L/T]
+    theta_r: float
+        Residual soil water content [-]
+    theta_s: float
+        Saturated soil water content [-]
+    n: float
+        Dimensionless measure of the pore-size distribution
+    alpha: float
+        Empirical shape parameter that is physically related
+        to the inverse of the air-entry pressure [1/L]
+    l: float, optional
+        Pore-connectivity or tortuosity parameter [-]
+        Default value 0.5 is commonly used and is based on the Mualem model.
+
+    Attributes
+    ----------
+    m: float, optional
+        Calculated from n as m = 1 - 1/n [-] (Mualem restriction)
+
+    References
+    ----------
     van Genuchten, M. Th. (1980) - A Closed-form Equation for Predicting the
     Hydraulic Conductivity of Unsaturated Soil.
+
+    Mualem, Y. (1976). A new model for predicting the hydraulic conductivity
+    of unsaturated porous media.
+
     """
 
     k_s: float
@@ -136,7 +164,23 @@ class Genuchten:
 class Brooks:
     """Brooks and Corey Soil Model.
 
+    Parameters
+    ----------
+    k_s: float
+        Saturated hydraulic conductivity [L/T]
+    theta_r: float
+        Residual soil water content [-]
+    theta_s: float
+        Saturated soil water content [-]
+    h_b: float
+        Bubbling pressure or air-entry suction [L]
+    l: float
+        Pore-size distribution index [-]
+
+    References
+    ----------
     Brooks, R.H. and Corey, A.T. (1964) - Hydraulic Properties of Porous Media.
+
     """
 
     k_s: float
@@ -205,8 +249,29 @@ class Brooks:
 class Haverkamp:
     """Haverkamp Soil Model.
 
+    Parameters
+    ----------
+    k_s: float
+        Saturated hydraulic conductivity [L/T]
+    theta_r: float
+        Residual soil water content [-]
+    theta_s: float
+        Saturated soil water content [-]
+    alpha: float
+        Empirical constant for the soil water retention curve [-]
+    beta: float
+        Empirical exponent for the soil water retention curve.
+        In this implementation, it is also used as the exponent
+        for the hydraulic conductivity curve (assuming B = beta) [-]
+    a: float
+        Empirical constant for the hydraulic conductivity curve
+        (Often denoted as 'A' in literature)
+
+    References
+    ----------
     Haverkamp, R., Vauclin, M., Touma, J., Wierenga, P. J., & Vachaud, G. (1977).
     A comparison of numerical simulation models for one-dimensional infiltration.
+
     """
 
     k_s: float
@@ -247,10 +312,26 @@ class Haverkamp:
 class Gardner:
     """Gardner(-Kozeny) Soil Model.
 
-    Gardner, W.H. (1958) - Some steady-state solutions of the unsaturated
+    Parameters
+    ----------
+    k_s: float
+        Saturated hydraulic conductivity [L/T]
+    theta_s: float
+        Saturated soil water content [-]
+    m: float
+        Empirical exponent parameter for the soil water retention curve [1/L]
+    c: float
+        Empirical exponent parameter for the hydraulic conductivity curve [1/L]
+        (Often denoted as alpha in the classic Gardner exponential model)
+
+    References
+    ----------
+    Gardner, W. H. (1958). Some steady-state solutions of the unsaturated
     moisture flow equation with application to evaporation from soils.
[…]
