### 2026-05-25 a40049f
create test for roundtrip
