+++ tests/test_soilmodel.py
@@ -16,53 +16,7 @@ def assert_close(actual: FloatArray, expected: FloatArray) -> None:
     assert allclose(actual, expected, rtol=1e-8, atol=1e-8)
 
 
-@pytest.fixture
-def bro() -> pe.SoilModel:
-    """Fixture for a Brooks-Corey soil model with specific parameters for testing."""
-    return pe.Brooks(k_s=10, theta_r=0.01, theta_s=0.43, h_b=10, l=2)
-
-
-@pytest.fixture
-def sor() -> pe.SoilModel:
-    """Fixture for a Panday soil model with specific parameters for testing."""
-    return pe.Panday(k_s=10, theta_r=0.01, theta_s=0.43, alpha=0.02, beta=1.1, brook=3)
-
-
-@pytest.fixture
-def gar() -> pe.SoilModel:
-    """Fixture for a Gardner soil model with specific parameters for testing."""
-    return pe.Gardner(k_s=10, theta_s=0.43, c=0.02, m=1.1)
-
-
-@pytest.fixture
-def gr() -> pe.SoilModel:
-    """Fixture for a Rucker soil model with specific parameters for testing."""
-    return pe.Rucker(k_s=10, theta_r=0.01, theta_s=0.43, c=0.02, m=1.1)
-
-
-@pytest.fixture
-def gg() -> pe.SoilModel:
-    """Fixture for a Genuchten-Gardner soil model with specific parameters for testing."""
-    return pe.GenuchtenGardner(
-        k_s=10, theta_s=0.43, theta_r=0.01, c=0.02, alpha=0.04, n=1.4
-    )
-
-
-@pytest.fixture
-def hav() -> pe.soilmodel.Haverkamp:
-    """Fixture for a Haverkamp soil model with specific parameters for testing."""
-    # Example parameters similar in style to other model tests
-    return pe.Haverkamp(
-        k_s=10.0,
-        theta_r=0.01,
-        theta_s=0.43,
-        alpha=0.02,
-        beta=1.2,
-        a=0.5,
-    )
-
-
-def test_theta_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
+def test_theta_genuchten(genuchten: pe.SoilModel, h: FloatArray = h) -> None:
     """Test water content calculation for the van Genuchten model."""
     expected = array(
         [
@@ -77,10 +31,10 @@ def test_theta_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
             0.166007528803122,
         ]
     )
-    assert_close(gen.theta(h=h), expected)
+    assert_close(genuchten.theta(h=h), expected)
 
 
-def test_s_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
+def test_s_genuchten(genuchten: pe.SoilModel, h: FloatArray = h) -> None:
     """Test degree of saturation calculation for the van Genuchten model."""
     expected = array(
         [
@@ -95,10 +49,10 @@ def test_s_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
             0.3714464971502905,
         ]
     )
-    assert_close(gen.s(h=h), expected)
+    assert_close(genuchten.s(h=h), expected)
 
 
-def test_k_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
+def test_k_genuchten(genuchten: pe.SoilModel, h: FloatArray = h) -> None:
     """Test hydraulic conductivity calculation for the van Genuchten model."""
     expected = array(
         [
@@ -113,18 +67,18 @@ def test_k_genuchten(gen: pe.SoilModel, h: FloatArray = h) -> None:
             1.7373526209279636e-11,
         ]
     )
-    assert_close(gen.k(h=h), expected)
[…]
