### 2026-05-25 dcb1c18
fix inverse for rucker
