### 2026-05-25 50da4ce
add tests for genuchtenkool
