+++ tests/conftest.py
@@ -14,33 +14,41 @@ def genuchten() -> pe.SoilModel:
 @fixture
 def brooks() -> pe.SoilModel:
     """Fixture for a Brooks-Corey soil model with specific parameters for testing."""
-    return pe.Brooks(k_s=10, theta_r=0.01, theta_s=0.43, h_b=10, l=2)
+    return pe.Brooks(k_s=10.0, theta_r=0.01, theta_s=0.43, h_b=10, l=2)
 
 
 @fixture
 def panday() -> pe.SoilModel:
     """Fixture for a Panday soil model with specific parameters for testing."""
-    return pe.Panday(k_s=10, theta_r=0.01, theta_s=0.43, alpha=0.02, beta=1.1, brook=3)
+    return pe.Panday(
+        k_s=10.0, theta_r=0.01, theta_s=0.43, alpha=0.02, beta=1.1, brook=3
+    )
 
 
 @fixture
 def gardner() -> pe.SoilModel:
     """Fixture for a Gardner soil model with specific parameters for testing."""
-    return pe.Gardner(k_s=10, theta_s=0.43, c=0.02, m=1.1)
+    return pe.Gardner(k_s=10.0, theta_s=0.43, c=0.02, m=1.1)
 
 
 @fixture
 def rucker() -> pe.SoilModel:
     """Fixture for a Rucker soil model with specific parameters for testing."""
-    return pe.Rucker(k_s=10, theta_r=0.01, theta_s=0.43, c=0.02, m=1.1)
+    return pe.Rucker(k_s=10.0, theta_r=0.01, theta_s=0.43, c=0.02, m=1.1)
 
 
 @fixture
 def genuchtengardner() -> pe.SoilModel:
     """Fixture for a Genuchten-Gardner soil model with specific parameters for testing."""
     return pe.GenuchtenGardner(
-        k_s=10, theta_s=0.43, theta_r=0.01, c=0.02, alpha=0.04, n=1.4
+        k_s=10.0, theta_s=0.43, theta_r=0.01, c=0.02, alpha=0.04, n=1.4
     )
+
+
+@fixture
+def fredlund() -> pe.SoilModel:
+    """Fixture for a Fredlund soil model with specific parameters for testing."""
+    return pe.Fredlund(k_s=10.0, theta_s=0.43, a=0.5, n=1.4, m=1.1)
 
 
 @fixture
@@ -55,6 +63,7 @@ def haverkamp() -> pe.soilmodel.Haverkamp:
         beta=1.2,
         a=0.5,
     )
+
 
 @fixture
 def genuchtenkool() -> pe.SoilModel:
