+++ tests/test_soilmodel.py
@@ -16,7 +16,7 @@ def assert_close(actual: FloatArray, expected: FloatArray) -> None:
     assert allclose(actual, expected, rtol=1e-8, atol=1e-8)
 
 
-def test_theta_genuchten(genuchten: pe.SoilModel, h: FloatArray = h) -> None:
+def test_theta_genuchten(genuchten: pe.SoilModel) -> None:
     """Test water content calculation for the van Genuchten model."""
     expected = array(
         [
@@ -34,7 +34,7 @@ def test_theta_genuchten(genuchten: pe.SoilModel, h: FloatArray = h) -> None:
     assert_close(genuchten.theta(h=h), expected)
 
 
-def test_s_genuchten(genuchten: pe.SoilModel, h: FloatArray = h) -> None:
+def test_s_genuchten(genuchten: pe.SoilModel) -> None:
     """Test degree of saturation calculation for the van Genuchten model."""
     expected = array(
         [
@@ -52,7 +52,7 @@ def test_s_genuchten(genuchten: pe.SoilModel, h: FloatArray = h) -> None:
     assert_close(genuchten.s(h=h), expected)
 
 
-def test_k_genuchten(genuchten: pe.SoilModel, h: FloatArray = h) -> None:
+def test_k_genuchten(genuchten: pe.SoilModel) -> None:
     """Test hydraulic conductivity calculation for the van Genuchten model."""
     expected = array(
         [
@@ -78,7 +78,7 @@ def test_h_genuchten(genuchten: pe.SoilModel, theta: FloatArray = theta) -> None
     assert_close(genuchten.h(theta=theta), expected)
 
 
-def test_theta_brooks(brooks: pe.SoilModel, h: FloatArray = h) -> None:
+def test_theta_brooks(brooks: pe.SoilModel) -> None:
     """Test water content calculation for the Brooks-Corey model."""
     expected = array(
         [
@@ -96,13 +96,13 @@ def test_theta_brooks(brooks: pe.SoilModel, h: FloatArray = h) -> None:
     assert_close(brooks.theta(h=h), expected)
 
 
-def test_s_brooks(brooks: pe.SoilModel, h: FloatArray = h) -> None:
+def test_s_brooks(brooks: pe.SoilModel) -> None:
     """Test degree of saturation calculation for the Brooks-Corey model."""
     expected = array([1.0, 1.0, 1.0, 1.0, 0.01, 0.0001, 1e-06, 1e-08, 1e-10])
     assert_close(brooks.s(h=h), expected)
 
 
-def test_k_brooks(brooks: pe.SoilModel, h: FloatArray = h) -> None:
+def test_k_brooks(brooks: pe.SoilModel) -> None:
     """Test hydraulic conductivity calculation for the Brooks-Corey model."""
     expected = array(
         [
@@ -126,7 +126,7 @@ def test_h_brooks(brooks: pe.SoilModel, theta: FloatArray = theta) -> None:
     assert_close(brooks.h(theta=theta), expected)
 
 
-def test_theta_panday(panday: pe.SoilModel, h: FloatArray = h) -> None:
+def test_theta_panday(panday: pe.SoilModel) -> None:
     """Test water content calculation for the Panday model."""
     expected = array(
         [
@@ -144,7 +144,7 @@ def test_theta_panday(panday: pe.SoilModel, h: FloatArray = h) -> None:
     assert_close(panday.theta(h=h), expected)
 
 
-def test_s_panday(panday: pe.SoilModel, h: FloatArray = h) -> None:
+def test_s_panday(panday: pe.SoilModel) -> None:
     """Test degree of saturation calculation for the Panday model."""
     expected = array(
         [
[…]
