### 2026-05-25 da1122a
create test for get_soilmodel
