### 2026-05-26 eb7cb72
add Kosugi soilmodel
