+++ src/pedon/soilmodel.py
@@ -5,9 +5,9 @@ from typing import Protocol, runtime_checkable
 
 import matplotlib.pyplot as plt
 from numpy import abs as npabs
-from numpy import asarray, exp, full, linspace, log, log10
+from numpy import asarray, exp, full, linspace, log, log10, sqrt
 from scipy.integrate import trapezoid
-from scipy.special import lambertw
+from scipy.special import erfc, erfcinv, lambertw
 
 from ._typing import FloatArray, SoilModelNames
 from .plot import swrc as plot_swrc
@@ -609,6 +609,60 @@ class Fredlund:
 
 
 @dataclass
+class Kosugi:
+    """Kosugi 2-Parameter Lognormal Soil Model.
+
+    Parameters
+    ----------
+    k_s: float
+        Saturated hydraulic conductivity [L/T]
+    theta_r: float
+        Residual soil water content [-]
+    theta_s: float
+        Saturated soil water content [-]
+    h_m: float
+        Median capillary pressure head [L]
+    sigma: float
+        Standard deviation of log-transformed soil pore radii [-]
+
+    References
+    ----------
+    Kosugi, K. (1996). Lognormal distribution model for unsaturated soil
+    hydraulic properties.
+
+    """
+
+    k_s: float
+    theta_r: float
+    theta_s: float
+    h_m: float
+    sigma: float
+
+    def theta(self, h: FloatArray) -> FloatArray:
+        se = 0.5 * erfc(log(h / self.h_m) / (sqrt(2.0) * self.sigma))
+        return self.theta_r + se * (self.theta_s - self.theta_r)
+
+    def s(self, h: FloatArray) -> FloatArray:
+        return (self.theta(h) - self.theta_r) / (self.theta_s - self.theta_r)
+
+    def k_r(self, h: FloatArray, s: FloatArray | None = None) -> FloatArray:
+        if s is None:
+            s = self.s(h)
+
+        return s**0.5 * (0.5 * erfc(erfcinv(2.0 * s) + self.sigma / sqrt(2.0))) ** 2
+
+    def k(self, h: FloatArray, s: FloatArray | None = None) -> FloatArray:
+        return self.k_s * self.k_r(h=h, s=s)
+
+    def h(self, theta: FloatArray) -> FloatArray:
+        se = (theta - self.theta_r) / (self.theta_s - self.theta_r)
+        return self.h_m * exp(sqrt(2.0) * self.sigma * erfcinv(2.0 * se))
+
+    def plot(self, ax: plt.Axes | None = None) -> plt.Axes:
+        return plot_swrc(self, ax=ax)
+
+
+@dataclass
 class GenuchtenGardner:
     """Combination soil model.
 
@@ -790,6 +844,7 @@ def get_soilmodel(
         "Rucker": Rucker,
         "Panday": Panday,
         "Fredlund": Fredlund,
+        "Kosugi": Kosugi,
         "GenuchtenGardner": GenuchtenGardner,
         "GenuchtenKool": GenuchtenKool,
     }
