+++ src/pedon/__init__.py
@@ -15,6 +15,7 @@ from .soilmodel import Genuchten as Genuchten
 from .soilmodel import GenuchtenGardner as GenuchtenGardner
 from .soilmodel import GenuchtenKool as GenuchtenKool
 from .soilmodel import Haverkamp as Haverkamp
+from .soilmodel import Kosugi as Kosugi
 from .soilmodel import Panday as Panday
 from .soilmodel import Rucker as Rucker
 from .soilmodel import SoilModel as SoilModel
