### 2026-05-26 b33fb0f
add test for kosugi
