### 2026-05-26 cca37ee
add campbell soilmodel
