+++ src/pedon/__init__.py
@@ -9,6 +9,7 @@ from ._version import show_versions as show_versions
 from .soil import Soil as Soil
 from .soil import SoilSample as SoilSample
 from .soilmodel import Brooks as Brooks
+from .soilmodel import Campbell as Campbell
 from .soilmodel import Fredlund as Fredlund
 from .soilmodel import Gardner as Gardner
 from .soilmodel import Genuchten as Genuchten
