### 2026-05-26 44e4e24
add kosugi and campbell to types
