+++ src/pedon/_typing.py
@@ -6,12 +6,14 @@ from numpy.typing import NDArray
 FloatArray = float | NDArray[float64]
 SoilModelNames = Literal[
     "Genuchten",
-    "GenuchtenGardner",
     "Brooks",
     "Panday",
     "Gardner",
     "Rucker",
     "Haverkamp",
     "Fredlund",
+    "Kosugi",
+    "Campbell",
     "GenuchtenKool",
+    "GenuchtenGardner",
 ]
