### 2026-05-26 fb763ab
add campbell tests
