### 2026-05-26 3b82808
improve type hints for fixtures
