+++ tests/test_soilmodel.py
@@ -27,7 +27,7 @@ def test_get_soilmodel(
     assert smt.__name__ == soilmodel_name
 
 
-def test_theta_genuchten(genuchten: pe.SoilModel) -> None:
+def test_theta_genuchten(genuchten: pe.Genuchten) -> None:
     """Test water content calculation for the van Genuchten model."""
     expected = array(
         [
@@ -45,7 +45,7 @@ def test_theta_genuchten(genuchten: pe.SoilModel) -> None:
     assert_close(genuchten.theta(h=h), expected)
 
 
-def test_s_genuchten(genuchten: pe.SoilModel) -> None:
+def test_s_genuchten(genuchten: pe.Genuchten) -> None:
     """Test degree of saturation calculation for the van Genuchten model."""
     expected = array(
         [
@@ -63,7 +63,7 @@ def test_s_genuchten(genuchten: pe.SoilModel) -> None:
     assert_close(genuchten.s(h=h), expected)
 
 
-def test_k_genuchten(genuchten: pe.SoilModel) -> None:
+def test_k_genuchten(genuchten: pe.Genuchten) -> None:
     """Test hydraulic conductivity calculation for the van Genuchten model."""
     expected = array(
         [
@@ -81,7 +81,7 @@ def test_k_genuchten(genuchten: pe.SoilModel) -> None:
     assert_close(genuchten.k(h=h), expected)
 
 
-def test_h_genuchten(genuchten: pe.SoilModel, theta: FloatArray = theta) -> None:
+def test_h_genuchten(genuchten: pe.Genuchten) -> None:
     """Test pressure head calculation for the van Genuchten model."""
     expected = array(
         [244927639.31052694, 139271.6694845786, 1998.5290326524432, 61.672352811394724]
@@ -89,7 +89,7 @@ def test_h_genuchten(genuchten: pe.SoilModel, theta: FloatArray = theta) -> None
     assert_close(genuchten.h(theta=theta), expected)
 
 
-def test_theta_brooks(brooks: pe.SoilModel) -> None:
+def test_theta_brooks(brooks: pe.Brooks) -> None:
     """Test water content calculation for the Brooks-Corey model."""
     expected = array(
         [
@@ -107,13 +107,13 @@ def test_theta_brooks(brooks: pe.SoilModel) -> None:
     assert_close(brooks.theta(h=h), expected)
 
 
-def test_s_brooks(brooks: pe.SoilModel) -> None:
+def test_s_brooks(brooks: pe.Brooks) -> None:
     """Test degree of saturation calculation for the Brooks-Corey model."""
     expected = array([1.0, 1.0, 1.0, 1.0, 0.01, 0.0001, 1e-06, 1e-08, 1e-10])
     assert_close(brooks.s(h=h), expected)
 
 
-def test_k_brooks(brooks: pe.SoilModel) -> None:
+def test_k_brooks(brooks: pe.Brooks) -> None:
     """Test hydraulic conductivity calculation for the Brooks-Corey model."""
     expected = array(
         [
@@ -131,13 +131,13 @@ def test_k_brooks(brooks: pe.SoilModel) -> None:
     assert_close(brooks.k(h=h), expected)
 
 
-def test_h_brooks(brooks: pe.SoilModel, theta: FloatArray = theta) -> None:
+def test_h_brooks(brooks: pe.Brooks) -> None:
     """Test pressure head calculation for the Brooks-Corey model."""
     expected = array([21.60246899, 14.86783883, 12.03443336, 10.37749043])
     assert_close(brooks.h(theta=theta), expected)
 
 
-def test_theta_panday(panday: pe.SoilModel) -> None:
[…]
