### 2026-05-26 972ea84
fixes for copilot
