+++ tests/conftest.py
@@ -52,7 +52,7 @@ def fredlund() -> pe.SoilModel:
 
 
 @fixture
-def kosugi() -> pe.SoilModel:
+def kosugi() -> pe.Kosugi:
     """Fixture for a Kosugi soil model with specific parameters for testing."""
     return pe.Kosugi(k_s=10.0, theta_r=0.01, theta_s=0.43, h_m=100.0, sigma=1.2)
 
