+++ docs/examples/index.md
@@ -30,4 +30,5 @@ This collection of Jupyter Notebooks demonstrates how to use `pedon` for analyzi
 05_curve_fitting
 06_hypags_curvefitting
 07_hysteresis
+08_modflow
 ```
