### 2026-05-27 0d9d895
add new soilmodels
