+++ paper/paper.bib
@@ -337,6 +337,39 @@
   year    = {1977}
 }
 
+@article{campbell_model_1974,
+  author  = {Campbell, G. S.},
+  title   = {A simple method for determining unsaturated conductivity from moisture retention data.},
+  journal = {Soil Science},
+  volume  = {117},
+  issue   = {6},
+  pages   = {311--314},
+  doi     = {10.1097/00010694-197406000-00001},
+  year    = {1974}
+}
+
+@article{kosugi_model_1996,
+  author  = {Kosugi, K.},
+  title   = {Lognormal Distribution Model for Unsaturated Soil Hydraulic Properties},
+  journal = {Water Resources Research},
+  volume  = {32},
+  number  = {9},
+  pages   = {2697--2703},
+  doi     = {10.1029/96WR01776},
+  year    = {1996}
+}
+
+@article{kool_parker_1987,
+  author  = {Kool, J. B. and Parker, J. C.},
+  title   = {Development and evaluation of closed-form expressions for hysteretic soil hydraulic properties},
+  journal = {Water Resources Research},
+  volume  = {23},
+  number  = {1},
+  pages   = {105--114},
+  doi     = {10.1029/WR023i001p00105},
+  year    = {1987}
+}
+
 @article{collenteur_signatures_2025,
   author  = {Collenteur, R. A. and Vonk, M. A. and Haaf, E.},
   title   = {Quantification and Analysis of Hydrograph Behavior Using Groundwater Signatures},
