+++ paper/paper.bib
@@ -422,3 +422,36 @@
   year    = {2021},
   doi     = {10.1016/j.cageo.2021.104897}
 }
+
+@incollection{langevin_modflow6_2017,
+  author    = {Langevin, Christian and Hughes, Joseph and Banta, Edward and Niswonger, Richard and Panday, Sorab and Provost, Alden},
+  title     = {Documentation for the {MODFLOW 6} Groundwater Flow Model},
+  booktitle = {Techniques and Methods},
+  publisher = {U.S. Geological Survey},
+  year      = {2017},
+  chapter   = {6-A55},
+  address   = {Reston, VA},
+  doi       = {10.3133/tm6A55}
+}
+
+@article{hughes_flopy_2024,
+  author  = {Hughes, J. D. and Langevin, C. D. and Paulinski, S. R. and Larsen, J. D. and Brakenhoff, D.},
+  title   = {FloPy Workflows for Creating Structured and Unstructured MODFLOW Models},
+  journal = {Groundwater},
+  volume  = {62},
+  number  = {1},
+  pages   = {124-139},
+  doi     = {10.1111/gwat.13327},
+  year    = {2024}
+}
+
+@article{bakker_flopy_2016,
+  author  = {Bakker, M. and Post, V. and Langevin, C. D. and Hughes, J. D. and White, J. T. and Starn, J. J. and Fienen, M. N.},
+  title   = {Scripting MODFLOW Model Development Using Python and FloPy},
+  journal = {Groundwater},
+  volume  = {54},
+  number  = {5},
+  pages   = {733-739},
+  doi     = {10.1111/gwat.12413},
+  year    = {2016}
+}
