### 2026-06-01 5535671
replace mypy badge with ty
