### 2026-06-01 0a37d12
fix all dois and urls in the code and README
