+++ README.md
@@ -13,26 +13,29 @@
 [![Formatter and Linter: ruff](https://img.shields.io/badge/linter-ruff-darkgreen?style=flat-square)](https://github.com/charliermarsh/ruff)
 
 Pedon (*from Greek: πέδον, pedon -> soil*) is a Python package for working with unsaturated soil hydraulic properties. It provides an object-oriented framework for soils, soil samples, and soil hydraulic models, making it easy to describe, analyze, and parameterize soil water retention and hydraulic conductivity behavior. At its core, Pedon treats each soil model as a Python class that defines a soil water retention curve and an unsaturated hydraulic conductivity function. The package currently includes implementations of several widely used models:
-  - [Mualem-Van Genuchten](https://www.soilphysics.okstate.edu/teaching/soil-6583/references-folder/van%20Genuchten%201980.pdf)
-  - [Brooks-Corey](https://www.wipp.energy.gov/library/cra/2009_cra/references/Others/Brooks_Corey_1964_Hydraulic_Properties_ERMS241117.pdf)
-  - [Fredlund & Xing](https://projects.mans.edu.eg/heepf/ilppp/cources/12/pdf%20course/1/pressure/osmotic%20soilsalinity22.pdf)
+  - [Mualem-Van Genuchten](https://doi.org/10.2136/sssaj1980.03615995004400050002x)
+  - [Brooks-Corey](https://mountainscholar.org/items/3c7b98df-13e3-486c-9d1e-949a7a869f76)
+  - [Fredlund & Xing](https://doi.org/10.1139/t94-061)
   - [Haverkamp](https://doi.org/10.2136/sssaj1977.03615995004100020024x)
   - [Gardner](https://doi.org/10.1097/00010694-195804000-00006)
-  - Panday ([MODFLOW USG-Transport](https://www.gsienv.com/product/modflow-usg/))
+  - [Rucker](https://doi.org/10.1016/j.advwatres.2005.01.004)
+  - [Kosugi](https://doi.org/10.1029/96WR01776)
+  - [Campbell](https://doi.org/10.1097/00010694-197406000-00001)
+  - Panday ([MODFLOW USG-Transport](https://www.gsienv.com/software/modflow-usg/modflow-usg/))
 
 Pedon allows these models to be evaluated directly and fitted to data. When laboratory measurements of soil water retention and/or unsaturated hydraulic conductivity are available, model parameters can be estimated by fitting analytical curves to the data using nonlinear least squares, following the same methodology as the widely used [RETC](https://www.pc-progress.com/Documents/programs/retc.pdf).
 
 To support parameterization without direct measurements, Pedon includes access to several established soil hydraulic parameter datasets:
-  - [HYDRUS](https://www2.pc-progress.com/downloads/Pgm_Hydrus3D5/HYDRUS_user_Manual_V5.pdf)
-  - [VS2D](https://pubs.usgs.gov/wri/1983/4099/report.pdf)
-  - Staring Series ([2001](https://edepot.wur.nl/43272) & [2018](https://edepot.wur.nl/512761))
+  - [HYDRUS](https://www.pc-progress.com/downloads/pgm_hydrus1d/hydrus1d-4.08.pdf)
+  - [VS2D](https://doi.org/10.3133/wri904025)
+  - Staring Series: [2001](https://edepot.wur.nl/43272) and [2018](https://doi.org/10.18174/512761)
 
 In addition, Pedon implements multiple pedotransfer functions that estimate soil hydraulic parameters from easily measured soil properties:
[…]
