+++ paper/paper.bib
@@ -1,11 +1,23 @@
-@article{peche_hypags_2024,
+@article{peche_hypags_2023,
+  author  = {Peche, A. and Houben, G. J.},
+  title   = {Estimating Characteristic Grain Sizes and Effective Porosity from Hydraulic Conductivity Data},
+  journal = {Groundwater},
+  volume  = {61},
+  number  = {4},
+  pages   = {574--585},
+  doi     = {doi.org/10.1111/gwat.13266},
+  year    = {2023}
+}
+
+
+@article{peche_genuchten_2024,
   author  = {Peche, A. and Houben, G. and Altfelder, S.},
   title   = {Approximation of van Genuchten Parameter Ranges from Hydraulic Conductivity Data},
   journal = {Groundwater},
   volume  = {62},
   number  = {3},
   pages   = {469--479},
-  doi     = {10.1111/gwat.13365},
+  doi     = {doi.org/10.1111/gwat.13365},
   year    = {2024}
 }
 
@@ -414,7 +426,7 @@
 }
 
 @article{memari_pyswr_2021,
-  author  = {S. S. Memari and T. P. Clement},
+  author  = {Memari, S. S. and Clement, T. P.},
   title   = {PySWR - A Python code for fitting soil water retention functions},
   journal = {Computers & Geosciences},
   volume  = {156},
@@ -424,7 +436,7 @@
 }
 
 @incollection{langevin_modflow6_2017,
-  author    = {Langevin, Christian and Hughes, Joseph and Banta, Edward and Niswonger, Richard and Panday, Sorab and Provost, Alden},
+  author    = {Langevin, C. D. and Hughes, J. D. and Banta, E. and Niswonger, R. and Panday, S. and Provost, A.},
   title     = {Documentation for the {MODFLOW 6} Groundwater Flow Model},
   booktitle = {Techniques and Methods},
   publisher = {U.S. Geological Survey},
