+++ paper/paper.bib
@@ -5,7 +5,7 @@
   volume  = {61},
   number  = {4},
   pages   = {574--585},
-  doi     = {doi.org/10.1111/gwat.13266},
+  doi     = {10.1111/gwat.13266},
   year    = {2023}
 }
 
@@ -17,7 +17,7 @@
   volume  = {62},
   number  = {3},
   pages   = {469--479},
-  doi     = {doi.org/10.1111/gwat.13365},
+  doi     = {10.1111/gwat.13365},
   year    = {2024}
 }
 
@@ -466,4 +466,14 @@
   pages   = {733-739},
   doi     = {10.1111/gwat.12413},
   year    = {2016}
+}
+
+@article{brink_swap_2026,
+  author  = {{van den Brink}, J. M. and {van Dam}, J.C. and Mulder, H. M. and Ritsema, C. J. and Bartholomeus, R. P.},
+  title   = {Effect of increasing conceptual model detail on simulated crop yield-drainage base relations},
+  journal = {Agricultural Water Management},
+  volume  = {329},
+  pages   = {110383},
+  year    = {2026},
+  doi     = {10.1016/j.agwat.2026.110383}
 }
