+++ paper/paper.md
@@ -122,7 +122,7 @@ bc = pe.SoilSample(h=h, theta=theta, k=k).fit(pe.Brooks)
 `pedon` contributes to the field of groundwater modeling by providing a modular, object-oriented framework that integrates hydraulic soil models, established parameter databases, and pedotransfer functions into a single, reproducible workflow. Existing tools such as unsatfit [@seki_unsatfit_2023], PySWR [@memari_pyswr_2021], and RETC [@genuchten_retc_1991] are utilities primarily focused on least-squares fitting of predefined soil models. In contrast, `pedon` enables researchers to define custom soil model formulations and systematically evaluate them against different datasets and pedotransfer functions. This level of interoperability is unique and provides a broader framework for the analysis of unsaturated soil hydraulic properties. At the same time, pedon is developed with a collaborative mindset, exemplified by the integration and extension of the soil model parameter estimation algorithms from HYPAGS [@peche_hypags_2024].
 
 # Research impact statement
-Soil hydraulic models and their parameters are essential for simulating variably saturated groundwater flow. Determining these parameters experimentally is difficult, time-consuming, and uncertain [@genuchten_retc_1991; @brandhorst_uncertainty_2017]. Therefore, soil hydraulic model parameters are often approximated or estimated from reference databases. `pedon` bundles soil models and parameter sources in a single framework, enabling efficient parameter derivation without extensive literature searches or ad hoc reimplementation. `pedon` is already used in scientific workflows for variably saturated groundwater flow modeling, including published studies by @vonk_nonlinear_2024 and @collenteur_signatures_2025. Here it facilitates easy coupling to MODFLOW 6 [@langevin_modflow6_2017] and MODFLOW-USG Transport [@panday_mfusgt_2026] via Python [@bakker_flopy_2016; @hughes_flopy_2024]. `pedon` is also a dependency of the Python package [`dutchsoils`](https://github.com/markvdbrink/dutchsoils), which is used in an academic context to process Dutch soil datasets [@heinen_bofek_2022].
+Soil hydraulic models and their parameters are essential for simulating variably saturated groundwater flow. Determining these parameters experimentally is difficult, time-consuming, and uncertain [@genuchten_retc_1991; @brandhorst_uncertainty_2017]. Therefore, soil hydraulic model parameters are often approximated or estimated from reference databases. `pedon` bundles soil models and parameter sources in a single framework, enabling efficient parameter derivation without extensive literature searches or ad hoc reimplementation. `pedon` is already used in scientific workflows for variably saturated groundwater flow modeling, including published studies by @vonk_nonlinear_2024 and @collenteur_signatures_2025. Here it facilitates easy coupling to MODFLOW 6 [@langevin_modflow6_2017] and MODFLOW-USG Transport [@panday_mfusgt_2026] via Python [@bakker_flopy_2016; @hughes_flopy_2024]. Additional `pedon` was used in @brink_swap_2026 for processing Dutch soil datasets [@heinen_bofek_2022].
 
 # AI usage disclosure
 GitHub Copilot was used during development for reviewing pull requests, writing unit tests and documentation, providing code completion, and sanity-checking proposed bug fixes. ChatGPT was used for this manuscript to review references, identify linguistic and grammatical errors, and verify compliance with the Journal of Open Source Software requirements. All AI-generated outputs were reviewed by the authors, who take full responsibility for the accuracy and originality of the works.
