### 2026-06-01 c2dc08e
fix some typing mypy errors
