### 2026-06-01 a43ad35
update refs to mypy and ruff
