### 2026-06-01 7053599
properly deprecate plot_hcf
