### 2026-06-01 e3c5b26
create resolve_soilmodel method
