+++ src/pedon/_params.py
@@ -31,20 +31,9 @@ def get_params(
     sm: Type[SoilModel] | SoilModel | SoilModelNames,
 ) -> DataFrame:
     """Get the parameter bounds for a specific soil model."""
-    if isinstance(sm, type) and issubclass(sm, SoilModel):
-        smn = sm.__name__
-    elif isinstance(sm, SoilModel):
-        smn = getattr(sm, "__name__", sm.__class__.__name__)
-        sm = type(sm)
-    elif isinstance(sm, str):
-        smn = sm
-        sm = get_soilmodel(smn)
-    else:
-        raise TypeError(
-            f"Argument must either be Type[SoilModel] | SoilModel | str, not {type(sm)}"
-        )
+    smn, sm_cls = resolve_soilmodel(sm)
 
-    params = _get_default_params(sm)
+    params = _get_default_params(sm_cls)
 
     param_bounds = {
         "Genuchten": {
