### 2026-06-01 db696ae
create sourcenames literal
