### 2026-06-01 6f9e6a3
create coverage badge
