+++ pyproject.toml
@@ -76,7 +76,7 @@ ignore_missing_imports = true
 
 [tool.pytest.ini_options]
 pythonpath = ["src"]
-addopts = "--cov=pedon --cov-report=term-missing --cov-branch"
+addopts = "--cov=pedon --cov-report=term-missing --cov-report=json --cov-branch"
 
 
 [tool.ruff]
