+++ .github/workflows/tests.yml
@@ -50,6 +50,7 @@ jobs:
 
     name: ${{ matrix.name }}
     env:
+      BADGE_GIST_COVERAGE_TOKEN: ${{ secrets.BADGE_GIST_COVERAGE_TOKEN }}
       # Color Output
       # Rich (pip)
       FORCE_COLOR: 1
@@ -96,12 +97,11 @@ jobs:
           PY
 
       - name: Update Coverage Badge
-        if: matrix.toxenv == 'py313' && github.event_name == 'push' && github.ref == 'refs/heads/dev'
+        if: matrix.toxenv == 'py313' && github.event_name == 'push' && github.ref == 'refs/heads/dev' && env.BADGE_GIST_COVERAGE_TOKEN != ''
         uses: Schneegans/dynamic-badges-action@v1.7.0
         with:
-          auth: ${{ secrets.BADGE_GIST_COVERAGE_TOKEN }}
+          auth: ${{ env.BADGE_GIST_COVERAGE_TOKEN }}
           gistID: "a891dd883f258c46826ba657cf9a89cf"
           filename: coverage.json
           label: coverage
           message: ${{ steps.coverage.outputs.coverage }}%
-          color: 34D058
