### 2026-06-01 c63f39c
add joss badge
