### 2026-06-01 9a38bf7
some textual clarifications
