+++ paper/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 Researchers and engineers working with unsaturated soils need estimates of soil hydraulic model parameters for variably saturated groundwater flow models. `pedon` provides a Python toolkit that brings together soil hydraulic models, parameter databases, pedotransfer functions, and fitting routines, making soil analysis faster, more reproducible, and easier to integrate into existing groundwater modeling workflows.
 
 # Soil hydraulic models
-A soil hydraulic model (soil model in short) is a parametric description of soil hydraulic functions: the soil water retention curve (SWRC) and the unsaturated hydraulic conductivity function (HCF). These relate soil water content and flow to pressure head and vice versa for use in variably saturated groundwater flow models. At this time, `pedon` provides the following soil models:
+A soil hydraulic model (or soil model for short) is a parametric description of soil hydraulic functions: the soil water retention curve (SWRC) and the unsaturated hydraulic conductivity function (HCF). These relate soil water content and flow to pressure head and vice versa for use in variably saturated groundwater flow models. At this time, `pedon` provides the following soil models:
 
 - `pedon.Genuchten`: Mualem-van Genuchten [@genuchten_mualem_1980]
 - `pedon.Brooks`: Brooks-Corey [@brooks_corey_1964]
@@ -69,13 +69,13 @@ theta = mg.theta(h)  # water content (-) at pressure head values
 k = mg.k(h)  # hydraulic conductivity (cm/d) at pressure head values
 ```
 
-The object-oriented design and duck typing provides a clear and consistent structure in which users can define custom soil model classes. Additionally, `pedon` only depends on well-maintained packages in the Python scientific ecosystem such as NumPy [@numpy_article_2020], SciPy [@scipy_paper_2020], Matplotlib [@matplotlib_paper_2007], and Pandas [@pandas_software_2020; @pandas_paper_2010].
+The object-oriented design and duck typing (relying on object methods rather than explicit types) provide a clear and consistent structure in which users can define custom soil model classes. Additionally, `pedon` only depends on well-maintained packages in the Python scientific ecosystem such as NumPy [@numpy_article_2020], SciPy [@scipy_paper_2020], Matplotlib [@matplotlib_paper_2007], and Pandas [@pandas_software_2020; @pandas_paper_2010].
 
 # Soil hydraulic parameters
 Soil hydraulic parameters depend on soil type and determine the shape of a soil model’s SWRC and HCF. They are rarely measured directly and are usually derived from reference datasets, empirical relationships, or laboratory measurements. `pedon` links these parameters to soil models and provides a framework to obtain them from existing datasets, easily measured soil properties, and direct measurements of soil water content and hydraulic conductivity.
 
 ## Parameter datasets
-`pedon` includes a dataset of Brooks–Corey and Mualem–van Genuchten parameters for a wide range of soils. At present, this dataset is compiled from three established soil hydraulic parameter databases:
+`pedon` includes a dataset of Brooks–Corey and Mualem–van Genuchten parameters for a wide range of soils. Currently, this dataset is compiled from three established soil hydraulic parameter databases:
 
 - Average parameter values for twelve major soil textural groups defined by @carsel_dataset_1988, also used in the HYDRUS software for variably saturated flow modeling [@simunek_hydrus1d_2009];
 - A dataset from the VS2D software [@healy_vs2d_1990] containing both Brooks–Corey and Mualem–van Genuchten parameters;
@@ -106,7 +106,7 @@ hypags: pe.Genuchten = pe.SoilSample(k=ks).hypags()
 ```
 
 ### Soil hydraulic measurements
[…]
