### 2026-06-01 fe4bf03
fix deprecationwarning
