+++ paper/paper.bib
@@ -148,7 +148,6 @@
   year    = {1958}
 }
 
-
 @article{bakker_gardner_2009,
   author  = {Bakker, M. and Nieber, J. L.},
   title   = {Damping of Sinusoidal Surface Flux Fluctuations with Soil Depth},
@@ -207,6 +206,17 @@
   year      = {1989}
 }
 
+@article{looy_pedotransfer_2017,
+  author  = {Van Looy, K. and Bouma, J. and Herbst, M. and Koestel, J. and Minasny, B. and Mishra, U. and Montzka, C. and Nemes, A. and Pachepsky, Y. A. and Padarian, J. and Schaap, M. G. and Tóth, B. and Verhoef, A. and Vanderborght, J. and {van der Ploeg}, M. J. and Weihermüller, L. and Zacharias, S. and Zhang, Y. and Vereecken, H.},
+  title   = {Pedotransfer Functions in Earth System Science: Challenges and Perspectives},
+  journal = {Reviews of Geophysics},
+  volume  = {55},
+  number  = {4},
+  pages   = {1199-1256},
+  doi     = {10.1002/2017RG000581},
+  year    = {2017}
+}
+
 @article{wosten_pedotransfer_1999,
   author  = {W{\"o}sten, J. H. M. and Nemes, A. and Lilly, A. and Le Bas, C.},
   title   = {Development and Use of a Database of Hydraulic Properties of European Soils},
@@ -257,6 +267,16 @@
   pages   = {163--176},
   doi     = {10.1016/S0022-1694(01)00466-8},
   year    = {2001}
+}
+
+@article{zhang_rosetta3_2017,
+  author  = {Zhang, Y. and Schaap, M. G.},
+  title   = {Weighted recalibration of the {Rosetta} pedotransfer model with improved estimates of hydraulic parameter distributions and summary statistics {(Rosetta3)}},
+  journal = {Journal of Hydrology},
+  volume  = {547},
+  pages   = {39--53},
+  year    = {2017},
+  doi     = {10.1016/j.jhydrol.2017.01.004}
 }
 
 @article{scipy_paper_2020,
@@ -412,6 +432,17 @@
   pages   = {360--370},
   doi     = {10.1016/j.advwatres.2017.10.022},
   year    = {2017}
+}
+
+@article{baroni_uncertainty_2010,
+  author  = {Baroni, G. and Facchi, A. and Gandolfi, C. and Ortuani, B. and Horeschi, D. and van Dam, J. C.},
+  title   = {Uncertainty in the determination of soil hydraulic parameters and its influence on the performance of two hydrological models of different complexity},
+  journal = {Hydrology and Earth System Sciences},
+  volume  = {14},
+  year    = {2010},
+  number  = {2},
+  pages   = {251--270},
+  doi     = {10.5194/hess-14-251-2010}
 }
 
 @article{seki_unsatfit_2023,
