+++ paper/paper.md
@@ -92,7 +92,7 @@ staring = pe.Soil("B01").from_name(pe.Genuchten, source="Staring_2018")
 `pedon` provides two approaches for obtaining soil model parameters from soil data. The first uses pedotransfer functions based on easily measured soil properties. The second relies on direct measurements of soil water content and hydraulic conductivity.
 
 ### Pedotransfer functions
-Pedotransfer functions relate easily measured soil properties (e.g. sand, silt, clay or organic matter content and bulk density) to soil model parameters [@bouma_pedotransfer_1989]. `pedon` implements functions from the literature, including those of @wosten_pedotransfer_1999, @wosten_staringreeks_2001, @cosby_pedotransfer_1984, and @cooper_pedotransfer_2021. It also provides access to parameter databases such as Rosetta [@schaap_rosetta_2001] (optional dependency) and HYPAGS [@peche_hypags_2023, @peche_genuchten_2024], the latter enabling estimation from a single value of saturated hydraulic conductivity or representative grain diameters.
+Pedotransfer functions relate easily measured soil properties (e.g. sand, silt, clay or organic matter content and bulk density) to soil model parameters [@bouma_pedotransfer_1989; @looy_pedotransfer_2017]. `pedon` implements functions from the literature, including those of @wosten_pedotransfer_1999; @wosten_staringreeks_2001; @cosby_pedotransfer_1984, and @cooper_pedotransfer_2021. It also provides access to parameter databases such as Rosetta [@schaap_rosetta_2001; @zhang_rosetta3_2017] (optional dependency) and HYPAGS [@peche_hypags_2023; @peche_genuchten_2024], the latter enabling estimation from a single value of saturated hydraulic conductivity or representative grain diameters.
 
 ```python
 # Estimate parameters using Cosby's pedotransfer function
@@ -100,7 +100,7 @@ sand_p = 40.0  # sand (%)
 clay_p = 10.0  # clay (%)
 cosby: pe.Brooks = pe.SoilSample(sand_p=sand_p, clay_p=clay_p).cosby()
 
-# Estimate parameters from saturated conductivity via HYPAGS
+# Estimate parameters using HYPAGS
 ks = 1e-4  # saturated hydraulic conductivity (m/s)
 hypags: pe.Genuchten = pe.SoilSample(k=ks).hypags()
 ```
@@ -116,13 +116,13 @@ The same fitting procedure can translate between soil models. The SWRC and HCF g
 bc = pe.SoilSample(h=h, theta=theta, k=k).fit(pe.Brooks)
 ```
 
-![Resulting Brooks-Corey SWRC after fitting on the Mualem-van Genuchten soil model \label{fig:swrc_fit}](figures/swrc_fit.png){height=5.4cm}
+![Resulting Brooks-Corey SWRC and HCF after fitting on the Mualem-van Genuchten soil model \label{fig:swrc_fit}](figures/fit_swrc_hcf.png){height=5.0cm}
 
 # State of the field
-`pedon` contributes to the field of groundwater modeling by providing a modular, object-oriented framework that integrates hydraulic soil models, established parameter databases, and pedotransfer functions into a single, reproducible workflow. Existing tools such as unsatfit [@seki_unsatfit_2023], PySWR [@memari_pyswr_2021], and RETC [@genuchten_retc_1991] are utilities primarily focused on least-squares fitting of predefined soil models. In contrast, `pedon` enables researchers to define custom soil model formulations and systematically evaluate them against different datasets and pedotransfer functions. This level of interoperability is unique and provides a broader framework for the analysis of unsaturated soil hydraulic properties. At the same time, pedon is developed with a collaborative mindset, exemplified by the integration and extension of the soil model parameter estimation algorithms from HYPAGS [@peche_hypags_2024].
[…]
