+++ paper/paper.bib
@@ -508,3 +508,36 @@
   year    = {2026},
   doi     = {10.1016/j.agwat.2026.110383}
 }
+
+@article{clapp_hornberger_1978,
+  title   = {Empirical equations for some soil hydraulic properties},
+  author  = {Clapp, R. B. and Hornberger, G. M.},
+  journal = {Water Resources Research},
+  volume  = {14},
+  number  = {4},
+  pages   = {601--604},
+  year    = {1978},
+  doi     = {10.1029/WR014i004p00601}
+}
+
+@article{saxton_generalized_1986,
+  author  = {Saxton, K. E. and Rawls, W. J. and Romberger, J. S. and Papendick, R. I.},
+  title   = {Estimating generalized soil-water characteristics from texture},
+  journal = {Soil Science Society of America Journal},
+  volume  = {50},
+  number  = {4},
+  pages   = {1031--1036},
+  year    = {1986},
+  doi     = {10.2136/sssaj1986.03615995005000040039x}
+}
+
+@article{saxton_rawls_2006,
+  author  = {Saxton, K. E. and Rawls, W. J.},
+  title   = {Soil Water Characteristic Estimates by Texture and Organic Matter for Hydrologic Solutions},
+  journal = {Soil Science Society of America Journal},
+  volume  = {70},
+  number  = {5},
+  pages   = {1569--1578},
+  year    = {2006},
+  doi     = {10.2136/sssaj2005.0117}
+}
