### 2026-06-02 11ee371
add weynants ptf
