+++ src/pedon/soil.py
@@ -549,11 +549,6 @@ class SoilSample:
             Density adjustment factor (normally between 0.9 and 1.3) to account
             for compaction or loosening. Default is 1.0 (normal density).
 
-        Returns
-        -------
-        Brooks
-            Brooks-Corey soil model with estimated parameters.
-
         References
         ----------
         Saxton, K. E., Rawls, W. J., Romberger, J. S., & Papendick, R. I. (1986).
@@ -571,48 +566,47 @@ class SoilSample:
         assert self.om_p is not None, msg
 
         # Sand and Clay are calculated as fractions, OM is kept as a percentage
-        s = self.sand_p / 100.0
-        c = self.clay_p / 100.0
-        om = self.om_p
+        s_f = self.sand_p / 100.0
+        c_f = self.clay_p / 100.0
 
         # 1500 kPa moisture (Wilting Point)
         th1500t = (
-            -0.024 * s
-            + 0.487 * c
-            + 0.006 * om
-            + 0.005 * (s * om)
-            - 0.013 * (c * om)
-            + 0.068 * (s * c)
+            -0.024 * s_f
+            + 0.487 * c_f
+            + 0.006 * self.om_p
+            + 0.005 * (s_f * self.om_p)
+            - 0.013 * (c_f * self.om_p)
+            + 0.068 * (s_f * c_f)
             + 0.031
         )
         th1500 = th1500t + (0.14 * th1500t - 0.02)
 
         # 33 kPa moisture (Field capacity)
         th33t = (
-            -0.251 * s
-            + 0.195 * c
-            + 0.011 * om
-            + 0.006 * (s * om)
-            - 0.027 * (c * om)
-            + 0.452 * (s * c)
+            -0.251 * s_f
+            + 0.195 * c_f
+            + 0.011 * self.om_p
+            + 0.006 * (s_f * self.om_p)
+            - 0.027 * (c_f * self.om_p)
+            + 0.452 * (s_f * c_f)
             + 0.299
         )
         th33 = th33t + (1.283 * th33t**2 - 0.374 * th33t - 0.015)
 
         # Saturation - 33 kPa moisture
         ths_33t = (
-            0.278 * s
-            + 0.034 * c
-            + 0.022 * om
-            - 0.018 * (s * om)
-            - 0.027 * (c * om)
-            - 0.584 * (s * c)
+            0.278 * s_f
+            + 0.034 * c_f
+            + 0.022 * self.om_p
+            - 0.018 * (s_f * self.om_p)
+            - 0.027 * (c_f * self.om_p)
+            - 0.584 * (s_f * c_f)
             + 0.078
         )
         ths_33 = ths_33t + (0.636 * ths_33t - 0.107)
 
         # Saturation moisture & density
-        ths = th33 + ths_33 - 0.097 * s + 0.043
+        ths = th33 + ths_33 - 0.097 * s_f + 0.043
         rho_n = (1 - ths) * 2.65
 
         # Apply density factor adjustment
@@ -629,36 +623,31 @@ class SoilSample:
 
         # Air entry tension (bubbling pressure) in kPa
         psi_et = (
-            -21.67 * s
-            - 27.93 * c
+            -21.67 * s_f
+            - 27.93 * c_f
             - 81.97 * ths_33
-            + 71.12 * (s * ths_33)
-            + 8.29 * (c * ths_33)
-            + 14.05 * (s * c)
+            + 71.12 * (s_f * ths_33)
+            + 8.29 * (c_f * ths_33)
+            + 14.05 * (s_f * c_f)
             + 27.16
         )
[…]
