### 2026-06-02 84d3730
update rosetta refs
