### 2026-06-02 3f57b66
add test for clapp
