### 2026-06-02 afafd76
add tests for ptfs
