### 2026-06-02 d215702
add tests for topsoil=False
