+++ tests/test_ptf.py
@@ -29,6 +29,19 @@ def test_wosten(ss: pe.soil.SoilSample) -> None:
     assert sm.m == pytest.approx(0.10788745337697181, rel=REL)
 
 
+def test_wosten_no_topsoil(ss: pe.soil.SoilSample) -> None:
+    """Test Wösten pedotransfer function without topsoil adjustment."""
+    sm = ss.wosten(topsoil=False)
+    assert isinstance(sm, pe.Genuchten)
+    assert sm.k_s == pytest.approx(0.10889706346920575, rel=REL)
+    assert sm.theta_r == pytest.approx(0.01, rel=REL)
+    assert sm.theta_s == pytest.approx(0.3522969025688723, rel=REL)
+    assert sm.alpha == pytest.approx(0.001298720038387063, rel=REL)
+    assert sm.n == pytest.approx(1.0907448358614613, rel=REL)
+    assert sm.l == pytest.approx(-3.6143956260446446, rel=REL)
+    assert sm.m == pytest.approx(0.08319529268253811, rel=REL)
+
+
 def test_wosten_sand(ss: pe.soil.SoilSample) -> None:
     """Test Wösten sand-specific pedotransfer function."""
     sm = ss.wosten_sand(topsoil=True)
@@ -37,6 +50,19 @@ def test_wosten_sand(ss: pe.soil.SoilSample) -> None:
     assert sm.theta_r == pytest.approx(0.01, rel=REL)
     assert sm.theta_s == pytest.approx(0.4959, rel=REL)
     assert sm.alpha == pytest.approx(0.0204414, rel=REL)
+    assert sm.n == pytest.approx(2.2182, rel=REL)
+    assert sm.l == pytest.approx(2.0, rel=REL)
+    assert sm.m == pytest.approx(0.5491840230817779, rel=REL)
+
+
+def test_wosten_sand_no_topsoil(ss: pe.soil.SoilSample) -> None:
+    """Test Wösten sand-specific pedotransfer function without topsoil adjustment."""
+    sm = ss.wosten_sand(topsoil=False)
+    assert isinstance(sm, pe.Genuchten)
+    assert sm.k_s == pytest.approx(20.793, rel=REL)
+    assert sm.theta_r == pytest.approx(0.01, rel=REL)
+    assert sm.theta_s == pytest.approx(0.4959, rel=REL)
+    assert sm.alpha == pytest.approx(0.0242778, rel=REL)
     assert sm.n == pytest.approx(2.2182, rel=REL)
     assert sm.l == pytest.approx(2.0, rel=REL)
     assert sm.m == pytest.approx(0.5491840230817779, rel=REL)
@@ -138,6 +164,19 @@ def test_toth_topsoil(ss: pe.soil.SoilSample) -> None:
     assert sm.n == pytest.approx(1.2221, rel=REL)
     assert sm.l == pytest.approx(0.5, rel=REL)
     assert sm.m == pytest.approx(0.18173635545372713, rel=REL)
+
+
+def test_toth_no_topsoil(ss: pe.soil.SoilSample) -> None:
+    """Test Tóth pedotransfer function without topsoil adjustment."""
+    sm = ss.toth(topsoil=False)
+    assert isinstance(sm, pe.Genuchten)
+    assert sm.k_s == pytest.approx(0.5623, rel=REL)
+    assert sm.theta_r == pytest.approx(0.041, rel=REL)
+    assert sm.theta_s == pytest.approx(0.4203, rel=REL)
+    assert sm.alpha == pytest.approx(0.0043157, rel=REL)
+    assert sm.n == pytest.approx(1.2524, rel=REL)
[…]
