### 2026-06-02 d168082
add refs to readme
