+++ paper/paper.bib
@@ -9,7 +9,6 @@
   year    = {2023}
 }
 
-
 @article{peche_genuchten_2024,
   author  = {Peche, A. and Houben, G. and Altfelder, S.},
   title   = {Approximation of van Genuchten Parameter Ranges from Hydraulic Conductivity Data},
@@ -520,7 +519,7 @@
   doi     = {10.1029/WR014i004p00601}
 }
 
-@article{saxton_generalized_1986,
+@article{saxton_pedotransfer_1986,
   author  = {Saxton, K. E. and Rawls, W. J. and Romberger, J. S. and Papendick, R. I.},
   title   = {Estimating generalized soil-water characteristics from texture},
   journal = {Soil Science Society of America Journal},
@@ -531,7 +530,7 @@
   doi     = {10.2136/sssaj1986.03615995005000040039x}
 }
 
-@article{saxton_rawls_2006,
+@article{saxton_pedotransfer_2006,
   author  = {Saxton, K. E. and Rawls, W. J.},
   title   = {Soil Water Characteristic Estimates by Texture and Organic Matter for Hydrologic Solutions},
   journal = {Soil Science Society of America Journal},
@@ -540,4 +539,70 @@
   pages   = {1569--1578},
   year    = {2006},
   doi     = {10.2136/sssaj2005.0117}
+}
+
+@article{weynants_pedotransfer_2009,
+  author  = {Weynants, M. and Vereecken, H. and Javaux, M.},
+  title   = {Revisiting {V}ereecken Pedotransfer Functions: Introducing a Closed-Form Hydraulic Model},
+  journal = {Vadose Zone Journal},
+  volume  = {8},
+  number  = {1},
+  pages   = {86--95},
+  year    = {2009},
+  doi     = {10.2136/vzj2008.0062}
+}
+
+@article{toth_pedotransfer_2015,
+  author  = {T{\'{o}}th, B. and Weynants, M. and Nemes, A. and Mak{\'{o}}, A. and Bilas, G. and T{\'{o}}th, G.},
+  title   = {New generation of hydraulic pedotransfer functions for {Europe}},
+  journal = {European Journal of Soil Science},
+  volume  = {66},
+  number  = {1},
+  pages   = {226--238},
+  year    = {2015},
+  doi     = {10.1111/ejss.12192}
+}
+
+@article{hodnett_pedotransfer_2002,
+  author  = {Hodnett, M. G. and Tomasella, J.},
+  title   = {Marked differences between van {Genuchten} soil water-retention parameters for temperate and tropical soils: a new water-retention pedo-transfer functions developed for tropical soils},
+  journal = {Geoderma},
+  volume  = {108},
+  number  = {3-4},
+  pages   = {155--180},
+  year    = {2002},
+  doi     = {10.1016/S0016-7061(02)00105-2}
+}
+
+@article{ottoni_pedotransfer_2018,
+  author  = {Ottoni, M. V. and Ottoni Filho, T. B. and Schaap, M. G. and Lopes-Assad, M. L. R. C. and Rotunno Filho, O. C.},
+  title   = {Hydrophysical Database for {B}razilian Soils ({HYBRAS}) and Pedotransfer Functions for Water Retention},
+  journal = {Vadose Zone Journal},
+  volume  = {17},
+  number  = {1},
+  pages   = {1--17},
+  year    = {2018},
+  doi     = {10.2136/vzj2017.05.0095}
+}
+
+@article{vereecken_pedotransfer_1989,
+  author  = {Vereecken, H. and Maes, J. and Feyen, J. and Darius, P.},
+  title   = {Estimating the Soil Moisture Retention Characteristic from Texture, Bulk Density, and Carbon Content},
[…]
