### 2026-06-02 ad11fe6
smaller evaluation range
