+++ paper/paper.bib
@@ -16,18 +16,18 @@
   volume  = {62},
   number  = {3},
   pages   = {469--479},
-  doi     = {10.1111/gwat.13365},
+  doi     = {10.1111/gwat.133365},
   year    = {2024}
 }
 
 @manual{genuchten_retc_1991,
-  title        = {The {RETC} Code for Quantifying the Hydraulic Functions. {Version 1.0}},
   author       = {{van Genuchten}, M. Th. and Leij, F. J. and Yates, S. R.},
+  title        = {The {RETC} Code for Quantifying the Hydraulic Functions. {Version 1.0}},
   organization = {U.S. Salinity Laboratory, USDA, ARS},
   address      = {Riverside, CA},
   note         = {EPA Report 600/2-91/065},
-  year         = {1991},
-  url          = {https://cfpub.epa.gov/si/si_public_record_report.cfm?dirEntryId=130162}
+  url          = {https://cfpub.epa.gov/si/si_public_record_report.cfm?dirEntryId=130162},
+  year         = {1991}
 }
 
 @article{carsel_dataset_1988,
@@ -89,8 +89,8 @@
   title        = {The {HYDRUS-1D} software package for simulating the one-dimensional movement of water, heat, and multiple solutes in variably-saturated media. {Version 4.08}},
   organization = {University of California Riverside},
   address      = {Riverside, CA},
-  year         = {2009},
-  url          = {https://www.pc-progress.com/downloads/pgm_hydrus1d/hydrus1d-4.08.pdf}
+  url          = {https://www.pc-progress.com/downloads/pgm_hydrus1d/hydrus1d-4.08.pdf},
+  year         = {2009}
 }
 
 @article{fuentes_burdine_1992,
@@ -109,9 +109,9 @@
   title       = {Waterretentie- en Doorlatendheidskarakteristieken van Boven- en Ondergronden in Nederland: De Staringreeks; (Update 2018)},
   institution = {Wageningen Environmental Research},
   number      = {2978},
-  year        = {2020},
+  language    = {Dutch},
   doi         = {10.18174/512761},
-  language    = {Dutch}
+  year        = {2020}
 }
 
 @incollection{healy_vs2d_1990,
@@ -176,9 +176,9 @@
   volume    = {10},
   number    = {3},
   pages     = {400--404},
+  publisher = {American Society of Agricultural and Biological Engineers},
   doi       = {10.13031/2013.39683},
-  year      = {1967},
-  publisher = {American Society of Agricultural and Biological Engineers}
+  year      = {1967}
 }
 
 @article{rucker_gardner_2005,
@@ -219,6 +219,7 @@
 @article{wosten_pedotransfer_1999,
   author  = {W{\"o}sten, J. H. M. and Nemes, A. and Lilly, A. and Le Bas, C.},
   title   = {Development and Use of a Database of Hydraulic Properties of European Soils},
+  journal = {Microbiology},
   journal = {Geoderma},
   volume  = {90},
   pages   = {169--185},
@@ -231,9 +232,9 @@
   title       = {Waterretentie- en Doorlatendheidskarakteristieken van Boven- en Ondergronden in Nederland: De Staringreeks},
   institution = {Alterra},
   number      = {153},
-  year        = {2001},
   language    = {Dutch},
-  url         = {https://edepot.wur.nl/43272}
+  url         = {https://edepot.wur.nl/43272},
+  year        = {2001}
 }
 
 @article{cooper_pedotransfer_2021,
@@ -274,8 +275,8 @@
[…]
