+++ paper/paper.bib
@@ -52,6 +52,19 @@
   year    = {1931}
 }
 
+@article{mualem_model_1976,
+  author  = {Mualem, Yechezkel},
+  title   = {A new model for predicting the hydraulic conductivity of unsaturated porous media},
+  journal = {Water Resources Research},
+  volume  = {12},
+  number  = {3},
+  pages   = {513-522},
+  doi     = {https://doi.org/10.1029/WR012i003p00513},
+  year    = {1976}
+}
+
+
+
 @article{genuchten_mualem_1980,
   author  = {{van Genuchten}, M. Th.},
   title   = {A Closed-Form Equation for Predicting the Hydraulic Conductivity of Unsaturated Soils},
