### 2026-06-02 b8622b7
rename GenuchtenKool to Kool
