### 2026-06-02 719f131
trim the text a little bit
