+++ paper/paper.md
@@ -73,12 +73,12 @@ k = mg.k(h)  # hydraulic conductivity (cm/d) at pressure head values
 The object-oriented design and duck typing (relying on object methods rather than explicit types) provide a clear and consistent structure in which users can define custom soil model classes. Additionally, `pedon` only depends on well-maintained packages in the Python scientific ecosystem such as NumPy [@numpy_article_2020], SciPy [@scipy_paper_2020], Matplotlib [@matplotlib_paper_2007], and Pandas [@pandas_software_2020; @pandas_paper_2010].
 
 # Soil hydraulic parameters
-Soil hydraulic parameters depend on soil type and determine the shape of a soil model’s SWRC and HCF. They are rarely measured directly and are usually derived from reference datasets, empirical relationships, or laboratory measurements. `pedon` links these parameters to soil models and provides a framework to obtain them from existing datasets, easily measured soil properties, and direct measurements of soil water content and hydraulic conductivity.
+Soil hydraulic parameters determine the shape of a soil model’s SWRC and HCF, but are rarely measured directly. To address this, pedon links these parameters to soil models by providing a unified framework to derive them from reference datasets, empirical relationships, or direct (laboratory) measurements.
 
 ## Parameter datasets
-`pedon` includes a dataset of Brooks–Corey and Mualem–van Genuchten parameters for a wide range of soils. Currently, this dataset is compiled from three established soil hydraulic parameter databases:
+`pedon` includes a dataset of soil model parameters for a wide range of soils, currently compiled from four sources:
 
-- Average parameter values for twelve major soil textural groups defined by @carsel_dataset_1988, also used in the HYDRUS software for variably saturated flow modeling [@simunek_hydrus1d_2009];
+- Average Mualem-van Genuchten parameter values for twelve major soil textural groups defined by @carsel_dataset_1988, also used in the HYDRUS software for variably saturated flow modeling [@simunek_hydrus1d_2009];
 - A dataset from the VS2D software [@healy_vs2d_1990] containing both Brooks–Corey and Mualem–van Genuchten parameters;
 - The Staring series from the Netherlands [@wosten_staringreeks_2001; @heinen_staringreeks_2020; @heinen_bofek_2022], which describes soils using the Mualem–van Genuchten model based on hundreds of processed samples [@genuchten_mualem_1980; @wosten_texture_1988].
 - The @clapp_hornberger_1978 dataset, which provides Campbell soil model parameters.
@@ -141,9 +141,9 @@ bc = pe.SoilSample(h=h, theta=theta, k=k).fit(pe.Brooks)
 `pedon` contributes to the field of groundwater modeling by providing a modular, object-oriented framework that integrates hydraulic soil models, established parameter databases, and pedotransfer functions into a single, reproducible workflow. Existing tools such as unsatfit [@seki_unsatfit_2023], PySWR [@memari_pyswr_2021], and RETC [@genuchten_retc_1991] are utilities primarily focused on least-squares fitting of predefined soil models. In contrast, `pedon` enables researchers to define custom soil model formulations and systematically evaluate them against different datasets and pedotransfer functions. This level of interoperability is unique and provides a broader framework for the analysis of unsaturated soil hydraulic properties. At the same time, pedon is developed with a collaborative mindset, exemplified by the integration and extension of the soil model parameter estimation algorithms from HYPAGS [@peche_hypags_2023; @peche_genuchten_2024].
 
 # Research impact statement
[…]
