### 2026-06-02 a8ccb5b
remove some words
