+++ paper/paper.bib
@@ -97,13 +97,16 @@
   year    = {1988}
 }
 
-@manual{simunek_hydrus1d_2009,
-  author       = {Sim{\r{u}}nek, J. and {\v{S}}ejna, M. and Saito, H. and Sakai, M. and {van Genuchten}, M. Th.},
-  title        = {The {HYDRUS-1D} software package for simulating the one-dimensional movement of water, heat, and multiple solutes in variably-saturated media. {Version 4.08}},
-  organization = {University of California Riverside},
-  address      = {Riverside, CA},
-  url          = {https://www.pc-progress.com/downloads/pgm_hydrus1d/hydrus1d-4.08.pdf},
-  year         = {2009}
+
+@article{simunek_hydrus_2008,
+  author  = {Sim{\r{u}}nek, J. and {van Genuchten}, M. Th. and {\v{S}}ejna, M.},
+  title   = {Development and applications of the {HYDRUS} and {STANMOD} software packages and related codes},
+  journal = {Vadose Zone Journal},
+  volume  = {7},
+  number  = {2},
+  pages   = {587--600},
+  doi     = {10.2136/vzj2007.0077},
+  year    = {2008}
 }
 
 @article{fuentes_burdine_1992,
@@ -619,4 +622,26 @@
   pages   = {1--12},
   doi     = {10.1097/00010694-199001000-00001},
   year    = {1990}
+}
+
+@article{vereecken_modeling_2016,
+  author  = {Vereecken, H. and Schnepf, A. and Hopmans, J. W. and Javaux, M. and Or, D. and Roose, T. and Vanderborght, J. and Young, M. H. and Amelung, W. and Aitkenhead, M. and Allison, S. D. and Assouline, S. and Baveye, P. and Berli, M. and Br{\"u}ggemann, N. and Finke, P. and Flury, M. and Gaiser, T. and Govers, G. and Ghezzehei, T. and Hallett, P. and Hendricks Franssen, H. J. and Heppell, J. and Horn, R. and Huisman, J. A.},
+  title   = {Modeling Soil Processes: Review, Key Challenges, and New Perspectives},
+  journal = {Vadose Zone Journal},
+  volume  = {15},
+  number  = {7},
+  pages   = {vzj2016.04.0033},
+  doi     = {10.2136/vzj2015.09.0131},
+  year    = {2016}
+}
+
+@article{rawls_properties_1982,
+  author  = {Rawls, W. J. and Brakensiek, D. L. and Saxton, K. E.},
+  title   = {Estimation of Soil Water Properties},
+  journal = {Transactions of the ASAE},
+  volume  = {25},
+  number  = {5},
+  pages   = {1316--1320},
+  doi     = {10.13031/2013.33720},
+  year    = {1982}
 }
