+++ src/pedon/_typing.py
@@ -18,4 +18,6 @@ SoilModelNames = Literal[
     "GenuchtenGardner",
 ]
 
-SourceNames = Literal["HYDRUS", "VS2D", "Staring_2001", "Staring_2018", "Clapp"]
+SourceNames = Literal[
+    "HYDRUS", "VS2D", "Staring_2001", "Staring_2018", "Clapp", "Rawls"
+]
