+++ paper/paper.bib
@@ -635,7 +635,7 @@
   year    = {2016}
 }
 
-@article{rawls_properties_1982,
+@article{rawls_dataset_1982,
   author  = {Rawls, W. J. and Brakensiek, D. L. and Saxton, K. E.},
   title   = {Estimation of Soil Water Properties},
   journal = {Transactions of the ASAE},
