### 2026-06-03 2ce4517
add test for rawls dataset
