+++ tests/test_datasets.py
@@ -62,6 +62,17 @@ def test_soil_from_clapp() -> None:
     assert s.model.h_b == 3.5
     assert s.model.b == 3.0
 
+def test_soil_from_rawls() -> None:
+    """Test loading soil from the Rawls PTF."""
+    s = pe.soil.Soil("Sand").from_name(pe.Brooks, source="Rawls")
+    assert s.source == "Rawls"
+    assert s.description == "Sand"
+    assert isinstance(s.model, pe.Brooks)
+    assert s.model.k_s == 504.0
+    assert s.model.theta_r == 0.02
+    assert s.model.theta_s == 0.437
+    assert s.model.h_b == 7.26
+    assert s.model.l == 0.592
 
 def test_soil_from_staring_with_int_year() -> None:
     """Test integer year input for from_staring."""
