+++ src/pedon/soil.py
@@ -97,7 +97,7 @@ class SoilSample:
         default=None, repr=False
     )  # moisture content measurement
 
-    def from_staring(self, name: str, year: str = "2018") -> "SoilSample":
+    def from_staring(self, name: str, year: str = "2018") -> Self:
         """Get properties and measurements from Staring series.
 
         References
@@ -1047,6 +1047,136 @@ class SoilSample:
             theta_s=vgpar[1],
             alpha=10 ** vgpar[2],
             n=10 ** vgpar[3],
+        )
+
+    def rawls(self, cecc: float | None = None) -> Brooks:
+        """Estimate Brooks-Corey parameters using the Rawls & .
+
+        cecc: Cation Exchange Capacity of clay (cmol kg^-1). Required if `rho` is not provided.
+
+        References
+        ----------
+        Rawls, W.J. and Brakensiek, D.L. (1989). Estimation of Soil Water Retention
+        and Hydraulic Properties. In: Morel-Seytoux, H.J. (eds) Unsaturated Flow in
+        Hydrologic Modeling. https://doi.org/10.1007/978-94-009-2352-2_10
+
+        """
+        assert self.sand_p is not None, "Rawls PTF requires 'sand_p' to be set."
+        assert self.clay_p is not None, "Rawls PTF requires 'clay_p' to be set."
+        assert self.om_p is not None, "Rawls PTF requires 'om_p' to be set."
+
+        if self.rho is None:
+            assert cecc is not None, (
+                "Rawls PTF requires 'cecc' to be provided if 'rho' is not set."
+            )
+            cec = cecc / self.clay_p
+            print(cec)
+            assert 0.1 <= cec <= 0.9, (
+                f"Rawls PTF requires 'cecc/clay_p' to be between 0.1 and 0.9. Got {cec:.2f}."
+            )
+            bd = (
+                1.51 + 0.0025 * self.sand_p
+                - 0.0013 * self.sand_p * self.om_p
+                - 0.0006 * self.clay_p * self.om_p
+                - 0.0048 * self.clay_p * cec
+            )
+            theta_s = (2.65 - bd) / 2.65
+            eac = 1.0 - (
+                3.8
+                + 0.00019 * self.clay_p**2
+                - 0.0337 * self.sand_p # 0.337 in book, seems like a typod
+                + 0.126 * cec * self.clay_p
+                + self.om_p * (self.sand_p / 200.0) ** 2
+            ) / 100.0
+            # Rawls & Baumer 1989
+            theta_r = (
+                (0.2 + 0.1 * self.om_p + 0.25 * self.clay_p * cec**0.45) * bd / 100.0
+            )
+            c1 = (
+                0.17
+                + 0.181 * self.clay_p
+                - 0.00000069 * self.sand_p**2 * self.clay_p**2
+                - 0.00000041 * self.sand_p**2 * (100.0 - self.sand_p - self.clay_p) ** 2
+                + 0.000118 * self.sand_p**2 * bd**2
+                + 0.00069 * self.clay_p**2 * bd**2
+                + 0.000049 * self.sand_p**2 * self.clay_p
+                - 0.000085 * (100.0 - self.sand_p - self.clay_p) * self.clay_p**2
+            )
+            k_s = (
+                0.00035
+                * (theta_s * eac - theta_r) ** 3
+                / (1 - theta_s * eac) ** 2
+                * (bd / theta_r) ** 2
+                * c1**2
+            ) * 24.0  # cm/d
+        else:
+            bd = self.rho
+            theta_s = (2.65 - bd) / 2.65
+            theta_r = (
+                -0.0182482
+                + 0.00087269 * self.sand_p
[…]
