### 2026-06-03 c5a69f6
add tests for rawls
