### 2026-06-03 2b90b09
choose a more logical value for om_p
