+++ tests/test_ptf.py
@@ -12,7 +12,7 @@ REL = 1e-3
 def ss() -> pe.soil.SoilSample:
     """Fixture for a soil sample with specific properties for testing pedotransfer functions."""
     return pe.soil.SoilSample(
-        sand_p=50.0, silt_p=10.0, clay_p=40.0, rho=1.5, om_p=20.0, m50=150.0
+        sand_p=50.0, silt_p=10.0, clay_p=38.0, rho=1.5, om_p=2.0, m50=150.0
     )
 
 
@@ -20,94 +20,94 @@ def test_wosten(ss: pe.soil.SoilSample) -> None:
     """Test Wösten pedotransfer function with texture-dependent parameters."""
     sm = ss.wosten(topsoil=True)
     assert isinstance(sm, pe.Genuchten)
-    assert sm.k_s == pytest.approx(0.6547824638330241, rel=REL)
+    assert sm.k_s == pytest.approx(61.41854253249469, rel=REL)
     assert sm.theta_r == pytest.approx(0.01, rel=REL)
-    assert sm.theta_s == pytest.approx(0.3506329025688723, rel=REL)
-    assert sm.alpha == pytest.approx(0.0014029528494249415, rel=REL)
-    assert sm.n == pytest.approx(1.1209348010913704, rel=REL)
-    assert sm.l == pytest.approx(-3.6143956260446446, rel=REL)
-    assert sm.m == pytest.approx(0.10788745337697181, rel=REL)
+    assert sm.theta_s == pytest.approx(0.4016384367793987, rel=REL)
+    assert sm.alpha == pytest.approx(0.07053697798996453, rel=REL)
+    assert sm.n == pytest.approx(1.137061406702455, rel=REL)
+    assert sm.l == pytest.approx(-4.936831819765616, rel=REL)
+    assert sm.m == pytest.approx(0.12054002175655676, rel=REL)
 
 
 def test_wosten_no_topsoil(ss: pe.soil.SoilSample) -> None:
     """Test Wösten pedotransfer function without topsoil adjustment."""
     sm = ss.wosten(topsoil=False)
     assert isinstance(sm, pe.Genuchten)
-    assert sm.k_s == pytest.approx(0.10889706346920575, rel=REL)
+    assert sm.k_s == pytest.approx(10.843130926922838, rel=REL)
     assert sm.theta_r == pytest.approx(0.01, rel=REL)
-    assert sm.theta_s == pytest.approx(0.3522969025688723, rel=REL)
-    assert sm.alpha == pytest.approx(0.001298720038387063, rel=REL)
-    assert sm.n == pytest.approx(1.0907448358614613, rel=REL)
-    assert sm.l == pytest.approx(-3.6143956260446446, rel=REL)
-    assert sm.m == pytest.approx(0.08319529268253811, rel=REL)
+    assert sm.theta_s == pytest.approx(0.4033024367793987, rel=REL)
+    assert sm.alpha == pytest.approx(0.06618124289279576, rel=REL)
+    assert sm.n == pytest.approx(1.104333140070533, rel=REL)
+    assert sm.l == pytest.approx(-4.936831819765616, rel=REL)
+    assert sm.m == pytest.approx(0.09447614699298912, rel=REL)
 
 
 def test_wosten_sand(ss: pe.soil.SoilSample) -> None:
     """Test Wösten sand-specific pedotransfer function."""
     sm = ss.wosten_sand(topsoil=True)
     assert isinstance(sm, pe.Genuchten)
-    assert sm.k_s == pytest.approx(20.793, rel=REL)
[…]
