### 2026-06-03 fc277aa
ruff
