### 2026-06-03 fe4668a
save some words in parameter datasets
