+++ paper/paper.md
@@ -82,12 +82,13 @@ Soil hydraulic parameters determine the shape of a soil model’s SWRC and HCF, 
 - VS2D [@healy_vs2d_1990] and Rawls [@rawls_dataset_1982] for Brooks–Corey parameters.
 - Clapp [@clapp_hornberger_1978] for Campbell parameters.
 
-For example, a parameter set for a sandy soil (using the "B05" classification for the Staring series) can easily be obtained from any of the databases via the following code:
+For example, a parameter set for a sandy soil (using the "B05" classification for the Staring series) can easily be obtained from any of the databases via the following code, enabling direct comparison of the resulting SWRC's (Figure \ref{fig:dataset_swrc}).
+
 ```python
 hydrus = pe.Soil("Sand").from_name(pe.Genuchten, source="HYDRUS")
 staring = pe.Soil("B05").from_name(pe.Genuchten, source="Staring_2018")
-clapp = pe.Soil("Sand").from_name(pe.Campbell, source="Clapp")
 rawls = pe.Soil("Sand").from_name(pe.Brooks, source="Rawls")
+clapp = pe.Soil("Sand").from_name(pe.Campbell, source="Clapp")
 ```
 
 ![Comparison of SWRC's for a generic sandy soil using parameters retrieved from four different soil databases. \label{fig:dataset_swrc}](figures/dataset_swrc.png)
@@ -132,7 +133,7 @@ hypags: pe.Genuchten = pe.SoilSample(k=ks).hypags()
 ```
 
 ### Soil hydraulic measurements
-`pedon` can estimate soil model parameters directly when measurements of soil water retention and/or unsaturated hydraulic conductivity are available. A soil model, together with its SWRC and HCF, is fitted to the data by minimizing the difference between measured and simulated values. This uses a nonlinear least-squares algorithm from SciPy [@scipy_paper_2020] and follows the well-established methodology of the `RETC` software [@genuchten_retc_1991].
+`pedon` can estimate soil model parameters directly when measurements of soil water retention and/or unsaturated hydraulic conductivity are available. A soil model, together with its SWRC and HCF, is fitted to the data by minimizing the difference between measured and simulated values. This uses a nonlinear least-squares algorithm from SciPy [@scipy_paper_2020] and follows the well-established methodology of the RETC software [@genuchten_retc_1991].
 
 ### Soil model conversion
 The same fitting procedure can translate between soil models. The SWRC and HCF generated by one model are sampled over a range of pressure heads and refitted using another formulation. This enables direct model comparison (Figure \ref{fig:fit_swrc_hcf}) and facilitates integration with external tools when a different soil model is required [@vonk_nonlinear_2024].
@@ -148,7 +149,7 @@ bc = pe.SoilSample(h=h, theta=theta, k=k).fit(pe.Brooks)
 `pedon` contributes to the field of groundwater modeling by providing a modular, object-oriented framework that integrates hydraulic soil models, established parameter databases, and pedotransfer functions into a single, reproducible workflow. Existing tools such as unsatfit [@seki_unsatfit_2023], PySWR [@memari_pyswr_2021], and RETC [@genuchten_retc_1991] are utilities primarily focused on least-squares fitting of predefined soil models. In contrast, `pedon` enables researchers to define custom soil model formulations and systematically evaluate them against different datasets and pedotransfer functions. This level of interoperability is unique and provides a broader framework for the analysis of unsaturated soil hydraulic properties. At the same time, `pedon` is developed with a collaborative mindset, exemplified by the integration and extension of the soil model parameter estimation algorithms from HYPAGS [@peche_hypags_2023; @peche_genuchten_2024].
 
 # Research impact statement
[…]
