+++ paper/paper.bib
@@ -99,7 +99,7 @@
 
 
 @article{simunek_hydrus_2008,
-  author  = {Sim{\r{u}}nek, J. and {van Genuchten}, M. Th. and {\v{S}}ejna, M.},
+  author  = {{\v{S}}im{\r{u}}nek, J. and {van Genuchten}, M. Th. and {\v{S}}ejna, M.},
   title   = {Development and applications of the {HYDRUS} and {STANMOD} software packages and related codes},
   journal = {Vadose Zone Journal},
   volume  = {7},
@@ -644,4 +644,17 @@
   pages   = {1316--1320},
   doi     = {10.13031/2013.33720},
   year    = {1982}
+}
+
+@inbook{rawls_pedotransfer_1989,
+  author    = {Rawls, W. J. and Brakensiek, D. L.},
+  editor    = {Morel-Seytoux, H. J.},
+  title     = {Estimation of Soil Water Retention and Hydraulic Properties},
+  booktitle = {Unsaturated Flow in Hydrologic Modeling: Theory and Practice},
+  publisher = {Springer Netherlands},
+  address   = {Dordrecht},
+  pages     = {275--300},
+  isbn      = {978-94-009-2352-2},
+  doi       = {10.1007/978-94-009-2352-2_10},
+  year      = {1989}
 }
