### 2026-06-03 7f82144
add paper to documentation
