### 2026-06-03 3217902
Add local Docker run instructions to draft-pdf.yml
Added instructions for running the workflow locally with Docker.
