+++ paper/paper.md
@@ -37,25 +37,22 @@ Researchers and engineers working with unsaturated soils need estimates of soil 
 # Soil hydraulic models
 A soil hydraulic model (or soil model for short) is a parametric description of soil hydraulic functions: the soil water retention curve (SWRC) and the unsaturated hydraulic conductivity function (HCF). These relate soil water content and flow to pressure head and vice versa for use in variably saturated groundwater flow models. At this time, `pedon` provides the following soil models:
 
-- `pedon.Genuchten`: Mualem-van Genuchten [@mualem_model_1976; @genuchten_mualem_1980]
-- `pedon.Brooks`: Brooks-Corey [@brooks_corey_1964]
+- `pedon.Genuchten`: @mualem_model_1976 - @genuchten_mualem_1980
+- `pedon.Brooks`: @brooks_corey_1964
 - `pedon.Panday`: van Genuchten SWRC and Brooks-Corey HCF [@fuentes_burdine_1992; @panday_mfusgt_2026]
-- `pedon.Fredlund`: Fredlund-Xing [@fredlund_xing_1994]
-- `pedon.Haverkamp`: Haverkamp [@haverkamp_model_1977]
-- `pedon.Kosugi`: Kosugi [@kosugi_model_1996]
-- `pedon.Campbell`: Campbell [@campbell_model_1974]
-- `pedon.Gardner`: Gardner(-Kozeny) [@gardner_model_1958; @brutsaert_kozeny_1967; @bakker_gardner_2009; @mathias_gardner_2006]
+- `pedon.Fredlund`: @fredlund_xing_1994
+- `pedon.Haverkamp`: @haverkamp_model_1977
+- `pedon.Kosugi`: @kosugi_model_1996
+- `pedon.Campbell`: @campbell_model_1974
+- `pedon.Gardner`: @gardner_model_1958 HCF with a @kozeny_soil_1927 based SWRC [@brutsaert_kozeny_1967; @mathias_gardner_2006; @bakker_gardner_2009]
 - `pedon.Rucker`: van Genuchten-like SWRC and Gardner HCF [@rucker_gardner_2005]
 - `pedon.GenuchtenGardner`: van Genuchten SWRC and Gardner HCF [@genuchten_mualem_1980; @gardner_model_1958]
-- `pedon.Kool`: Mualem-van Genuchten SWRC and HCF with hysteresis in SWRC [@kool_parker_1987]
+- `pedon.Kool`: Mualem-van Genuchten SWRC and HCF with hysteresis [@kool_parker_1987]
 
 ## Software design
 The soil models are implemented as Python classes with model-specific methods for evaluating the SWRC and HCF. For example, the Mualem–van Genuchten soil model can be used as follows:
 
 ```python
-import numpy as np
-import pedon as pe
-
 # Mualem-van Genuchten parameters for Sandy Loam
 mg = pe.Genuchten(
     k_s=106.1,  # saturated conductivity (cm/d)
@@ -65,7 +62,7 @@ mg = pe.Genuchten(
     n=1.89,  # shape parameter (-)
 )
 
-h = np.logspace(-1, 4, 6)  # pressure head (cm)
+h = np.logspace(-1, 6, 8)  # pressure head (cm)
 theta = mg.theta(h)  # water content (-) at pressure head values
 k = mg.k(h)  # hydraulic conductivity (cm/d) at pressure head values
 ```
@@ -139,8 +136,8 @@ hypags: pe.Genuchten = pe.SoilSample(k=ks).hypags()
[…]
