+++ paper/paper.bib
@@ -658,3 +658,13 @@
   doi       = {10.1007/978-94-009-2352-2_10},
   year      = {1989}
 }
+
+@article{kozeny_soil_1927,
+  author   = {Kozeny, Josef},
+  title    = {Ueber kapillare Leitung des Wassers im Boden},
+  journal  = {Sitzungsberichte der Akademie der Wissenschaften in Wien, Mathematisch-Naturwissenschaftliche Klasse, Abteilung IIa},
+  language = {German},
+  volume   = {136},
+  pages    = {271--306},
+  year     = {1927}
+}
