### 2026-06-03 b8fc577
update gardner docstring references
