+++ paper/paper.bib
@@ -276,7 +276,7 @@
 
 @article{schaap_rosetta_2001,
   author  = {Schaap, M. G. and Leij, F. J. and {van Genuchten}, M. Th.},
-  title   = {Rosetta: A Computer Program for Estimating Soil Hydraulic Parameters with Hierarchical Pedotransfer Functions},
+  title   = {{Rosetta}: A Computer Program for Estimating Soil Hydraulic Parameters with Hierarchical Pedotransfer Functions},
   journal = {Journal of Hydrology},
   volume  = {251},
   number  = {3--4},
