### 2026-06-03 5d5c20f
make sure references in docstring are complete
