+++ src/pedon/soil.py
@@ -102,13 +102,15 @@ class SoilSample:
 
         References
         ----------
-        Wösten, J. H. M. and Veerman, G. J. and de Groot, W. J. M. and Stolte, J. (2001).
+        Wösten, J. H. M., Veerman, G. J., de Groot, W. J. M., & Stolte, J. (2001).
         Waterretentie- en Doorlatendheidskarakteristieken van Boven- en Ondergronden in
-        Nederland: De Staringreeks. url: https://edepot.wur.nl/43272
+        Nederland: De Staringreeks. Alterra, Report 153, Wageningen, The Netherlands.
+        url: https://edepot.wur.nl/43272
 
-        Heinen, M. and Bakker, G. and Wösten, J. H. M. (2020). Waterretentie- en
+        Heinen, M., Bakker, G., & Wösten, J. H. M. (2020). Waterretentie- en
         Doorlatendheidskarakteristieken van Boven- en Ondergronden in Nederland:
-        De Staringreeks; (Update 2018). doi: 10.18174/512761
+        De Staringreeks; (Update 2018). Wageningen Environmental Research, Report 2978,
+        Wageningen, The Netherlands. doi: 10.18174/512761
 
         """
         if year not in ("2001", "2018"):
@@ -182,8 +184,9 @@ class SoilSample:
 
         References
         ----------
-        van Genuchten, M. Th. and Leij, F. J. and Yates, S. R. (1991). The RETC Code for
-        Quantifying the Hydraulic Functions.
+        van Genuchten, M. Th., Leij, F. J., & Yates, S. R. (1991). The RETC Code for
+        Quantifying the Hydraulic Functions. Version 1.0. U.S. Salinity Laboratory,
+        USDA, ARS, Riverside, CA. EPA Report 600/2-91/065.
         url: https://cfpub.epa.gov/si/si_public_record_report.cfm?dirEntryId=130162
 
         """
@@ -269,8 +272,9 @@ class SoilSample:
 
         References
         ----------
-        Wosten et al (1999) - Development and use of a database of hydraulic
-        properties of European soils. doi: 10.1016/S0016-7061(98)00132-3
+        Wösten, J. H. M., Nemes, A., Lilly, A., & Le Bas, C. (1999). Development and
+        use of a database of hydraulic properties of European soils. Geoderma, 90, 169--185.
+        doi: 10.1016/S0016-7061(98)00132-3
 
         """
         assert (
@@ -374,8 +378,11 @@ class SoilSample:
             If True, applies the topsoil adjustment to the pedotransfer
             function. Default is False.
 
-        Wosten et al. (2001) - Waterretentie- en doorlatendheidskarakteristieken
-        van boven- en ondergronden in Nederland: de Staringreeks.
+        References
+        ----------
+        Wösten, J. H. M., Veerman, G. J., de Groot, W. J. M., & Stolte, J. (2001).
+        Waterretentie- en Doorlatendheidskarakteristieken van Boven- en Ondergronden in
+        Nederland: De Staringreeks. Alterra, Report 153, Wageningen, The Netherlands.
         url: https://edepot.wur.nl/43272
 
         """
@@ -461,8 +468,9 @@ class SoilSample:
 
         References
         ----------
-        Wosten et al. (2001) - Waterretentie- en doorlatendheidskarakteristieken
-        van boven- en ondergronden in Nederland: de Staringreeks.
+        Wösten, J. H. M., Veerman, G. J., de Groot, W. J. M., & Stolte, J. (2001).
+        Waterretentie- en Doorlatendheidskarakteristieken van Boven- en Ondergronden in
+        Nederland: De Staringreeks. Alterra, Report 153, Wageningen, The Netherlands.
[…]
