### 2026-06-03 b98a7c4
cleanup ptfs to get rid of some rounding
