+++ src/pedon/soil.py
@@ -299,7 +299,7 @@ class SoilSample:
             - 0.001183 * self.rho * self.om_p
             - 0.0001664 * ts * self.silt_p
         )
-        alpha_ = (
+        alpha = exp(
             -14.96
             + 0.03135 * self.clay_p
             + 0.0351 * self.silt_p
@@ -316,24 +316,27 @@ class SoilSample:
             - 0.4852 * self.rho * self.om_p
             + 0.00673 * ts * self.clay_p
         )
-        n_ = (
-            -25.23
-            - 0.02195 * self.clay_p
-            + 0.0074 * self.silt_p
-            - 0.1940 * self.om_p
-            + 45.5 * self.rho
-            - 7.24 * self.rho**2
-            + 0.0003658 * self.clay_p**2
-            + 0.002885 * self.om_p**2
-            - 12.81 * self.rho**-1
-            - 0.1524 * self.silt_p**-1
-            - 0.01958 * self.om_p**-1
-            - 0.2876 * log(self.silt_p)
-            - 0.0709 * log(self.om_p)
-            - 44.6 * log(self.rho)
-            - 0.02264 * self.rho * self.clay_p
-            + 0.0896 * self.rho * self.om_p
-            + 0.00718 * ts * self.clay_p
+        n = (
+            exp(
+                -25.23
+                - 0.02195 * self.clay_p
+                + 0.0074 * self.silt_p
+                - 0.1940 * self.om_p
+                + 45.5 * self.rho
+                - 7.24 * self.rho**2
+                + 0.0003658 * self.clay_p**2
+                + 0.002885 * self.om_p**2
+                - 12.81 * self.rho**-1
+                - 0.1524 * self.silt_p**-1
+                - 0.01958 * self.om_p**-1
+                - 0.2876 * log(self.silt_p)
+                - 0.0709 * log(self.om_p)
+                - 44.6 * log(self.rho)
+                - 0.02264 * self.rho * self.clay_p
+                + 0.0896 * self.rho * self.om_p
+                + 0.00718 * ts * self.clay_p
+            )
+            + 1
         )
         l_ = (
             0.0202
@@ -344,7 +347,8 @@ class SoilSample:
             + 0.00283 * self.rho * self.silt_p
             + 0.0488 * self.rho * self.om_p
         )
-        ks_ = (
+        lb = (10 * exp(l_) - 10) / (1 + exp(l_))
+        k_s = exp(
             7.755
             + 0.0352 * self.silt_p
             + 0.93 * ts
@@ -359,14 +363,13 @@ class SoilSample:
             + 0.02986 * ts * self.clay_p
             - 0.03305 * ts * self.silt_p
         )
-        theta_r = 0.01
         return Genuchten(
-            k_s=max(exp(ks_), 0),
-            theta_r=theta_r,
+            k_s=k_s,
+            theta_r=0.01,
             theta_s=theta_s,
-            alpha=exp(alpha_),
-            n=exp(n_) + 1,
-            l=(10 * exp(l_) - 10) / (1 + exp(l_)),
+            alpha=alpha,
+            n=n,
+            l=lb,
         )
 
     def wosten_sand(self, topsoil: bool = False) -> Genuchten:
@@ -452,15 +455,15 @@ class SoilSample:
             + 2.364 * self.silt_p**-1
             + 1.014 * log(self.silt_p)
         )
-        l = min(max(2 * (exp(l_) - 1) / (exp(l_) + 1), -2.0), 2.0)  # noqa: E741
[…]
