### 2026-06-05 f121d15
Add Brunswick (BW-VGM) soil hydraulic property model
