+++ src/pedon/soilmodel.py
@@ -1,14 +1,16 @@
 """Soil models for soil water retention and hydraulic conductivity."""
 
 from dataclasses import dataclass, field
-from typing import Protocol, runtime_checkable
+from typing import Protocol, runtime_checkable, Callable
 from warnings import warn
 
 import matplotlib.pyplot as plt
 from numpy import abs as npabs
-from numpy import asarray, exp, full, linspace, log, log10, maximum, sqrt
+from numpy import asarray, exp, full, linspace, log, log10, maximum, sqrt, power, where, clip, isnan
 from scipy.integrate import trapezoid
 from scipy.special import erfc, erfcinv, lambertw
+from scipy.interpolate import PchipInterpolator
+from scipy.optimize import brentq
 
 from ._typing import FloatArray, SoilModelNames
 from .plot import hcf
@@ -26,8 +28,7 @@ def plot_hcf(*args, **kwargs):
         stacklevel=2,
     )
     return hcf(*args, **kwargs)
-
-
+    
 @runtime_checkable
 class SoilModel(Protocol):
     """Protocol for soil models.
@@ -265,6 +266,241 @@ class Brooks:
     def plot(self, ax: plt.Axes | None = None) -> plt.Axes:
         return plot_swrc(self, ax=ax)
 
+def _sc_vgm(h: FloatArray, alpha: float, n: float) -> FloatArray:
+    """Capillary effective saturation S_c(h) using van Genuchten (1980).
+ 
+    Weber et al. (2019) Eq. 5 – BW-VGM variant.
+ 
+    Parameters
+    ----------
+    h : FloatArray
+        Pressure head [cm], convention h ≤ 0 for unsaturated conditions.
+    alpha : float
+        van Genuchten α parameter [1/cm].
+    n : float
+        van Genuchten n parameter [–].
+    """
+    h = asarray(h, dtype=float)
+    m = 1.0 - 1.0 / n
+    return where(
+        h <= 0,
+        1.0,
+        power(1.0 + power(alpha * npabs(h), n), -m),
+    )
+
+
+def _build_snc_interpolator(alpha: float, n: float) -> Callable:
+    """Build a monotone PCHIP interpolator for S_nc(h).
+ 
+    S_nc is 1 for h ≤ h_d and decreases linearly to 0 in log10(h)-space at
+    pF = 6.8 (oven-dry), following Weber et al. (2019) Eq. 6–7.
+ 
+    The desaturation head h_d = 100/α is the BW-VGM specific value placing
+    the noncapillary plateau two log-units above the VGM air-entry head 1/α.
+ 
+    The grid extends from pF = −2 (h = 0.01 cm) to pF = 6.8 so that the
+    interpolator covers all physically relevant wet-end values.
+    """
+    pF = linspace(-2.0, 6.8, 600)
+    h_vals = power(10.0, pF)
+    h_d = 10.0 ** (log10(1.0 / alpha) + 2.0)   # = 100 / alpha
+ 
+    snc_vals = where(
+        h_vals <= h_d,
+        1.0,
+        maximum(
+            0.0,
+            1.0 - (log10(h_vals) - log10(h_d)) / (6.8 - log10(h_d)),
+        ),
+    )
+    return PchipInterpolator(h_vals, snc_vals, extrapolate=False)
+
+
+@dataclass
+class Brunswick:
+    """Brunswick (BW-VGM) soil hydraulic property model.
+ 
+    Partitions pore space into capillary (c) and noncapillary (nc) fractions,
+    overcoming the inability of the van Genuchten–Mualem (VGM) model to
+    describe hydraulic properties in the dry range.
+ 
+    Parameters
+    ----------
+    k_sc : float
+        Saturated capillary hydraulic conductivity K_sc [L/T].
+    k_snc : float
[…]
