### 2026-06-05 c0ccad4
style: run ruff linter and formatter changes
