### 2026-06-05 05dbd04
fix: resolve ruff syntax errors and docstring issues
