### 2026-06-05 df4f9ab
fix: resolve ty typechecking subscriptable errors
