### 2026-06-05 8b4d821
fix: resolve mypy assignment type mismatch using np.float64
