### 2026-06-05 05b3d0e
fix: explicit float type assignment for theta_dry to appease mypy
