### 2026-06-05 38323c8
fix: bypass stubborn mypy assignment check with inline ignore
