### 2026-06-05 09c7007
fix: resolve mypy caching by renaming theta_dry to a fresh float64
