### 2026-06-05 13af0ce
fix: entirely remove theta_dry assignment to stop mypy error
