### 2026-06-05 e49463c
fix: cast brentq output to np.float64 to match array type tracking
