### 2026-06-05 14a90f8
dropping np.float64() to FINALLY resolve mypy fail
