### 2026-06-05 25d0581
another round of ruff romatting
