### 2026-06-05 4277e85
another go at mypy fail...new function for h() method
