### 2026-06-05 fc6aff2
another round of ruff formatting
