### 2026-06-05 1a63b04
start with Gerke model
