+++ src/pedon/soilmodel.py
@@ -1,6 +1,7 @@
 """Soil models for soil water retention and hydraulic conductivity."""
 
 from dataclasses import dataclass, field
+from logging import getLogger
 from typing import Protocol, runtime_checkable
 from warnings import warn
 
@@ -8,11 +9,14 @@ import matplotlib.pyplot as plt
 from numpy import abs as npabs
 from numpy import asarray, exp, full, linspace, log, log10, maximum, sqrt
 from scipy.integrate import trapezoid
+from scipy.optimize import brentq
 from scipy.special import erfc, erfcinv, lambertw
 
 from ._typing import FloatArray, SoilModelNames
 from .plot import hcf
 from .plot import swrc as plot_swrc
+
+logger = getLogger(__name__)
 
 
 def plot_hcf(*args, **kwargs):
@@ -937,6 +941,163 @@ class Kool:
         return plot_swrc(self, ax=ax)
 
 
+@dataclass
+class Gerke:
+    """Gerke and van Genuchten (1993) Dual-Porosity Soil Model.
+
+    Parameters
+    ----------
+    k_sf: float
+        Saturated hydraulic conductivity of the fracture pore system [L/T].
+    theta_rf: float
+        Residual soil water content of the fracture pore system [-].
+    theta_sf: float
+        Saturated soil water content of the fracture pore system [-].
+    alpha_f: float
+        Inverse of the air-entry pressure for the fracture pore system [1/L].
+    n_f: float
+        Pore-size distribution parameter for the fracture pore system [-].
+    l_f: float
+        Pore-connectivity parameter for the fracture pore system [-].
+    k_sm: float
+        Saturated hydraulic conductivity of the matrix pore system [L/T].
+    theta_rm: float
+        Residual soil water content of the matrix pore system [-].
+    theta_sm: float
+        Saturated soil water content of the matrix pore system [-].
+    alpha_m: float
+        Inverse of the air-entry pressure for the matrix pore system [1/L].
+    n_m: float
+        Pore-size distribution parameter for the matrix pore system [-].
+    l_m: float
+        Pore-connectivity parameter for the matrix pore system [-].
+    w_f: float
+        Volumetric weighting factor for the fracture domain (V_f / V_t) [-].
+
+    References
+    ----------
+    Gerke, H. H., & van Genuchten, M. Th. (1993). A dual-porosity model for simulating the
+    preferential movement of water and solutes in structured porous media. Water Resources
+    Research, 29(2), 305--319. doi: 10.1029/92WR02611
+
+    """
+
+    k_sf: float
+    theta_rf: float
+    theta_sf: float
+    alpha_f: float
+    n_f: float
+    k_sm: float
+    theta_rm: float
+    theta_sm: float
+    alpha_m: float
+    n_m: float
+    w_f: float  # Volumetric weighting factor for the fracture domain
+    l_f: float = 0.5
+    l_m: float = 0.5
+    fracture: Genuchten = field(init=False, repr=False)
+    matrix: Genuchten = field(init=False, repr=False)
+
+    def __post_init__(self):
+        """Construct the sub-domain Genuchten models from raw parameters."""
+        self.fracture = Genuchten(
+            k_s=self.k_sf,
+            theta_r=self.theta_rf,
+            theta_s=self.theta_sf,
+            alpha=self.alpha_f,
+            n=self.n_f,
+            l=self.l_f,
+        )
+        self.matrix = Genuchten(
+            k_s=self.k_sm,
+            theta_r=self.theta_rm,
+            theta_s=self.theta_sm,
+            alpha=self.alpha_m,
+            n=self.n_m,
+            l=self.l_m,
+        )
+
+    @property
+    def w_m(self) -> float:
+        """Volumetric weighting factor for the matrix domain."""
[…]
