### 2026-06-05 a1eee61
clean up the code a little bit
