### 2026-06-05 df5c362
add ref to Diamantopoulos
