### 2026-06-07 799ecef
split k methods for brunswick
