### 2026-06-07 f48763d
add brunswick tests
