### 2026-06-07 29051ad
add tests for brunswick
