+++ tests/test_soilmodel.py
@@ -765,15 +765,15 @@ def test_theta_brunswick(brunswick: pe.Brunswick) -> None:
     """Test water content calculation for the Brunswick model."""
     expected = array(
         [
-            0.4299964564950973,
-            0.4299553180235971,
-            0.4294409651859182,
-            0.4234860792740441,
-            0.3838829869422752,
-            0.3026052868139805,
-            0.2169877079249958,
-            0.1358848269504815,
-            0.05901260369850514,
+            0.4999991703271416,
+            0.4999668602375778,
+            0.4986879600880817,
+            0.4569001652583161,
+            0.2292138441347721,
+            0.1007588362446942,
+            0.05524125928920474,
+            0.03159644072915797,
+            0.013452718105112795,
         ]
     )
     assert_close(brunswick.theta(h=h), expected)
@@ -783,15 +783,15 @@ def test_s_brunswick(brunswick: pe.Brunswick) -> None:
     """Test degree of saturation calculation for the Brunswick model."""
     expected = array(
         [
-            0.999991770448857,
-            0.9998960767990403,
-            0.9986999306182265,
-            0.9848513419865673,
-            0.8927511370748727,
-            0.7037332388697222,
-            0.5046225807093044,
-            0.3160112250011194,
-            0.1372386039267259,
+            0.9999983399655377,
+            0.9999337204751556,
+            0.9973759201761635,
+            0.9138003475166322,
+            0.4584276777446077,
+            0.2015176724893884,
+            0.11048251857840948,
+            0.06319288145831595,
+            0.02690543621022558,
         ]
     )
     assert_close(brunswick.s(h=h), expected)
@@ -801,15 +801,15 @@ def test_k_brunswick(brunswick: pe.Brunswick) -> None:
     """Test hydraulic conductivity calculation for the Brunswick model."""
     expected = array(
         [
-            6.643330678829654,
-            6.069187965262085,
-            5.500031787890718,
-            4.819180158250977,
-            2.8161051083697294,
-            0.40209260425917124,
-            0.010016240370640382,
-            5.707812304107276e-05,
-            1.0014906168992358e-07,
+            97.92956247426391,
+            91.85479004815869,
+            69.58948878220202,
+            15.721644887915476,
+            0.04643513652798862,
+            5.159091342945608e-05,
+            3.879127256802016e-07,
+            3.406731262036317e-09,
+            2.890078846460336e-11,
         ]
     )
     assert_close(brunswick.k(h=h), expected)
@@ -818,7 +818,7 @@ def test_k_brunswick(brunswick: pe.Brunswick) -> None:
 def test_h_brunswick(brunswick: pe.Brunswick) -> None:
     """Test pressure head calculation for the Brunswick model."""
     expected = array(
-        [288837.3076926304, 16022.374305032258, 1071.412573040744, 54.02731930871524]
+        [1025.3797500583578, 141.1015529468394, 50.26648558718838, 20.517268056602314]
     )
     assert_close(brunswick.h(theta=theta), expected)
 
