### 2026-06-07 00a7c42
ruff
