+++ src/pedon/_typing.py
@@ -16,6 +16,7 @@ SoilModelNames = Literal[
     "Campbell",
     "Kool",
     "GenuchtenGardner",
+    "Gerke",
 ]
 
 SourceNames = Literal[
