### 2026-06-08 abc933b
add gerke and brunswick soil models
