+++ paper/paper.bib
@@ -660,7 +660,7 @@
 }
 
 @article{kozeny_soil_1927,
-  author   = {Kozeny, Josef},
+  author   = {Kozeny, J.},
   title    = {Ueber kapillare Leitung des Wassers im Boden},
   journal  = {Sitzungsberichte der Akademie der Wissenschaften in Wien, Mathematisch-Naturwissenschaftliche Klasse, Abteilung IIa},
   language = {German},
@@ -668,3 +668,38 @@
   pages    = {271--306},
   year     = {1927}
 }
+
+@article{gerke_dualporosity_1993,
+  author  = {Gerke, H. H. and van Genuchten, M. Th.},
+  title   = {A dual-porosity model for simulating the preferential movement of water and solutes in structured porous media},
+  journal = {Water Resources Research},
+  volume  = {29},
+  number  = {2},
+  pages   = {305-319},
+  doi     = {10.1029/92WR02339},
+  year    = {1993}
+}
+
+@article{weber_brunswick_2019,
+  author  = {Weber, T. K. D. and Durner, W. and Streck, T. and Diamantopoulos, E.},
+  title   = {A Modular Framework for Modeling Unsaturated Soil Hydraulic Properties Over the Full Moisture Range},
+  journal = {Water Resources Research},
+  volume  = {55},
+  number  = {6},
+  pages   = {4994-5011},
+  doi     = {10.1029/2018WR024584},
+  year    = {2019}
+}
+
+
+@article{weber_brunswick_2020,
+  author  = {Weber, T. K. D. and Finkel, M. and {da Conceição Gonçalves}, M. and Vereecken, H. and Diamantopoulos, E.},
+  title   = {Pedotransfer Function for the Brunswick Soil Hydraulic Property Model and Comparison to the van Genuchten-Mualem Model},
+  journal = {Water Resources Research},
+  volume  = {56},
+  number  = {9},
+  pages   = {e2019WR026820},
+  doi     = {10.1029/2019WR026820},
+  year    = {2020}
+}
+
