+++ _sources/_api/pedon.rst.txt
@@ -7,6 +7,7 @@ Submodules
 .. toctree::
    :maxdepth: 2
 
+   pedon.plot
    pedon.soil
    pedon.soilmodel
 
