+++ _sources/examples/04_hypags.ipynb.txt
@@ -1,157 +1,188 @@
 {
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "e9486bc5",
-   "metadata": {},
-   "source": [
-    "# Parameter Estimation via HYPAGS\n",
-    "\n",
-    "*Aaron Peche & Martin Vonk (2025)*\n",
-    "\n",
-    "## Overview\n",
-    "\n",
-    "HYPAGS (HYdraulic Parameters And Grain Sizes) is a parameter estimation method that can estimate full van Genuchten soil hydraulic parameters from minimal input data:\n",
-    "- **A single saturated conductivity value (k)** or **representative grain diameters (d10/d20)**\n",
-    "\n",
-    "This is useful when:\n",
-    "- You have measured soil conductivity but no water retention curve\n",
-    "- You only have grain size information from grain size analysis\n",
-    "- You need a quick estimate but lack the data for pedotransfer functions\n",
-    "\n",
-    "Peche, A., Houben, G. and Altfelder, S. (2024), Approximation of van Genuchten Parameter Ranges from Hydraulic Conductivity Data. Groundwater, 62: 469-479. https://doi.org/10.1111/gwat.13365"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "31240834",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "import matplotlib.pyplot as plt\n",
-    "\n",
-    "import pedon as pe"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "2118579c",
-   "metadata": {},
-   "source": [
-    "## Inputs Available for HYPAGS\n",
-    "\n",
-    "HYPAGS can work with different types of input data. The method will use whichever measurement is provided. \n",
-    "\n",
-    "The accuracy depends on the input type. Conductivity measurements are generally more reliable because they directly affect water movement, while grain size is an indirect proxy.\n",
-    "\n",
-    "## Example 1: Estimate from Saturated Conductivity"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "96011f57",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "# Estimate from saturated conductivity\n",
-    "# Note: HYPAGS expects k in m/s\n",
-    "k_value = 1e-5  # m/s\n",
-    "ss_k = pe.SoilSample(k=k_value)\n",
-    "hyp_k = ss_k.hypags()\n",
-    "\n",
-    "print(\"Estimated van Genuchten parameters from k value:\")\n",
-    "print(hyp_k)"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "314319fd",
-   "metadata": {},
-   "source": [
-    "## Example 2: Estimate from the Representative D10 Grain Diameter"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "19f95608",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "# Estimate from grain diameter\n",
-    "# d10: 10th percentile diameter\n",
-    "d10_value = 1e-5  # m\n",
-    "ss_d10 = pe.SoilSample(d10=d10_value)\n",
-    "hyp_d10 = ss_d10.hypags()\n",
-    "\n",
-    "print(\"Estimated van Genuchten parameters from d10:\")\n",
-    "print(hyp_d10)"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "43867e55",
-   "metadata": {},
-   "source": [
[…]
