+++ _sources/examples/09_paper.ipynb.txt
@@ -1,0 +1,373 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "3979c593",
+   "metadata": {},
+   "source": [
+    "# Article for Journal of Open Source Software\n",
+    "Martin Vonk & Aaron Peche (2026)\n",
+    "\n",
+    "This notebook replicates the results presented in the article submitted to the Journal of Open Source Software (JOSS). The article can be found here: https://joss.theoj.org/papers/095927c33334e8ebb4612243ea40fdae doi.org/...\n",
+    "\n",
+    "JOSS is a developer-friendly, open-access academic journal (ISSN 2475-9066) dedicated to research software packages and features a formal peer-review process. The pre-review and review of the Pedon package are publicly available in issues https://github.com/openjournals/joss-reviews/issues/9769 and https://github.com/openjournals/joss-reviews/issues/10409, respectively.\n",
+    "\n",
+    "## Setup"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "id": "906c9e13",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "from pathlib import Path\n",
+    "\n",
+    "import matplotlib as mpl\n",
+    "import matplotlib.pyplot as plt\n",
+    "import matplotlib.ticker as mplticker\n",
+    "import numpy as np\n",
+    "from cycler import cycler\n",
+    "\n",
+    "import pedon as pe\n",
+    "\n",
+    "rc_params = {\n",
+    "    \"axes.prop_cycle\": cycler(\n",
+    "        color=[\n",
+    "            \"#3f90da\",\n",
+    "            \"#ffa90e\",\n",
+    "            \"#bd1f01\",\n",
+    "            \"#94a4a2\",\n",
+    "            \"#832db6\",\n",
+    "            \"#a96b59\",\n",
+    "            \"#e76300\",\n",
+    "            \"#b9ac70\",\n",
+    "            \"#717581\",\n",
+    "            \"#92dadd\",\n",
[…]
