+++ _sources/examples/08_modflow.ipynb.txt
@@ -1,0 +1,332 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "de12f680",
+   "metadata": {},
+   "source": [
+    "# Generating Input for MODFLOW\n",
+    "\n",
+    "This notebook shows how to convert hydraulic soil properties from `pedon` models into parameters that can be used in FloPy-based MODFLOW workflows.\n",
+    "\n",
+    "The workflow is:\n",
+    "1. Build or load a soil hydraulic model in `pedon`.\n",
+    "2. Evaluate water retention and conductivity over a pressure-head range.\n",
+    "3. Fit a MODFLOW-oriented parameterization (here: `Panday`).\n",
+    "4. Create MODFLOW-USG / MODFLOW 6 package inputs.\n",
+    "\n",
+    "## FloPy references\n",
+    "- Bakker, M., Post, V., Langevin, C. D., Hughes, J. D., White, J. T., Starn, J. J., & Fienen, M. N. (2016). *Scripting MODFLOW Model Development Using Python and FloPy*. Groundwater, 54(5), 733-739. https://doi.org/10.1111/gwat.12413/epdf\n",
+    "- Hughes, J. D., Langevin, C. D., Paulinski, S. R., Larsen, J. D., & Brakenhoff, D. (2024). *FloPy Workflows for Creating Structured and Unstructured MODFLOW Models*. Groundwater, 62, 124-139. https://doi.org/10.1111/gwat.13327"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "id": "0b6f8fd6",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import flopy as fp\n",
+    "import numpy as np\n",
+    "import pandas as pd\n",
+    "\n",
+    "import pedon as pe"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "eee696d7",
+   "metadata": {},
+   "source": [
+    "## Converting soil models\n",
+    "\n",
[…]
