+++ _sources/examples/07_hysteresis.ipynb.txt
@@ -1,0 +1,163 @@
+{
+    "cells": [
+        {
+            "cell_type": "markdown",
+            "id": "a97813c2",
+            "metadata": {},
+            "source": [
+                "# Hysteresis Model\n",
+                "\n",
+                "## Packages"
+            ]
+        },
+        {
+            "cell_type": "code",
+            "execution_count": 1,
+            "id": "b9dc3c04",
+            "metadata": {},
+            "outputs": [],
+            "source": [
+                "import matplotlib.pyplot as plt\n",
+                "\n",
+                "import pedon as pe"
+            ]
+        },
+        {
+            "cell_type": "markdown",
+            "id": "ce1c0ccc",
+            "metadata": {},
+            "source": [
+                "## Implementation\n",
+                "\n",
+                "There is only one hysteresis model available in `pedon` which is based on the work of Kool and Parker (1987). This model is implemented in the `Kool` class. The parameters for this model are the same as for the `Genuchten` model, but with an additional parameter $\\xi$ which controls the degree of hysteresis by multiplying the (dry) air-entry pressure $\\alpha$. The 'ink-bottle-effect' (pores empty at larger suction than filling up) leads to a shift of the primary wetting curve compared with the primary drying curve. The van Genuchten shape parameters n and m remain similar for both, drying and wetting (Bashir et al., 2015). Also, no air entrapment is assumed - values for the residual and saturated water contents remain similar (Kool & Parker, 1987; Bashir et al., 2015). The following table from Bashir et al. (2015) shows the parameters for the three soil types used i
[…]

[…]
