+++ paper/paper.bib
@@ -16,7 +16,7 @@
   volume  = {62},
   number  = {3},
   pages   = {469--479},
-  doi     = {10.1111/gwat.133365},
+  doi     = {10.1111/gwat.13365},
   year    = {2024}
 }
 
@@ -59,11 +59,9 @@
   volume  = {12},
   number  = {3},
   pages   = {513-522},
-  doi     = {https://doi.org/10.1029/WR012i003p00513},
+  doi     = {WR012i003p00513},
   year    = {1976}
 }
-
-
 
 @article{genuchten_mualem_1980,
   author  = {{van Genuchten}, M. Th.},
@@ -690,7 +688,6 @@
   doi     = {10.1029/2018WR024584},
   year    = {2019}
 }
-
 
 @article{weber_brunswick_2020,
   author  = {Weber, T. K. D. and Finkel, M. and {da Conceição Gonçalves}, M. and Vereecken, H. and Diamantopoulos, E.},
