### 2026-06-08 5bfdc8b
bump version to 0.1.0
