+++ src/pedon/_version.py
@@ -1,7 +1,7 @@
 from importlib import metadata
 from platform import python_version
 
-__version__ = "0.0.12"
+__version__ = "0.1.0"
 
 
 def get_versions() -> dict[str, str]:
