+++ paper/paper.bib
@@ -59,7 +59,7 @@
   volume  = {12},
   number  = {3},
   pages   = {513-522},
-  doi     = {WR012i003p00513},
+  doi     = {10.1029/WR012i003p00513},
   year    = {1976}
 }
 
