+++ paper/paper.md
@@ -138,7 +138,8 @@ hypags: pe.Genuchten = pe.SoilSample(k=ks).hypags()
 The same fitting procedure can translate between soil models. The SWRC and HCF generated by one model are sampled over a range of pressure heads and refitted using another formulation. This enables direct model comparison (Figure \ref{fig:fit_swrc_hcf}) and facilitates integration with external tools when a different soil model is required [@vonk_nonlinear_2024].
 
 ```python
-# Fitting a Brooks-Corey soil model to the SWRC and HCF of a Mualem-van Genuchten soil model
+# Fitting a Brooks-Corey soil model to the SWRC and HCF of a
+# Mualem-van Genuchten soil model
 bcf = pe.SoilSample(h=h, theta=theta, k=k).fit(pe.Brooks)
 ```
 
