+++ _sources/examples/09_paper.ipynb.txt
@@ -211,7 +211,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 7,
+   "execution_count": null,
    "id": "5f4ae504",
    "metadata": {},
    "outputs": [
@@ -229,8 +229,8 @@
    "source": [
     "fig, ax = plt.subplots(figsize=(2.6, 3.35), layout=\"constrained\")\n",
     "pe.plot.swrc(wosten, label=\"Wösten\", ax=ax)\n",
-    "pe.plot.swrc(saxton, label=\"Saxton\", ax=ax)\n",
     "pe.plot.swrc(rosetta, label=\"Rosetta\", ax=ax)\n",
+    "pe.plot.swrc(saxton, label=\"Saxton\", ax=ax)\n",
     "ax.set_ylim(1e-2, 1e6)\n",
     "ax.set_xlim(0.0, 0.46)\n",
     "ax.xaxis.set_minor_locator(mplticker.MultipleLocator(0.05))\n",
