### 2026-06-11 10c6176
update type hint for matplotlib Axes
