### 2026-06-11 8ed5988
add air entry pressure to Gardner
