### 2026-06-11 2a41e17
create soilmodelconverter class
