+++ src/pedon/soilmodel.py
@@ -1403,3 +1403,286 @@ def resolve_soilmodel(
     raise TypeError(
         f"Argument must either be Type[SoilModel] | SoilModel | str, not {type(sm)}"
     )
+
+
+class SoilModelConverter:
+    """Namespace class for soil hydraulic model conversion methods.
+
+    All conversion methods are static and take a SoilModel instance as their
+    first argument ``sm``. Use :meth:`list_methods` to discover which conversions
+    are available for a given model type — it reads the ``sm`` type annotation of
+    each static method directly, so no extra registration is needed.
+
+    Example:
+    -------
+    >>> vg = Genuchten(k_s=1e-5, alpha=2.0, n=2.5, theta_s=0.45, theta_r=0.05)
+    >>> SoilModelConverter.list_methods(vg)
+    ['ghezzehei', 'morel', 'peche']
+    >>> gardner = SoilModelConverter.peche(vg)
+    >>> print(gardner.c)  # Output: 4.4
+
+    """
+
+    @classmethod
+    def list_methods(cls, sm: SoilModel) -> list[str]:
+        """Get all available conversion methods for the given SoilModel instance.
+
+        Inspects the ``sm`` parameter type annotation of every static method on
+        this class and returns those where ``isinstance(sm, annotation)`` is True.
+
+        Parameters
+        ----------
+        sm : SoilModel
+            The soil model instance to check conversions for.
+
+        Returns
+        -------
+        list[str]
+            Sorted list of method names applicable for the given model type.
+
+        Example
+        --------
+        >>> vg = Genuchten(k_s=1e-5, alpha=2.0, n=2.5, theta_s=0.45, theta_r=0.05)
+        >>> SoilModelConverter.list_methods(vg)
+        ['ghezzehei', 'morel', 'peche']
+
+        """
+        from inspect import getmembers_static
+        from typing import get_type_hints
+
+        return sorted(
+            name
+            for name, obj in getmembers_static(cls)
+            if isinstance(obj, staticmethod)
+            and not name.startswith("_")
+            and isinstance(sm, get_type_hints(obj.__func__).get("sm", ()))
+        )
+
+    @staticmethod
+    def morel(sm: Genuchten | Brooks | Panday) -> Panday | Brooks | Genuchten:
+        """Convert between van Genuchten and Brooks-Corey models using the Morel-Seytoux
+        equivalence method, which preserves the effective capillary drive.
+
+        This method implements the parameter equivalence from Morel-Seytoux et al. (1996),
+        providing conversion between Brooks-Corey (BC) and van Genuchten (vG) parameters
+        while preserving the maximum value of the effective capillary drive Hc.M,
+        defined as the integral of relative permeability over capillary pressure.
+
+        The conversion is based on two criteria:
+        1. Preserve the effective capillary drive (primary criterion)
+        2. Preserve asymptotic behavior at low water contents (secondary criterion)
+
+        Parameters
+        ----------
+        sm : Genuchten or Brooks
+            The soil model instance to convert.
+
+        Returns
+        -------
+        Brooks or Genuchten
+            Converted soil model instance:
+            - If input is Genuchten, returns Brooks model
+            - If input is Brooks, returns Genuchten model
+
+        References
+        ----------
+        Morel-Seytoux, H. J., Meyer, P. D., Nachabe, M., Touma, J., van Genuchten, M. T.,
[…]
