### 2026-06-11 bef43e4
add tests for converter
