### 2026-06-11 a6e19ce
test protocol conformance
