### 2026-06-18 06b9737
ensure proper capitalization in references
