+++ paper/paper.bib
@@ -11,7 +11,7 @@
 
 @article{peche_genuchten_2024,
   author  = {Peche, A. and Houben, G. and Altfelder, S.},
-  title   = {Approximation of van Genuchten Parameter Ranges from Hydraulic Conductivity Data},
+  title   = {Approximation of van {Genuchten} Parameter Ranges from Hydraulic Conductivity Data},
   journal = {Groundwater},
   volume  = {62},
   number  = {3},
@@ -120,7 +120,7 @@
 
 @techreport{heinen_staringreeks_2020,
   author      = {Heinen, M. and Bakker, G. and W{\"o}sten, J. H. M.},
-  title       = {Waterretentie- en Doorlatendheidskarakteristieken van Boven- en Ondergronden in Nederland: De Staringreeks; (Update 2018)},
+  title       = {Waterretentie- en Doorlatendheidskarakteristieken van Boven- en Ondergronden in {Nederland}: De {Staringreeks}; (Update 2018)},
   institution = {Wageningen Environmental Research},
   number      = {2978},
   language    = {Dutch},
@@ -130,7 +130,7 @@
 
 @incollection{healy_vs2d_1990,
   author    = {Healy, R. W.},
-  title     = {Simulation of Solute Transport in Variably Saturated Porous Media with Supplemental Information on Modifications to the U.S. Geological Survey Computer Program VS2D},
+  title     = {Simulation of Solute Transport in Variably Saturated Porous Media with Supplemental Information on Modifications to the {U.S. Geological Survey} Computer Program {VS2D}},
   booktitle = {Water-Resources Investigations Report},
   publisher = {U.S. Geological Survey},
   address   = {Denver, CO},
@@ -174,7 +174,7 @@
 
 @article{mathias_gardner_2006,
   author  = {Mathias, S. A. and Butler, A. P.},
-  title   = {Linearized Richards' Equation Approach to Pumping Test Analysis in Compressible Aquifers},
+  title   = {Linearized {Richards'} Equation Approach to Pumping Test Analysis in Compressible Aquifers},
   journal = {Water Resources Research},
   volume  = {42},
   number  = {6},
@@ -197,7 +197,7 @@
 
 @article{rucker_gardner_2005,
   author  = {Rucker, D. F. and Warrick, A. W. and Ferr{\'e}, T. P. A.},
-  title   = {Parameter Equivalence for the Gardner and van Genuchten Soil Hydraulic Conductivity Functions for Steady Vertical Flow with Inclusions},
+  title   = {Parameter Equivalence for the {Gardner} and van {Genuchten} Soil Hydraulic Conductivity Functions for Steady Vertical Flow with Inclusions},
   journal = {Advances in Water Resources},
   volume  = {28},
   number  = {7},
@@ -232,7 +232,7 @@
 
 @article{wosten_pedotransfer_1999,
   author  = {W{\"o}sten, J. H. M. and Nemes, A. and Lilly, A. and Le Bas, C.},
-  title   = {Development and Use of a Database of Hydraulic Properties of European Soils},
+  title   = {Development and Use of a Database of Hydraulic Properties of {European} Soils},
   journal = {Microbiology},
   journal = {Geoderma},
   volume  = {90},
@@ -243,7 +243,7 @@
 
 @techreport{wosten_staringreeks_2001,
   author      = {W{\"o}sten, J. H. M. and Veerman, G. J. and {de Groot}, W. J. M. and Stolte, J.},
-  title       = {Waterretentie- en Doorlatendheidskarakteristieken van Boven- en Ondergronden in Nederland: De Staringreeks},
[…]
