+++ paper/paper.bib
@@ -700,3 +700,11 @@
   year    = {2020}
 }
 
+@software{zenodo_pedon_2026,
+  author    = {Vonk, M. A. and Peche, A.},
+  title     = {{pedon}: A {Python} package for analyzing unsaturated soil hydraulic properties},
+  year      = {2026},
+  publisher = {Zenodo},
+  version   = {0.1.0},
+  doi       = {10.5281/zenodo.20590480}
+}
