+++ README.md
@@ -16,7 +16,7 @@
 [![Typed: Ty](https://img.shields.io/badge/typing-ty-darkgreen?style=flat-square)](https://docs.astral.sh/ty/)
 [![Formatter and Linter: Ruff](https://img.shields.io/badge/linter-ruff-darkgreen?style=flat-square)](https://docs.astral.sh/ruff/)
 
-Pedon (*from Greek: πέδον, pedon -> soil*) is a Python package for working with unsaturated soil hydraulic properties. It provides an object-oriented framework for soils, soil samples, and soil hydraulic models, making it easy to describe, analyze, and parameterize soil water retention and hydraulic conductivity behavior. At its core, Pedon treats each soil model as a Python class that defines a soil water retention curve and an unsaturated hydraulic conductivity function. The package currently includes implementations of several widely used models:
+`pedon` (*from Greek: πέδον, pedon -> soil*) is a Python package for working with unsaturated soil hydraulic properties. It provides an object-oriented framework for soils, soil samples, and soil hydraulic models, making it easy to describe, analyze, and parameterize soil water retention and hydraulic conductivity behavior. At its core, `pedon` treats each soil model as a Python class that defines a soil water retention curve and an unsaturated hydraulic conductivity function. The package currently includes implementations of several widely used models:
   - [Mualem-Van Genuchten](https://doi.org/10.2136/sssaj1980.03615995004400050002x)
   - [Brooks-Corey](https://mountainscholar.org/items/3c7b98df-13e3-486c-9d1e-949a7a869f76)
   - [Fredlund & Xing](https://doi.org/10.1139/t94-061)
@@ -27,15 +27,15 @@ Pedon (*from Greek: πέδον, pedon -> soil*) is a Python package for working 
   - [Campbell](https://doi.org/10.1097/00010694-197406000-00001)
   - Panday ([MODFLOW USG-Transport](https://www.gsienv.com/software/modflow-usg/modflow-usg/))
 
-Pedon allows these models to be evaluated directly and fitted to data. When laboratory measurements of soil water retention and/or unsaturated hydraulic conductivity are available, model parameters can be estimated by fitting analytical curves to the data using nonlinear least squares, following the same methodology as the widely used [RETC](https://www.pc-progress.com/Documents/programs/retc.pdf).
+`pedon` allows these models to be evaluated directly and fitted to data. When laboratory measurements of soil water retention and/or unsaturated hydraulic conductivity are available, model parameters can be estimated by fitting analytical curves to the data using nonlinear least squares, following the same methodology as the widely used [RETC](https://www.pc-progress.com/Documents/programs/retc.pdf).
 
-To support parameterization without direct measurements, Pedon includes access to several established soil hydraulic parameter datasets:
+To support parameterization without direct measurements, `pedon` includes access to several established soil hydraulic parameter datasets:
   - [HYDRUS](https://www.pc-progress.com/downloads/pgm_hydrus1d/hydrus1d-4.08.pdf)
   - [VS2D](https://doi.org/10.3133/wri904025)
   - Staring Series: [2001](https://edepot.wur.nl/43272) and [2018](https://doi.org/10.18174/512761)
   - [Clapp & Hornberger (1978)](https://doi.org/10.1029/WR014i004p00601)
 
-In addition, Pedon implements multiple pedotransfer functions that estimate soil hydraulic parameters from easily measured soil properties:
+In addition, `pedon` implements multiple pedotransfer functions that estimate soil hydraulic parameters from easily measured soil properties:
   - Van Genuchten: [Wösten](https://doi.org/10.1016/S0016-7061(98)00132-3)
[…]
