+++ paper/paper.bib
@@ -53,12 +53,12 @@
 }
 
 @article{mualem_model_1976,
-  author  = {Mualem, Yechezkel},
+  author  = {Mualem, Y.},
   title   = {A new model for predicting the hydraulic conductivity of unsaturated porous media},
   journal = {Water Resources Research},
   volume  = {12},
   number  = {3},
-  pages   = {513-522},
+  pages   = {513--522},
   doi     = {10.1029/WR012i003p00513},
   year    = {1976}
 }
@@ -220,12 +220,12 @@
 }
 
 @article{looy_pedotransfer_2017,
-  author  = {Van Looy, K. and Bouma, J. and Herbst, M. and Koestel, J. and Minasny, B. and Mishra, U. and Montzka, C. and Nemes, A. and Pachepsky, Y. A. and Padarian, J. and Schaap, M. G. and Tóth, B. and Verhoef, A. and Vanderborght, J. and {van der Ploeg}, M. J. and Weihermüller, L. and Zacharias, S. and Zhang, Y. and Vereecken, H.},
+  author  = {Van Looy, K. and Bouma, J. and Herbst, M. and Koestel, J. and Minasny, B. and Mishra, U. and Montzka, C. and Nemes, A. and Pachepsky, Y. A. and Padarian, J. and Schaap, M. G. and T\{'o}th, B. and Verhoef, A. and Vanderborght, J. and {van der Ploeg}, M. J. and Weiherm\{"u}ller, L. and Zacharias, S. and Zhang, Y. and Vereecken, H.},
   title   = {Pedotransfer Functions in Earth System Science: Challenges and Perspectives},
   journal = {Reviews of Geophysics},
   volume  = {55},
   number  = {4},
-  pages   = {1199-1256},
+  pages   = {1199--1256},
   doi     = {10.1002/2017RG000581},
   year    = {2017}
 }
@@ -233,7 +233,6 @@
 @article{wosten_pedotransfer_1999,
   author  = {W{\"o}sten, J. H. M. and Nemes, A. and Lilly, A. and Le Bas, C.},
   title   = {Development and Use of a Database of Hydraulic Properties of {European} Soils},
-  journal = {Microbiology},
   journal = {Geoderma},
   volume  = {90},
   pages   = {169--185},
@@ -497,7 +496,7 @@
   journal = {Groundwater},
   volume  = {62},
   number  = {1},
-  pages   = {124-139},
+  pages   = {124--139},
   doi     = {10.1111/gwat.13327},
   year    = {2024}
 }
@@ -508,7 +507,7 @@
   journal = {Groundwater},
   volume  = {54},
   number  = {5},
-  pages   = {733-739},
+  pages   = {733--739},
   doi     = {10.1111/gwat.12413},
   year    = {2016}
 }
@@ -673,7 +672,7 @@
   journal = {Water Resources Research},
   volume  = {29},
   number  = {2},
-  pages   = {305-319},
+  pages   = {305--319},
   doi     = {10.1029/92WR02339},
   year    = {1993}
 }
@@ -684,7 +683,7 @@
   journal = {Water Resources Research},
   volume  = {55},
   number  = {6},
-  pages   = {4994-5011},
+  pages   = {4994--5011},
   doi     = {10.1029/2018WR024584},
   year    = {2019}
 }
