+++ paper/paper.bib
@@ -11,7 +11,7 @@
 
 @article{peche_genuchten_2024,
   author  = {Peche, A. and Houben, G. and Altfelder, S.},
-  title   = {Approximation of van {Genuchten} Parameter Ranges from Hydraulic Conductivity Data},
+  title   = {Approximation of {van Genuchten} Parameter Ranges from Hydraulic Conductivity Data},
   journal = {Groundwater},
   volume  = {62},
   number  = {3},
@@ -197,7 +197,7 @@
 
 @article{rucker_gardner_2005,
   author  = {Rucker, D. F. and Warrick, A. W. and Ferr{\'e}, T. P. A.},
-  title   = {Parameter Equivalence for the {Gardner} and van {Genuchten} Soil Hydraulic Conductivity Functions for Steady Vertical Flow with Inclusions},
+  title   = {Parameter Equivalence for the {Gardner} and {van Genuchten} Soil Hydraulic Conductivity Functions for Steady Vertical Flow with Inclusions},
   journal = {Advances in Water Resources},
   volume  = {28},
   number  = {7},
@@ -220,7 +220,7 @@
 }
 
 @article{looy_pedotransfer_2017,
-  author  = {Van Looy, K. and Bouma, J. and Herbst, M. and Koestel, J. and Minasny, B. and Mishra, U. and Montzka, C. and Nemes, A. and Pachepsky, Y. A. and Padarian, J. and Schaap, M. G. and T\{'o}th, B. and Verhoef, A. and Vanderborght, J. and {van der Ploeg}, M. J. and Weiherm\{"u}ller, L. and Zacharias, S. and Zhang, Y. and Vereecken, H.},
+  author  = {Van Looy, K. and Bouma, J. and Herbst, M. and Koestel, J. and Minasny, B. and Mishra, U. and Montzka, C. and Nemes, A. and Pachepsky, Y. A. and Padarian, J. and Schaap, M. G. and T{\'o}th, B. and Verhoef, A. and Vanderborght, J. and {van der Ploeg}, M. J. and Weiherm{\"u}ller, L. and Zacharias, S. and Zhang, Y. and Vereecken, H.},
   title   = {Pedotransfer Functions in Earth System Science: Challenges and Perspectives},
   journal = {Reviews of Geophysics},
   volume  = {55},
@@ -355,13 +355,13 @@
   journal = {Geoderma},
   volume  = {427},
   pages   = {116123},
-  issn    = {0016--7061},
+  issn    = {0016-7061},
   doi     = {10.1016/j.geoderma.2022.116123},
   year    = {2022}
 }
 
 @article{wosten_texture_1988,
-  author  = {Wösten, J. H. M. and {van Genuchten}, M. Th.},
+  author  = {W{\"o}sten, J. H. M. and {van Genuchten}, M. Th.},
   title   = {Using Texture and Other Soil Properties to Predict the Unsaturated Soil Hydraulic Functions},
   journal = {Soil Science Society of America Journal},
   volume  = {52},
@@ -459,7 +459,7 @@
 }
 
 @article{seki_unsatfit_2023,
-  author  = {Seki, K. and Toride, N. and van Genuchten, M. Th.},
+  author  = {Seki, K. and Toride, N. and {van Genuchten}, M. Th.},
   title   = {Evaluation of a general model for multimodal unsaturated soil hydraulic properties},
   journal = {Journal of Hydrology and Hydromechanics},
   volume  = {71},
@@ -513,7 +513,7 @@
 }
 
 @article{brink_swap_2026,
[…]
