### 2026-06-24 7e83595
add references to the journal of open source software
