+++ CITATION.cff
@@ -1,0 +1,34 @@
+cff-version: "1.2.0"
+authors:
+  - family-names: Vonk
+    given-names: M. A.
+    orcid: "https://orcid.org/0009-0007-3528-2991"
+  - family-names: Peche
+    given-names: A.
+    orcid: "https://orcid.org/0000-0003-3761-2441"
+doi: 10.5281/zenodo.20590480
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+    - family-names: Vonk
+      given-names: M. A.
+      orcid: "https://orcid.org/0009-0007-3528-2991"
+    - family-names: Peche
+      given-names: A.
+      orcid: "https://orcid.org/0000-0003-3761-2441"
+  date-published: 2026-06-23
+  doi: 10.21105/joss.10409
+  issn: 2475-9066
+  issue: 122
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 10409
+  title: "pedon: A Python package for analyzing unsaturated soil
+    hydraulic properties"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.10409"
+  volume: 11
+title: "pedon: A Python package for analyzing unsaturated soil hydraulic
+  properties"
